[FIX] account: contruct string tax depending on company
Before this commit, if you have two companies and two diferent taxes (each one for company) the field tax_string does not discriminate taxes.
Example:
Configure a product with Sales Price = $100
Add a 21% tax for the first company
Add a 21% tax for the second company
The tax_string is (= $ 142.00 Incl. Taxes) but is worng

The tax_string must depend on company. So, tax_string must be (= $ 121.00 Incl. Taxes)