[FIX] website: close navbar dropdown on page click
Issue:
In edit mode, the "on click" dropdown remains open even after clicking
outside of the dropdown.

With this pull request [1], we manually handle the dropdown to keep it
open when clicking a dropdown item, but it prevents the dropdown from
closing when clicking outside of the menu.

Before this commit, we only handled the on-page click event for the
HoverableDropdown [2], but not for the clickable dropdown. To maintain
the same behavior for both clickable and hoverable dropdowns, this
commit binds the onPageClick event once we entered in edit mode.

[1]: https://github.com/odoo/odoo/commit/fe67b34f1a03d838df82a1307cf6467f994a8f3b
[2]: https://github.com/odoo/odoo/commit/455e03c743d6f1db70fab4e82b2fdbc963bc22ef

task-3370847

closes odoo/odoo#172677

X-original-commit: 25b7491d15a5e3249f0e6c649a37cbd75b5b3ae7
Signed-off-by: Robin Lejeune (role) <role@odoo.com>