[FIX] microsoft_account: raise the Badrequest if the post data doesn't available
[FIX] microsoft_account: accessing the url without any post data
When a user tries to access the URL directly, at that time the value of
dictionary `kw` is not available. So the error will be generated.

Error : KeyError: `state`

This commit will solve the above issue by raising the `BadRequest` if the value
of dictionary `kw` does not available.

sentry-4377121133

closes odoo/odoo#157720

X-original-commit: 382b64c2f14073cfff1a8ef9290b5c7834d52188
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>