[IMP] Add account_id.internal_group in depends of _compute_amount_residual
This pr adds the account_id.internal_group field in the depends of the _compute_amount_residual method.
The residual amount of an account line should be recalculated when the internal_group of an account changes. While this is not common practice, it may be necessary (mainly in migrated installations from previous versions).
If recomputed, invoices may remain in the process of payment, when in fact they are paid and reconciled.