[FIX] website: fix language switchers options visibility
Before this commit, if a website had at least two languages and the
copyright area was disabled, a redundant "Language Selector" button
would appear. This is not necessary anymore since the "Language
Selector" option is already included within the copyright options.

Steps to reproduce:

- Navigate to the website editor
- In the "Theme" tab, click "Add a Language" to add another language
- Type "French" or another language and install it
- A language selector will appear in the copyright area
- Click on "or edit master"
- Click on the "copyright" area at the bottom of the page
- Disable the copyright switch button in the right panel
- A "Language Selector" will appear below the "Copyright" switch button
  -> redundant

For the same reason, the "None" option for the language selector in the
navbar has been removed. The new "Header" options include a "Show/Hide"
element specifically for the language selector.

This commit addresses these behaviours.

task-3576937

closes odoo/odoo#140441

Signed-off-by: Benjamin Vray (bvr) <bvr@odoo.com>