[FIX] web, *: fix supportedOptions in some fields
*: analytic, mass_mailing

This commit fixes the supportedOptions object in some fields. Instead of using
a 'supportedTypes' key when descbribing the type of fields that an option must use,
the 'availableTypes' key must be used. This was simply misspelled.

A help has also been added for the fold_field option, following the fix for task-3474817

closes odoo/odoo#143127

Related: odoo/enterprise#51241
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>