[FIX] hr_attendance: generated unique kiosk key for every company
This issue occurs when a customer creates multiple companies. Then, after
installing the attendance module, after selecting the kiosk mode, an error
is generated. This is because, before installing the attendance module,
if the customer created a company, a single kiosk key is generated. This
key is different for each company, but it is the same.

step to reproduce-
- First created two company.
- Install the `Attendances` & Open it.
- Click on the Kiosk Mode from Menu.
- Error will be generated.

Error-ValueError
      Expected singleton: res.company(2, 1, 4, 3)

after this commit,Each attendance kiosk key generated during the company's
creation is unique.

sentry-4617967282

closes odoo/odoo#145382

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>