[FIX] hr_recruitment: limit job title size
step to reproduce:
- go to recruitment
- create a job with a title with a abnormally long title
-> the title is not limited & overflows

Bug explaination:
The parent has no defined width, so the title cannot be limited by
the original css rule.

Expected behavior:
The title size should be limited or be able to wrapped.

Bug resolution:
Remove the o_text_overflow class and thus permit the title to be
wrapped.

behavior after this commit:
The title is able to be wrapped.

task-3598895

closes odoo/odoo#142480

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>