[IMP] website, *: add aria-label on social media anchors
*: website_hr_recruitment

In order to improve the accessibility of the website pages, an
`aria-label` attribute has been added on some social media anchors to
avoid a warning of type `Links do not have a discernible name`. The
`options.js` file has also been adapted in order to correctly set the
`aria-label` attribute of added social media anchors.

Part-of: odoo/odoo#140453
Co-authored-by: qsm-odoo <qsm@odoo.com>