[FIX] tools: re-enable es.po as a base for Spanish locales
Starting Odoo 17, es_419 is used as a base for every Spanish locale
(es_VE, es_AR, es_BO...), but a lot of localizations still have their
translations written for es_ES only.

After this commit, both es_ES and es_419 will be used as base languages
for Spanish variations, ensuring that the translations will be loaded
whatever the file they where put into and the language of the user.

closes odoo/odoo#165279

Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>