[FIX] account: branch user cannot add products
Create a branch under the main company
Select an internal user and keep only this branch on the fields
"Allowed companies" and "Default Company"
Access with the user
Create an invoice
Add a product

Issue: user will be blocked by error
"""
Access Error

Due to security restrictions, you are not allowed to access 'Companies' (res.company) records.

Records: YourCompany (id=1)
User: Marc Demo (id=6)

This restriction is due to the following rules:
- company rule employee

Note: this might be a multi-company issue.

Contact your administrator to request access if necessary.
"""

This occurs because the accounting info is taken from the parent company,
but the branch user has no access to the parent company records

opw-3627439

closes odoo/odoo#148028

X-original-commit: fab2dc3ded2b638c8160bbc84ec14669c4e8546c
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>