[FIX] l10n_in: invisible the l10n_in_pan field if company is not indian
Before this commit:
If Indian-Accounting is installed in the user's system but the user is in
another company, still the l10n_in_pan(PAN) field is visible.

Technical Reason:
attrs was not there to invisible the field.
Adding invisible attrs if the active company is not Indian.

After this commit:
l10n_in_pan(PAN) field is not visible if the active company is not Indian.

Task-3635445

closes odoo/odoo#149496

X-original-commit: 5fb795f5ab3b54b2c37fe1b734f949cc3ce9a6de
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Naman Shah (nash) <nash@odoo.com>