[FIX] mail: fix non deterministic im status test
Before this PR the "Can handle im_status of unknown partner" test was
sometimes failing. This test is not related to im_status: it tests
that receiving a `mail.record/insert` notification properly inserts
the data in the store. This flow is already thoroughly tested as many
discuss functionalities rely on this mechanism. This PR removes this
test.

fixes runbot-52628

closes odoo/odoo#147705

X-original-commit: 39a368238f5fc9b8d8d4115f14a05499322e92e6
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>