[FIX] point_of_sale: fix pos_restaurant in free trial
When you install point_of_sale and then pos_restaurant, you end up with
a pos_config with an archived journal.

Then, if you change the chart of account via demo data, it doesn't
delete the archived log and crashes.

To correct this problem, we add `with_context(active_test=False)` to
the pos.config search.

closes odoo/odoo#147889

Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>