[FIX] payment_adyen: ignore missing street numbers
Commit ceb8b78 introduced a mechanism to throw an error if some address
fields were empty. The mechanism was introduced to prevent us from
sending False as an address detail to Adyen when it was not set since it
might negatively affect the fraud check score.
It however caused some payments not to go through because it relied on
Odoo's utils to extract the house number from the address, which would
fail in some circumstances (e.g., when the number is placed before the
street name). In some other cases, empty address details that were
previously accepted by Adyen caused the error to be thrown.
With this commit, we now pass Unknown instead of False or empty string
for the address details that are not set and that are not required by
Adyen in practice, unlike what is stated in AVS documentation. This
delegates the responsibility to block the payment to Adyen as the address
check and the related error are removed from Odoo.

closes odoo/odoo#149365

X-original-commit: 771b7ebd55646d8c96938e562802b47cb92d18fb
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: anko-odoo <anko@odoo.com>