[FIX] website: convert images added to Image Wall to webp
Before this commit, when images are added to an Image Wall or to an
Image Gallery, they remained in their original format.

After this commit, images added to an Image Wall or an Image Gallery
are converted to webp.

Thanks to [1], the Image Gallery dataset entries related to the
conversion are preserved.

[1]: https://github.com/odoo/odoo/commit/e73a1a96dcf8786c01205ae0a97b10e38ab9afeb

task-3666894

X-original-commit: 6aaddd38b0c39dda312fe79ab181e6c849c37da0
Part-of: odoo/odoo#170037
Signed-off-by: Benjamin Vray (bvr) <bvr@odoo.com>