[FIX] portal, website: allow to grant portal access for certain users
Context:
- Have a db with two websites (website1 and website2) with two different
domains.
- For those two websites, make sure that the "Shared Customer Accounts"
setting is disabled. The user is so forced to create an account per
website.
- As a visitor, create an account in each website using the same email
address.

-> No issue; it is possible to create two accounts with the same email
address.

Steps to reproduce the issue:
- Create two contacts (e.g. partner or customer) with the same email
address. One is linked to website1, the other is linked to website2.
- Select them, click on "Action" and then on "Grant portal access".
- Click on "Grant Access" for the first contact; it works as expected.

-> It is not possible to grant access for the second contact.

The goal of this commit is to ensure that the two described behaviors
lead to the same results. More specifically, it should be possible to
grant portal access for two partners that have:
- The same email address but are linked to different websites that have
the "Shared Customer Accounts" setting disabled.
- The same email address. One user is linked to a website that has the
"Shared Customer Accounts" setting disabled and the other is not linked
to a website.

It should however not be possible to grant portal access for two
partners that have:
- The same email address and are linked to the same website that has the
"Shared Customer Accounts" setting disabled.
- The same email address and are not linked to a website.

It should also not be possible to grant portal access for a partner that
is not linked to a website if it exists a user with the same email
address that is linked to the current website. Indeed, in this
situation, the partner is redirected to the current website at the
creation of its account and an "Access Denied" message would then be
displayed.

To solve the problem, the method `_get_similar_user_domain()` has been
implemented in the "website" module. Its goal is to build the domain
needed to find the users that have the same email than partners
depending on their linked website characteristics. The
`_is_portal_similar_than_user()` method has also been introduced. Its
goal is to check if the credentials of a portal user and a user are the
same. If it is the case, `email_state` field of the portal user is set
to `exist`.

task-3640503

closes odoo/odoo#166730

X-original-commit: b8a9ce02493093418f491cba7968884964944a89
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Signed-off-by: Colin Louis (loco) <loco@odoo.com>