[IMP] website: properly declare the stacking context of highlight SVGs
A `z-index: -1` is used on highlight SVG to make sure they appear behind
the text they highlight. However, it should still appear above
the background of the column/snippet behind that text. To make sure of
that, a new stacking context was created, using `z-index: 1`. While this
works... it also changes the z-index value, making it possible for
highlighted text to appear above other elements for no apparent reason.
This is replaced by a more appropriate `isolation: isolate;`.

task-3285817

Part-of: odoo/odoo#142968
Co-authored-by: qsm-odoo <qsm@odoo.com>