[IMP] hr_expense: editable expense lines
In #138387 we modified the expense lines tree in expense reports so that the user can click on the lines to see an attached receipt in the viewer at the side of the webpage. The tree was made to be readonly to make clicking the lines easier and to make it harder for the user to change a field by mistake. This commit reverts the lines to being editable again while leaving some fields as readonly like quantity and taxes as those can invalidate the expense.

task-3685651

closes odoo/odoo#149588

Signed-off-by: Quentin De Paoli <qdp@odoo.com>