[IMP] l10n_lu: Adding tags
In this commit we will adds tags to some accounts. The tags added correspond to
a mapping that was added in this PR: https://github.com/odoo/enterprise/commit/81c427ba1a505e1e9d4a64dd6227bbb747214142
This mapping was useful to separate the tax and net amount and then use it in
the appendix A of the annual report.

We decided the use tags instead of this mapping so that it's clearer for the
users and to give control to them.

closes odoo/odoo#149752

Task: 3646832
Related: odoo/enterprise#54525
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>