[FIX] mail: allow complete email for default from
Default_from on alias domain is currently managed like bounce or catchall.
We consider it is always only a left-part of an email address that should
be completed with alias domain name.

However and notably at migration default_from could contain a complete
email address. In that case better be defensive and keep the domain part
of the value.

Task-3690919

closes odoo/odoo#149887

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>