[FIX] web_editor: factorize open Cropper from toolbar
Before this commit, the cropper tool was opened by a closure that spawned the Cropper
tool aside from the editable document.

This was anoying as the editable document can be in an iframe, hence the Cropper would have
been virtually unusable.

This commit factorizes the code in a method to allow overrides. Which is done in
web_studio:ReportEditorWysiwyg.

opw-3484065

closes odoo/odoo#150904

Related: odoo/enterprise#55049
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>