[IMP] l10n_sg: taxes update
Update the taxes and taxes-related models due to new
tax rules.
This concerns a few accounts,taxes, tax groups, and the
tax report.

We also use this occasion to set non-important taxes as
inactive by default to reduce slightly the amount of
taxes that are shown out of the box.

Tax id #3585950

closes odoo/odoo#179715

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>