[FIX] base: can delete manual models with base fields
Create an ir.model custom (state = "manual" -- for example via studio) that has a mix
of manual fields (named x_...) and of base fields (originating from some mixin).

Unlink all linked views or object, and try to unlink that model eventually.

Before this commit, an error was raised because base fields couldn't be deleted, even though the table was empty.

After this commit, the deletion works.

Note that this commit is a fix of https://github.com/odoo/odoo/pull/130420/ , which added partial support for this
and a backport of 7550bcd61e52bc9c3de007d06cf95c91eaec893a which fixed the former PR in 17.0

opw-3558590

closes odoo/odoo#151282

X-original-commit: 3e84f3e9b0da397d82d5add0e37be58f0e30c7be
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>