[IMP] test_mail: make access tests independent from discuss models
Purpose is to make tests independent from channel model which has its own
access rules. We now use a test model with simple rules, allowing to
simulate models with

  * public access;
  * portal access;
  * internal access;
  * admin access;

Tests are rewritten to be more concise and precise and use this new model.

This is done in stable to keep coherency in tests codebase. It also eases
writing bugfixes as tests won't have to be modified during forward port
process. Finally it allows to backport improvements or fixes in later
versions if it applies to previous versions.

Task-4320561

X-original-commit: odoo/odoo@6370ca168c001cf4c44aa67f4c0bc48e9b90e5f1
Part-of: odoo/odoo#189226
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>