[FIX] pos_self_order: prevent error with online payment method
Prior to this commit, using an online payment method with an image in
self-ordering mode resulted in an issue preventing the opening of the
mobile menu. The specific error encountered was:
`TypeError: Object of type bytes is not JSON serializable`

This issue arose due to the image field being of type bytes. Since the
image field is not necessary for self-ordering, this commit addresses
the problem by preventing the loading of the image for a payment
method, resolving the serialization error.

opw-3624798

closes odoo/odoo#151388

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>