[IMP] {stock_}account: margin and inventory measures in invoice analysis
In order to better analyze profitability, we added an extra measure to
the Invoice Analysis report to show the "Margin" on every invoice line
based on the product cost price.

In order to have a simplified inventory valuation without fully using
the Inventory app, we also added an "Inventory Value" measure that also
uses the product cost price to show the change in inventory value based
on incoming and outgoing accounting documents.

An extra filter "Inventory Valuation" was added as well to show the
"Inventory Value" values per storable product and per month.

task-3708415

closes odoo/odoo#151805

Signed-off-by: Cedric Snauwaert <csn@odoo.com>