[IMP] website: list pages before controllers in sitemap
This shouldn't change anything regarding SEO, but is worth a try.
It will also impact the link suggestion when creating a link in the
editor, but having pages listed first also have sense there, or at least
it won't be worst.

The idea for the SEO part is that if the sitemap order (if too long)
would have some impact as crawlers might have a limited crawling budget
for your website and it will only crawl the first pages it finds.
Are those "first pages" impacted by the sitemap order? It's almost sure
it's not, as Google definitely knows how to crawl on its own, and is
even probably ignoring the sitemap most of the time.

Also, pages:
- Are probably always important content since you created manually a page to
  write something, while (some) controllers might just be content you
  care less about. Pages are probably always important while we can't
  say that for controllers.
- Should be fewer in number than controllers most of the time
- Have a lastmod set, as opposed to controllers
- May be created at any time, meaning a new crawler visit is needed,
  while controllers are almost never added in production, installing a
  new module is something very rare. Exception is about record's
  controllers which are "created" at any time like pages (eg a new
  product controller page)

For all those reasons, this commit reverse the pages vs controllers
order in the sitemap.

This is coming from our prod where some pages are yet not indexed while
they have been published months ago.

closes odoo/odoo#152350

Signed-off-by: Jérémy Kersten <jke@odoo.com>