[IMP] account: new ActionableErrors widget for warnings display
The new widget is meant to be a warnings header for wizards and forms,
where flows like EDI can list a series of errors and actions for
the users to fix the roadblocks.

The HTML result is an `<ul>` list with a point for each warning
and a link to a given action. Clicking the link will fire the execution
of a Python method on the backend, also passing back to it a series of
parameters the component has stored on setup time.

During setup, ActionableErrors accepts this structure as incoming data:
{
    'error_key' : {
         'message': _("Description of the warning"),
         'action_text': _("Text of the link"),
         'action_model': 'model.on.which.the.method.is.called',
         'action_name': 'action_python_method_name',
         'action_params': [...],'
    },
    ...
}

A map is used instead of an array, so that the Owl framework
could be more precise in rendering changes in future development.
(i.e. remove a warning after clicking the action link
without re-rendering the whole widget)

This Widget will be used both for `l10n_it_edi` and `account_intrastat`
and can be embedded in many other occasions.

Part-of: odoo/odoo#142596