[FIX] calendar: Adjust meeting dates in an all_day setup
Steps to reproduce:

- Calendar > create an all_day meeting with an additional attendee
- Save > Edit > change starting date to a later day > traceback

Cause of the issue:

The write call of the method _onchange_date is applied to a pseudo record.
However, the write method expects a record with an integer id to
correctly _send_mail_to_attendees down the line.

Fix:

Since we don't want to send_mail_to_attendees anyway, we skip this part
of the write method using the already existing contextual escape
'is_calendar_event_new'.

opw-3733753

closes odoo/odoo#154091

X-original-commit: f206781b3f4cababa4e69c8c9586f0f5406acc83
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>