[FIX] website: always set the first carousel slide as the active one
Before this commit, the first carousel slide was set as the active one
only at the `cleanForSave` of the `Carousel` options.

This means that:
- saving the editor after sliding the carousel without having started
its options (i.e. by just clicking on a carousel arrow) will save the
current slide as the first one to be displayed;
- saving the editor after sliding the carousel but with the `Carousel`
options started will save the first slide as the active one (because of
the `cleanForSave`).

These behaviors are not consistent.

This commit fixes this inconsistency by setting the first slide as the
active one for each dropped carousel, every time we save the editor.

Note that this behavior is tested in the last commit of this PR, which
adds a general tour testing the carousel.

task-3744613
Related to opw-3675019

X-original-commit: 91ec5f77d9574bd40b1a8b1b8572e76876abd14d
Part-of: odoo/odoo#173684
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Signed-off-by: Soukéina Bojabza (sobo) <sobo@odoo.com>