[FIX] point_of_sale: total show Nan in coins pop-up
Before this commit:
===================
In coins/bills pop pup, if user adds some coin and removes quantity from any of
the fields (blank space) then it's showing `NAN` and not calculating the other
inputted value which is completely annoying.

After this commit:
==================
If user removes quantity from any of the fields (blank space) then the input
will count `0` in the backend and calculate accordingly.

task - 3728728

closes odoo/odoo#153770

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>