[FIX] project_timesheet_holidays: changing public holidays deletes auto-cancelled linked timesheets
Steps to reproduce:
-------------------
1. Create and validate an arbitrary Time Off, checking that it also creates a timesheet entry.
2. Create a new public holiday spanning the entire duration of the Time Off created in (1)
3. Attempt to delete the timesheet entry created in (1), if it is not already deleted.

Although you will be asked to cancel the timeoff first, that is not possible because the timeoff will be auto-cancelled by this point, making the deletion of the timesheet impossible.

This PR adds an override call to unlink the timesheets before the leave (the inital validated timeoff) is force canceled.

Task-3697074

X-original-commit: 5e77f9c7e336a6080bf779f1a16ba7568c141808
Part-of: odoo/odoo#168223
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>