[FIX] account_payment: allow duplicate acquirer per company
The payment acquirer must always be linked to a single journal, one per company. It's currently an hard constraint we have on the journal.
So after the duplication, both are always linked to the same journal but only one appear on it.
This "fix" is improving that allowing the duplication of an existing acquirer and to be able to link them to different journals or on the same one.

opw-3704407

closes odoo/odoo#166352

X-original-commit: dba5afa1d6281260bde6e9a38db22ac548120676
Related: odoo/enterprise#62989
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>