[FIX] account: Amounts to Settle views and filters adjustments
This commit replaces the Display Name column by the Name column in the
Amounts to Settle view and adds the Journal Entry to the search options.
It also includes a small modification to the Amounts to Settle opening
action to adjust the amls that are shown.

task-3640027

closes odoo/odoo#148476

Related: odoo/enterprise#54137
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>