[FIX] odoo: parser recover True
If the xml node has not ascci characters then it will raise a
tracebeack wgen running cleanup_xml_node method.

In this case the cleanup_xml_node should read the string using
a parser to try to recover the data and be able to generate
get the clean version of the node.