[FIX] survey: fix charts size on results page
Purpose
=======

When you go on any survey results page, you notice charts are smaller
than they used to be on version 16. But there are still readable. The
annoying part is when you click on the `print` btn or press `CTRL + P`,
the charts disappear from both the page and the window preview.

How to fix
==========

- add classes `w-100` and `h-auto` to the canvas elements containing
the charts

Also, we empty the before/after print events in the public widget. They
are no longer useful, but someone may have hooked in those function.
We want to avoid the fix breaks someone else code.

see #135683
task-3667969

closes odoo/odoo#152705

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>