[FIX] website_sale: add !important to ribbon color
Before this commit, the bg-primary class was overwriting the bg_color
set by the user after saving changes as it has the !important tag.

Steps to reproduce the issue:
    1. Go to Shop page on website
    2. Enable edit mode
    3. Pick a product and create a new Badge (ribbon)
    4. Change the background color and save changes
        -> color reverts to bg-primary

This commit fixes this behavior by adding the !important tag to the
bg_color so that it properly overwrites the bg-primary color.

opw-3734598

closes odoo/odoo#155274

X-original-commit: 9ee115b58342b3e0dbc11081e7ff752c10f8bfa9
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: Simon Goffaux (sigo) <sigo@odoo.com>