[FIX] payment(_demo,_custom): show 'Secured by' element
If payment_custom or payment_demo is installed 'Secured by' element is
overriden and becomes hidden for all payment providers.

We removed xpath as it is more convenient for the users to see which
provider is behind a payment method. After this commit 'Secured by'
element is shown for all providers except for custom.

closes odoo/odoo#155014

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>