[IMP] mail: allow to configure email queue cron batch size
In this commit we add a configuration parameter allowing to control the number
of emails fetched each time the email queue cron is invoked. We now have two
parameters to control email sending :

  * 'mail.mail.queue.batch.size' which is the email queue size that will
    be used to search for outgoing emails (matching optional filters and IDs)
    and then given to 'send';
  * 'mail.session.batch.size' which is the batch size of a given segment of
    email to send using a specific server, used when splitting emails to send
    based on configuration (see '_split_by_mail_configuration');

Also fix 'mail.session.batch.size': when being 0 it does not call '_send'
while it should be considered as False (aka a maximum of 1000). This
is due to the usage of "split_every" which make iterators of length 0
if given 0 as step.

Task-3768123: Mail: Configurable cron queue size
Part of Task-3084943: Event: Improve communication scheduler scalability

Part-of: odoo/odoo#177046
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>