[FIX] hr: many2one_avatar field uses wrong model
Steps to reproduce
==================

- Install Timesheets
- Login as Admin
- Edit the access rights of Mark Demo:
  * Timesheets: "User: all timesheets"
  * Employees: "None"
- Logout and login as Demo
- Go to a project task
- Switch to the Timesheets notebook
- Add a new line
- Click on the employee field > Search More

=> An error occurred

Cause of the issue
==================

The hr.employee.public model is an SQL view of the hr.employee table
with differents permissions.

Since the user doesn't have the hr.group_hr_user group, the model
hr.employee.public should be used and not hr.employee.

Since [0], the model is switched depending on whether the user has the
appropriate group.

The EmployeeFieldRelationMixin is used and defines a getter for the
relation. That relation was not propagated to the Many2OneField.

Solution
========

Pass the relation to the many2OneProps.

---

[0]: https://github.com/odoo/odoo/pull/136786

opw-3765393

closes odoo/odoo#155743

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>