[FIX] product: Validate URLs on documents
Added a validation for URLs on product.document to resolve an issue
where incomplete links on documents will redirect to a local URL instead
of an external URL. Thus leading to a page that does not exist.

e.g.: youtube.com instead of https://www.youtube.com

This seemed like the proper course of action as there would be no reason
to not validate as any URL within the database can still be formatted properly and work the same.
api.constrains does not seem to work upon creation on this model, so I
opted to use the api.onchange.

opw-3698591

closes odoo/odoo#155806

Signed-off-by: Ryan Cen (ryce) <ryce@odoo.com>
Co-authored-by: Morgane Demesmaeker (edm) <69820391+Demesmaeker@users.noreply.github.com>