[FIX] website: restore default selected value for form selections type
Since the PR [1] changed the rendering engine of qweb, the default
"selected" value set on selects field on the form snippet were lost once
the page is saved.

This commit builds upon the changes made in [this commit] by reinstating
the default "selected" value.

Steps to replicate:

- Go to Website -> Edit.
- Drop a Form snippet onto the page.
- Click on the 'Company' field.
- In Field > Type, opt for "Selection".
- Choose option 3 from the options list to establish it as the default.
- Save the modifications.

Issue: The expected default value for the selection field is not
retained after saving.

[1]: https://github.com/odoo/odoo/pull/130467
[this commit]: https://github.com/odoo/odoo/commit/b42e9cc686e7d3ccf82cd091a5dc24028fff8a2b

task-3767819

closes odoo/odoo#155838

Signed-off-by: Romain Derie (rde) <rde@odoo.com>