[FIX] im_livechat: show live chat on top of other elements
The live chat should always be on top of other elements: some sites
that embeds it use z-index for various purposes. Currently, the live
chat does not uses z-index, thus is it can be hidden by other
elements.

This PR sets the z-index to the largest value handled by the browser.

opw-3732033

closes odoo/odoo#155903

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>