[FIX] account: Do not select wrong partner sharing same name with another
In a single database, you could have two partners who are called John Doe.

Before this commit, any statement line where the partner_name was set
with 'John Doe' would return the last one being created, due to the _order
attribute on res.partner model, even if the statement line was generated
from a payment of the other 'John Doe' (ie first one created).

With this commit, we ensure that the wrong partner is not selected, in
case we cannot differentiate one from the other.

closes odoo/odoo#157754

X-original-commit: 421a0ae53de8c947d68fb1c22066d4d73c1ce9b7
Related: odoo/enterprise#58699
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>