[FIX] pos_self_order: prevent multi-company access error
Before this commit, if the POS user didn't have access to the first
company, an access error would be raised when accessing the self-order
page. The steps to reproduce this issue are as follows:

1. Add another company and set up POS self-order.
2. Remove the other company from the Marc user.
3. Open the POS in the new company.
4. Try to access the mobile menu.
This would result in an access error.

The issue was that the first company existed in the
`allowed_company_ids` in the context. To solve this, we need to
correctly set the `allowed_company_ids` in the context.

opw-3744500

closes odoo/odoo#156357

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>