[FIX] hr_expense: replace JS test by tour
In odoo/pull/138387, a new feature was added where clicking on an expense line in an expense sheet will highlight the receipts of that line in the attachment viewer if any. A JS test was included for this feature that mocked the fetching of the sheet's attachment. Since this changed in the backend in odoo/pull/142029, the feature no longer worked but the test never broke since it was mocking the backend. This commit adds a tour to prevent that from happening again.

closes odoo/odoo#155417

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>