[FIX] point_of_sale: improve product search in large databases
Before this commit, searching for a product in the product screen was
problematic when dealing with large databases. The search function was
triggered too frequently, leading to performance issues.

With this commit, I have increased the debounce time for the product
search. This reduces the frequency of search operations when user is
typing.

opw-3788796

closes odoo/odoo#156806

Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>