[FIX] website_event: `modal_ticket_registration` overflow
This commit handle an issue causing the registration inner elements to
don't overflow as expected when multiple tickets are rendered on screen.

Note: This could probably be improved by review the overall modal layout
in order to make a different use of Bootstrap default classes.

task-3718417
part of task-3698364

Part-of: odoo/odoo#152434