[FIX] web: make resequence work after a resequence crash
Steps to reproduce:
- Go to a form view with a one2many with a handle widget
- Move a line in the one2many that make Odoo crash with the onchange

To generate a faulty view, create a form view with a one2many and add
an automation "On UI change" to make the resequence crash.

Current behaviour
-----------------
After the error being thrown, the user will not be able to
resequence this line again.

Expected Behaviour
------------------
After the error, the user will still be able to resequence
this line

closes odoo/odoo#193781

X-original-commit: 554f635ee8c5cd8e735a4d9c23cd88f8bad7dbf6
Signed-off-by: Simon Genin (ges) <ges@odoo.com>
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>