[FIX] web: make resequence work after a resequence crash
Steps to reproduce:
- Go to a form view with a one2many with a handle widget
- Move a line in the one2many that make Odoo crash with the onchange

To generate a faulty view, create a form view with a one2many and add
an automation "On UI change" to make the resequence crash.

Current behaviour
-----------------
After the error being thrown, the user will not be able to
resequence this line again.

Expected Behaviour
------------------
After the error, the user will still be able to resequence
this line

closes odoo/odoo#159678

X-original-commit: ae6854e433dc20e891448d6bffb9c5064970de42
Signed-off-by: Simon Genin (ges) <ges@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>