[FIX] mail: add get modelParam in ActivityController
We prepare the following fix by adding a modelParam getter that can be
overridden by sub-classes.

How to reproduce:
1. Schedule an activity for a document.
2. Go to activity view.
3. Select the document to preview in the inspector.
4. Rename the document.
5. Refresh the page.

The document name is not renamed.

In the kanban and list view, the document is updated because multiEdit model
parameter is enabled. It is not the case for the activity view.

Similarly to what is done in list and kanban controller, we add a modelParam
getter to the activity controller to allow to override the parameters in
subclasses which will allows to enable the multiEdit parameter in the document
activitty controller.

Task-3817795

closes odoo/odoo#164052

X-original-commit: 78353766ed636358ead8207e88a570ef80ec01fc
Related: odoo/enterprise#61792
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Pierre-Yves Dufays (pydu) <pydu@odoo.com>