[FIX] web_editor: resolved NaN issue in font-size dropdown in toolbar
Commit [1] resolved an issue related to the behavior of small font sizes, which
caused 'NaN' to appear in the font-size dropdown within the floating toolbar.
This occurred due to the removal of a variable definition, resulting in the
inability to compute the font size. This commit rectifies the problem by
reintroducing the variable in the SCSS file to ensure correct rendering.

[1]: 7931d1a14a3e4e65463cc101536061fa259b615d

task-3801894

closes odoo/odoo#158811

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>