[FIX] mail: partners with no users are displayed in command palette
Before commit:

if a partner is a member of any channel (eg, whatsapp channel) it will
be displayed in the command palette even if he does not have any
associated user or it is inactive. which will cause issues as you can
not chat with partner who do not have any dedicated user.

After commit:

partners with no dedicated users and inactive users will not be
displayed in the command palette to chat with when you enter @​.

Task-3815150

closes odoo/odoo#166220

X-original-commit: 9a3f9441c33e683603602555680def7f756c04b7
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>