[FIX] website: reapply original source on hovered image in edit mode
Steps to reproduce the bug:
- Add an image on the website and add an animation of type "On Hover" on
it.
- Save and edit.
- Click on the image to load its options.
- Save.

-> Problem, the image src is the raw data.

When the user clicks on the image, the `ImageHandlerOption` snippet
option is started. Because the image has an hover effect, the
`_refreshPublicWidgets()` method is called. Due to it, the
`ImageShapeHoverEffet` public widget is restarted. The problem is that
because the src of the image has been modified in the `_applyOptions()`
method of the `ImageHandlerOption`, the src of the image will not be
replaced by the original one at the `destroy()` of the
`ImageShapeHoverEffet` public widget. At the re start of this public
widget, `this.originalImgSrc` will be set to a base 64 source. At the
save of the editor, the public widget is destroyed and the image source
is set to those raw data. Because the image has not been modified, it
keeps it as a src after the save of the editor. To solve the problem,
the `destroy()` method of the `ImageShapeHoverEffet` has been adapted to
correctly set the source of the image when it has not been modified. To
do so, the `originalSrcBeforeHover` attribute is used to keep track of
the original image src. It is needed for example if a user applies a
filter on the animated image and then clicks on 'undo'; at the filter
application, `ImageShapeHoverEffet` is destroyed. Because the image has
been modified, the src of the image will not be replaced by the original
one. Without the `originalSrcBeforeHover` attribute, the original src
would be lost and it would not be possible to reset the src to the
original if the user clicks on 'undo' and then save.

task-3820127

closes odoo/odoo#158515

Signed-off-by: Benoit Socias (bso) <bso@odoo.com>