[IMP] hr_holidays : improve dashboard's display
The warning message when you have a half-day off confirmed while the
duration of this leave's type is daily is moved
towards the bottom of the sheet to avoid overlap between this
message and the ribbon.
And to improve the readibility of the dialog view, the scroll bar is
removed when it's not needed.

closes odoo/odoo#159454

Task: 3820211
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>