[FIX] mrp: Handle kit with BoM having qty != 1
Use case:
- Create a Kit BoM with finished qty to 5 consuming 10 components
- Do a sale order for 3 units (3/5 of BoM)
- Update the sale order line to 4 units

Current behavior:
The delivery has a huge amount to deliver

Expected:
The delivery is for 8 units

It happens because the method `_compute_kit_quantities`
always expect a BoM for 1 units.
`bom_line_data['original_qty']` always contains the number of times the
BoM will be needed and not the quantity of finished products.
In order to have the number of component by unit of finished product we
have to introduce the BoM quantity in the formula

closes odoo/odoo#160149

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>