[FIX] web: x2many: do not lose custom footer in form view dialogs
Before this commit, if the form view of an x2many (non editable
inline) field contained a custom footer (i.e. a <footer> node in
its arch, to override the default footer of the FormViewDialog),
the custom footer was correctly displayed for the first dialog to
open, but as of the second one, the default footer was displayed
instead. The issue has been introduced by [1], since which we
directly receive the arch as an xmlDoc, shared between all
instances of FormViewDialog that can be opened from that x2many.
So manipulating it (e.g. to remove the <footer> from it), impacts
the subsequent dialogs that can be opened. This commit ensures that
we manipulate a copy.

[1] #136376

Task 3805791

closes odoo/odoo#160181

Signed-off-by: Francois Georis (fge) <fge@odoo.com>