[FIX] *: add indexes
Deleting a partner may take a long time because odoo has to check an
entire table to find few records or no records at all that reference the
partner and set it to null.

So we are adding an index btree not null to speed up the deletion of
partners.

closes odoo/odoo#160184

Task-id: 3759406
Related: odoo/enterprise#59932
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>