[FIX] account: sync user_groupby with groupby
When changing `groupby` we want to also update `user_groupby` to avoid
raising `_validate_groupby_no_child` or `_validate_formula`.
It can always be set independently after.

Detected while upgrading `account_financial_report_it_sp` because
`<field name="groupby" eval="False"/>` was set in the data but
`user_groupby` was never updated. (but the constraint was triggered
because of the write)

Note that the constraint was not fired when needed either.

closes odoo/odoo#160301

Related: odoo/enterprise#60420
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>