[FIX] website_sale: fix eCommerce categories form view mobile
This commit fixes a layout issue inside the eCommerce categories form
view on mobile devices.

Prior to this commit, a `.oe_left` class was applied to the content of
the sheet, moving it "out of the flow", resulting in a wrong sized
`form_sheet`.

To fix the issue, we remove the `.oe_left` class, we apply a `.col-md-4`
`col-lg-6` to handle the width on large devices, and add a `.pe-3` to ensure that
the labels are not placed right next to the image in mobile.

task-3847917

closes odoo/odoo#160422

X-original-commit: 6aca24b14330b7681f8307f818135bd2bf71ab08
Signed-off-by: Valentin Chevalier <vcr@odoo.com>