[FIX] website_sale: wrap shop card's bottom content
In mobile, having a somewhat 'large price' (like a few thousands
currency units) can cause the bottom of the card to overflow (esp.
if buttons like 'add to card' or 'add to wishlist' are present).

This commit modifies the card template so that content is normally
spaced and can wrap to a new line if need be.

opw-3706637

closes odoo/odoo#160628

Signed-off-by: Bouvy Damien (dbo) <dbo@odoo.com>