[FIX] web: retain the 'All Calendar' filter on breadcrumb jump
Steps to reproduce:
- open the Calendar module
- click on the everybody's calendar filter option to view all the meetings
- click on any meeting to open the form view for that meeting
- now navigate back through the breadcrumb
- the everybody's calendar filter is now unchecked

Issue:
The 'Everybody's Calendar' option is not set and one needs to check it again
every time he come back through the breadcrumb.

Cause:
​All the other filters are saved in the backend through orm calls but the All
filter is not saved since it does not have a recordId.

Solution:
The filter state is now stored in the view's local state, so when a user come back
through the breadcrumb the previous set value is taken from the state.

task-3792066

closes odoo/odoo#165924

X-original-commit: f303712e8eb63e4cd0d09029426cc3471d76bd3b
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>
Signed-off-by: Suyash Rajput (sura) <sura@odoo.com>