[FIX] web, website_sale: add missing `.float-start class
Commit[1] aimed to fix a layout issue inside the eCommerce categories
form view on mobile devices.

Prior to Commit[1], a `.float-start` class was applied to the content
of thesheet, moving it "out of the flow", resulting in a wrong sized
form_sheet.

To fix the issue, we removed the `.float-start` class, but doing such
a change introduced a risk for users's customizations.

======
After
======

To prevent this issue, we reintroduce the initial lines, and will provide
a fix in another PR.

Commit[1]: 9837396
Related to task-3847917

closes odoo/odoo#160858

X-original-commit: 0914a0bd179a6b644e51de25bfad0555ea0f40ca
Signed-off-by: Pierre Paridans (app) <app@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>