[FIX] mass_mailing: set html lang attribute in unsubscribe page
Steps to reproduce issue:

1. Download two languages, set one as the user's language, the other as the website's
2. Make sure MailHog is running
3. Create a Mailing Contact with two different Mailing Lists
4. Create a mail with a link to unsubscribe
5. Send mail to one of the Contact's Mailing lists
6. Through MailHog, click on the subscription cancelling link
7. The first line is in the user's language and the others in the website's language

Explanation:

While most of the webpage originates from an XML file, translated using the language of the website.
https://github.com/odoo/odoo/blob/2ad911dab1f34b1b6dda1016cc189f0fea64b12c/addons/mass_mailing/views/mass_mailing_templates_portal.xml#L118
The first line is passed through javascript and translated on the fly.
https://github.com/odoo/odoo/blob/efd178daee689192d4e930a075475587038b3e0d/addons/mass_mailing/static/src/js/unsubscribe.js#L50-L55
The line is translated thanks to an object fetched at the load of the page. The url where the object is located can receive parameters including the language, but defaults to the user's and is not set in our case.
https://github.com/odoo/odoo/blob/0ca2ca9e8914dda77166b62dc8c2e9dc1ca1d3a4/addons/web/static/src/legacy/js/core/translation.js#L73-L78

Suggested fix:

The language can be passed through the attribute `lang` of the document.
https://github.com/odoo/odoo/blob/658bc65b22dba86a35f6bba0f892427e1be4aed1/addons/web/static/src/legacy/js/core/session.js#L205-L211
https://github.com/odoo/odoo/blob/e251354ce4b7df1763247385413e98f975981479/addons/web/views/webclient_templates.xml#L10
This option is already used in `website` and, [since 16.0](https://github.com/odoo/odoo/commit/5136ac7e4daaadda396c45c8538a81f8d6c44c0e), is inherited once `website` is installed.
https://github.com/odoo/odoo/blob/c3c8003c6d1c4d694531c93e803e7e9cd7ebaa83/addons/website/views/website_templates.xml#L39-L50

opw-3748027

closes odoo/odoo#160921

X-original-commit: 2e1d11dd5a5609e43be6f1a36bc7f4e2ca20849d
Signed-off-by: Stroobant Paul (stpa) <stpa@odoo.com>