[FIX] web_editor, website: fix clicking on we-list button
When a "we-list" button, like the "add item" button, is clicked while a
list input is focused, the click event doesn't trigger.

Steps to reproduce the bug:

- Drag and drop a "Form" snippet onto the page.
- Click on the "+ Field" in the options.
- Select "Multiple Checkboxes" for the "Type" option of the field.
- Rename "Option 1" to "ABC" in the Checkbox list.
- Click directly on the "Add new Checkbox" button.
- Bug: nothing happens. The new checkbox is not added to the list.

This bug arises because when clicking the button, the "change" event of
the input is triggered first before the "click" event of the button.
Since the "change" event refreshes the widget, the "click" event is
lost.

This commit adds a "setTimeout" in the "change" event handler to
prioritize "click" events within the "we-list".

task-3859207

closes odoo/odoo#167326

X-original-commit: 8d1dfd7c4250c01ac55c698fd2b24fbbf314730c
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>