[FIX] account_journal: avoid duplicate mail alias
Steps to reproduce:

Create an account journal.
Rename the created journal (let's say from A to B).
Create a new account journal named A.
Changing the name of a journal doesn't change the name of the alias, An duplicate error will occur.
Additionally the error can happen if the user has manually added aliases with the same name before upgrading to 17.0.

Adding the code of the journal to the alias name on creation when needed.

closes odoo/odoo#159037

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>