[FIX] payment_(buckaroo, stripe): updated the PM based on provider codes
When processing a transaction, the payment method was searched based on
the received code (e.g., 'sepa_debit') that was compared with the
`payment` module's generic codes (e.g., 'sepa_direct_debit'). This
commit ensures that we now compare with provider-specific codes for
Buckaroo and Stripe.

In practice, this mistake had little to no impact as most provider codes
match the generic ones, and we fall back onto the payment method
selected by the user if we can not find a more accurate one based on the
code.

closes odoo/odoo#161883

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>