[IMP] mail: preventing line break in systray activity
Before this commit:
In systray activity, for some languages, when there were few activities (exceeded a certain
length of text), it would break the line to display the labels.

After this commit:
Now, to maintain the design integrity, the text has been truncated to visually display
the labels like "aujourd'hu.." and maintain the design consistency.

Task-3869856

closes odoo/odoo#162001

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>