[FIX] account,sale: credit limit in multi-currency setup
The partner credit limit (warning) may not be computed correctly in a
multi-currency setup.

Note that the partner credit is always in company currency.

(1) Consider a draft invoice. To determine the warning amount
    the current amount in document currency (not company currency)
    is added to the current partner credit (roughly speaking).
(2) To compute the credits from sales orders of the partner
    (credit_to_invoice) we just add the amount_to_invoice
    from each sales order of the partner.
    But this field is in the currency of the order
    and thus may be different from the company currency.

To fix these issues currency conversions were added for both cases.

X-original-commit: 47a09537b3a7ed54f120d3c50b5847b353bb33dd
Part-of: odoo/odoo#164971