[FIX] lunch: prevent indexerror when deleting topping
When the user tries to delete the toppings,
a traceback appears.

Steps to reproduce the error:
- Go to Lunch > Configuration > Vendors > Open any vendor
- Now add extra 2 (topping 2) > Save
- Delete that extra 2 (topping 2) > Save

Error: A traceback appears:
"IndexError: list index out of range"

https://github.com/odoo/odoo/blob/07f6d71e4dd3fac7b4cacc545819430aa48823cd/addons/lunch/models/lunch_supplier.py#L192-L193
Here when the user deletes the extra 2 (topping 2),
"topping" will be [2,2] instead of [2,2,False].
So when it tries to access topping[2],
It will lead to the above traceback.

sentry-5203693767

closes odoo/odoo#162207

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>