[FIX] phone_validation: Upgrade SN Senegal phonenumbers module metadata
The Python module 'phonenumbers' received an update,
which includes localization data for Senegal "SN" phone number patterns.
At odoo we still use non-updated version of the 'phonenumbers', because
'phonenumbers' is an external module tied to the Ubuntu module repository,
that is not expected to receive a bump to the latest version soon.

This commit brings changes from the newest phonenumbers version into the
odoo as a metadata patch.

[Reproduce]
- Ensure 'phonenumbers' version 8.12 is used by Odoo.
- Install the contacts module.
- Create a new contact.
- Insert "+221750142092" into the phone or mobile field.
- Bug: Observe that the number isn't formatted correctly
    (spaces are not inserted to divide the number).

[Issue Explanation]
In version 8.12 of the 'phonenumbers' module in the Ubuntu repository,
the regex used to recognize mobile numbers was: "7(?:[06-8]\d|21|90)\d{6}"
which failed to capture some numbers. In the latest version of
'phonenumbers', the regex has been updated to:
"7(?:(?:[06-8]\\d|21|90)\d|5(?:01|[19]0|25|[38]3|[4-7]\\d))\d{5}"

opw-3663379

X-original-commit: 2bd8bbc861bde42464cf1b6bee0d22ac849f2c22
Part-of: odoo/odoo#164168