[FIX] purchase_stock: compute the on_time_rate for cancelled PO
Steps to reproduce:

- Create a PO:
  - 1 x storable product 1
  - 1 x storable product 2
- Confirm the PO
- Change the quantity of the PO line of product 1 to 0 (this modifies
  the associated delivery accordingly)
- Click on the customer to open the res.partner form
- Click on the "On-Time Rate" smart button

> Traceback: psycopg2.errors.DivisionByZero

Cause of the issue:

Clicking on that smart button will call the "_read_group" method and the
following sql expression will be added to the associated query:
https://github.com/odoo/odoo/blob/b4620c123b753caf2b77472a8d6c42d767471c53/addons/purchase_stock/report/vendor_delay_report.py#L57-L60
The error is therefore rasied because the SUM(qty_total) associated to
product 1 was reset to 0 because of our change on the PO line.

Fix:

Since SUM(qty_total) of 0 signifies that no units of that product was
not expected to be delivered for that customer. As such they should be
removed from the products whose "On-Time Rate" is computed from the
vendor.delay.report's.

opw-3852055

closes odoo/odoo#166601

X-original-commit: bb834b63e00d3bd345e088dfbef2d93b3439657d
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>