[FIX] hr_holidays: avoid auto-cancel of valid leaves
How to reproduce the issue:
- create a leave type that allows negative amounts
- create an accrual plan (demo seniority plan works)
- create an accrual allocation for an employee on the leave type
- set the number of days to 1
- go on the employee dashboard for time off
- create a leave in the near future (in the current month)
- create a second leave in the same period
- ensure with the balance on the day of the latest leave that
the balance is negative
- run the cron "Time Off: Cancel invalid leaves"

Behaviour prior to this commit:
The latest leave is cancelled.

Intended behaviour:
Since the leave is valid according to the leave type configuration,
it should not have been cancelled. The leave cancelled should only be
made in the case of an excess over the limit in negatives.

Additional changes:
This commit updates the explanatory comment in the _get_consumed_leaves
method to explicit every key usage. Also removes an unused key in
that method.

task-3879156

closes odoo/odoo#162643

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>