[FIX] mass_mailing: don't show in mail list managment non public lists
Issue: In the portal page for managing your mailing lists, to subscribe
or unsubscribe, you can see also the ones that are supposely marked for
not appearing in this view, when this happen the appear as "Mailing list
.#number" which is not ideal, since it will still let us subscribe to it
when we shouldn be able to.

Stesp to reproduce:

1. Install mass_mailing.
2. Create atleast 2 lists, one with "Show in Preferences" ticked and
the other not.
3. Now go to "/mailing/my" to manage the lists and see how they appear.

Solution: It seems logical to filter out for the lists we are not
subscribed the non public lists out (the ones that are not marked as
"Show In Preferences"), with this we will only see the ones we should
have access to see and subscribe/unsubscribe.

opw-3877385

closes odoo/odoo#162835

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>