[FIX] web_editor: consider fake unit and step inputs as numeric inputs
Since commit [1], a fake unit can be specified on `InputUserValueWidget`
widgets, thanks to the `data-fake-unit` attribute. It allows to display
a unit whose purpose is only to show a dummy unit and which does not
need to be converted to other ones.

However, it is still a unit, so the value written in the input should
still be considered as numeric and should therefore be processed as such
(e.g. checking if it has the correct format or if it is well a number).
This is currently not the case, since only values with real units (i.e.
with `data-unit`) allows to go through these checks. This means that
fake unit inputs can accept any values, which can cause errors if only
numeric ones were expected (e.g. for numeric CSS properties).

The same goes for inputs without unit but still having a step specified.
Indeed, the `data-step` attribute allows to increase/decrease the value
by using the keyboard arrows, which implies that the value should be
considered as numeric. Note that no option is really broken by that, as
they generally either parsed the value themselves or accepted the value
without any traceback assuming that the user should write logical values
and should not be surprised if it does not work. But the comma use case
makes it worth being fixed.

This commit adds the fake unit and the step cases to the checks, so the
input values are considered as numeric and can properly be parsed to
numbers.

Steps to reproduce:
For the fake unit case:
- In website edit mode, go to the "Theme" tab in the right panel.
- In the "Paragraph" section, set the "Line Height" option to a decimal
  number but with a comma instead of a period.
- Save.
=> There is a style compilation error, because the value was used as is
in the CSS file, which is not correct. If it was a real unit, the value
would have been cut before the comma by the `parseFloat` function.

For the step case:
- In edit mode, drop the "Form" snippet.
- Add a new field and set its "Type" option to "File Upload".
- Set the "Max # Of Files" option to a decimal number but with a comma.
=> If we wrote a number bigger than 1 to allow uploading multiple files,
the field will stay as a single file input.

[1]: https://github.com/odoo/odoo/commit/388e4bb2bfcaebdd4ff30277fb49a034592d7086

opw-3635886

closes odoo/odoo#163963

X-original-commit: 94165b4036625a0bdf866d0b709bef51b0651428
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Signed-off-by: Soukéina Bojabza (sobo) <sobo@odoo.com>