[FIX] website_event: prevent double toast
Toast messages when redirecting to event registration pages always appear twice.
This can be seen when going to an event not yet started, going to "rooms"
and any of them.

This is due to the element containing the toast data being duplicated
for mobile and desktop views.

We surround the duplicate call with a temporary empty value
for the toast message, preventing it from being duplicated.

Introduced in 9f2442d09b62fcb3f687ec39928ea0e3ac49557e

Also updates layout of registration buttons to be more in lines with
other modals.

task-3889923

closes odoo/odoo#163105

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>