[IMP] point_of_sale: no % on receipt for fixed tax
Before this commit, the `point_of_sale.OrderReceipt` template would
always print taxes followed by a % symbol.

This should not be so for fixed taxes, whose amount is not a percentage.

This commit removes the % symbol after fixed taxes.
As an improvement it also displays the tax name in the first column,
since the tax amount is already displayed in the following column.

No task.

closes odoo/odoo#163391

Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>