[FIX] account_peppol: fix tests
Responses should be mocked according to requests that are made during tests.
The tests use a sample Peppol file, where the supplier is `0198:dk16356706`.
When the document is imported and its content is extracted,
the corresponding partner is created with these `peppol_eas` and `peppol_endpoint` values.
That triggers a participant check but the response is not mocked for this endpoint.

This commit adds a mock response for the endpoint that matches the test file.

no task, reported by gawa

closes odoo/odoo#163603

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>