[FIX] product: fix _compute_product_pricelist() for new record
To reproduce the problem:
- Enable pricelist feature
- Create a new Contact > no pricelist shown (all versions affected)
- Save > still no pricelist shown (>= 17.0 versions affected)
- Exit and go back to the Contact: the pricelist is set

The _compute_product_pricelist() doesn't work properly for new record,
then we create a new contact, pricelist is empty (until the save for
version < 17.0, until save/exit/return on the record for version >=
17.0).

In 17.0, we introduced web_save, which makes create() and read() in the
same request. This is why the version >= 17.0 is more problematic and
more confusing for the user. Prior to 17.0, it was not great either,
but was never considered a bug (some effort has been made in
https://github.com/odoo/odoo/pull/153326/files). So this fix targets
versions >= 17.0.

Fix the issue by taking care of new ids in the
_get_partner_pricelist_multi() method. Also, add a Form test case for
every possible case.

opw-3872299

closes odoo/odoo#163768

Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>