[FIX] web: fix clickbot kitchen display issue
The Kitchen Display app uses a conditional menu that can leads to the
frontend. In some situation, other apps are tested again during the
Kitchen Display test leading to a mess in the logs and a epic failure.

As no viable solution was found, the menu is now skippep with this
commit.

closes odoo/odoo#164006

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>