[FIX] mrp: set warehouse loc/types as readonly
In debug mode, the special locations and picking types are shown on the
warehouse 'Technical Information' tab. All the values shown there are
readonly, with the exception of the mrp-related fields.
This was likely forgotten at the time, and can lead to
misconfigurations, so we make it consistent with the other location and
picking types introduced in stock and other modules.

closes odoo/odoo#166403

X-original-commit: 73ca9bfa4e200c16935f51ad5a9e55913f4281ba
Signed-off-by: William Henrotin (whe) <whe@odoo.com>