[FIX] stock: fix multi-company rental schedule
The rental schedule (Gantt view) shows the display name of the rented products,
which is read in sudo mode (`models.py` > `_read_group_format_result`). This
display name includes the total product quantity, which considers all companies
in the DB when computed in sudo mode (and without specifying a warehouse or
location). However, we should only consider the selected companies.

This change adds a domain which limits the companies that are considered when
computing product quantities.

opw-3877326

closes odoo/odoo#164196

Related: odoo/enterprise#61871
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>