[IMP] stock, mrp: Transfer,MO cancel warning
It's too easy to cancel a picking or a manufacturing by mistake.
In the opposite it's hard to reverse the error since there is no
reset to draft or any easy way to remove the cancel state.

The best solution would be to have an easy way to fix the issue but
until a proper solution a simple warning message could be useful

closes odoo/odoo#160609

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>