[FIX] web: decrease placeholder color
This commit aims to decrease a bit the contrast of placeholder text
in order to avoid confusion with regular text.

Prior to this commit, placeholder texts were using a gray value from the
grayscale, which was very similar to the `$body-color` in some
context.

To fix that issue, the color of the placeholder has been reduced to make
it lighter, even though it doesn't meet `3.5:1` contrast ratio.

task-3686510

closes odoo/odoo#166637

X-original-commit: 23277c54dd110d10230116ae5a1d6d3e0667cea2
Related: odoo/enterprise#63140
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>