[FIX] sale(_product_configurator): sale warnings application
Warning messages are displayed (and blocking products removed) on the end of the product configurator flow, but were not displayed (and blocking products were not removed) if the grid/matrix was used.

This commit makes sure the warning information is loaded before the configurator opening.  That way, the warning can be shown early in the flow to the salesman, and the product is directly blocked if it shouldn't be sold anymore.

opw-3894805

closes odoo/odoo#164199

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>