[FIX] l10n_es_pos: don't generate pdf for simplified invoices
In case it's a simplified invoice, the 'to_invoice' is automatically set to True if a corresponding
journal is set. Simplified invoices then generates a PDF which takes a few seconds to generate and
slow downs the user experience. This commit avoids generating the PDF in case it's a simplified invoice.

opw-3875944

closes odoo/odoo#164783

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>