[FIX] web: input field: always format value on Enter/Tab
Have a numeric field in a form view. The displayed value is
formatted, according to the thousands and decimal separators.
Say the value is 1,240.00. If the user edits the input and types
the same value, i.e. 1240, and then press "Enter", "Tab" or "Shift-
Tab", the value won't be formatted as it should be, because we
detect no change of value (which is true, but the value should
be formatted anyway).

With this commit, the value is always formatted on "Enter", "Tab"
and "Shift-Tab".

Issue spotted by fp

closes odoo/odoo#165112

X-original-commit: d0e0c1126ea23bbb42620cc58c36185cd99d5b4c
Signed-off-by: Simon Genin (ges@odoo) <ges@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>