[FIX] mail: escaping search term for regexps
Before this commit, trying to search for a message containing a round bracket would produce a traceback.
- search for the string "("

This happens because the search term is not being escaped for the regexp.

This commit fixes the issue by escaping the search term. Same is being done on the mock server. Added tests.

closes odoo/odoo#164320

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>