[FIX] web: Keep single company mode when logging into a company
Before:
When loggin into a company, it would simply add the new company in the
list of the currently selected/active companies, this made it harder
to use or stay in single-company mode.

After:
Logging into a company will now switch companies if we are in single-company
mode, essentially making the user stay in single-company mode. The user
can still toggle companies to enter multi-company mode.

Task ID: 3927141

closes odoo/odoo#168759

X-original-commit: 1aa2e3d4f6f96be9c26ba510d6f7603ba397c435
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>