[FIX] spreadsheet_account: don't mutate evaluated cell
Steps to reproduce:
- create a new spreadsheet
- type in cell A1 "01/2024"
- in A2, type =ODOO.BALANCE("1", A1)
- select A1
=> boom

Since 4939846774 `dateRange` is mutated in place.
But `dateRange` is actually the evaluated cell object.
It has originally the `type: "number"` and a corresponding
value whose type is a number.
But those accounting functions replaces the value with
a string. From that point, there's a mismatch between
`cell.type` and `cell.value`. The type says it's a number
but it's actually not.

When selecting the cell and the bottom bar statistics are
recomputed, because it expects numbers, not strings.

closes odoo/odoo#165362

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>