[IMP] point_of_sale: update order's partner on InvoiceButton
This commit is used for the enterprise PR (odoo/enterprise#62261).

When creating an invoice for a previous order in the POS, the class
InvoiceButton handles missing partners, writing the selected partner
to pos.order. But since the partner was not set to the order, the JS
client cannot retrieve the partner directly.

This commit remedies that by setting the partner to the order

task-3865331

closes odoo/odoo#165464

Signed-off-by: Laurent Smet (las) <las@odoo.com>