[IMP] account: remove amounts to settle menuitem
Description of the issue this commit addresses:

The Customers and Vendors dropdown start to bloat and the Amounts to Settle
view can be reached via the dashboard or the Aged Payable/Aged Receivable
reports. Therefore, we can remove them to lower the amount of buttons.

Desired behavior after this commit is merged:

The menuitems in the Customers and the Vendors dropdown to reach the Amounts
to Settle view have been removed.

no task linked, feedback

closes odoo/odoo#165806

Signed-off-by: William André (wan) <wan@odoo.com>