[IMP] stock: add attribute `_get_default_warehouse_id` to res.users
After version >= `16.4` module **repair** is using attribute
`_get_default_warehouse_id` which is defined in module `sale_stock`.
There is no dependency between that 2 modules.

But as the module `sale_stock` is autoinstall=1 and its all dependencies
are being installed when we install `repair` the module **sale_stock**
is also being installed automatically. That is why when we install **repair**
we will not have any problem on using that attribute. If we uninstall that
`sale_stock` we will have problem because `repair` module will not be able to
find the attribute `_get_default_warehouse_id`.

For fixing the issue I added that attribute in module `stock` as it is
dependent for both of `repair` and `sale_stock`, and also I use the inheritence
in module `sale_stock`

Steps to reproduce:

1. Install `repair` on `16.0`
2. Uninstall `sale_stock`
3. Try to upgrade to `17.0`

You will see this error:

```
 File "/home/odoo/src/odoo/17.0/addons/repair/models/repair.py", line 28, in _default_picking_type_id
    return self._get_picking_type().get((self.env.company, self.env.user))
  File "/home/odoo/src/odoo/17.0/addons/repair/models/repair.py", line 563, in _get_picking_type
    default_warehouse = self.env.user.with_company(companies.id)._get_default_warehouse_id()
AttributeError: 'res.users' object has no attribute '_get_default_warehouse_id'
```
Problem issued by commit : odoo@eaa6e1a
upg-1565705

closes odoo/odoo#167309

X-original-commit: 9e8cf10559b750b4f847c1345783f2e896dfd399
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>