[FIX] point_of_sale: not show inactive variant attributes in the shop
Currently, if you select a product that has variants, the popup window to
choose the variants shows both active and inactive variants.

Steps to reproduce:
-------------------
* Make a quotation with the product `Desk Organizer` and select `Plastic`
as its variant -> Confirm
* Go to **Products**
* Select `Desk Organizer`
* Go to `Attributes & Variants`
* Remove the `Plastic` option
  > The variant is now inactive
* Go to the **Point of Sale** App
* Open a shop session
* Select the `Desk Organizer` product
> Observation: the `Plastic` variant is shown

Why the fix:
------------
We compare the workflow with sales, when doing a new quotation with the
`Desk Organizer` product, the inactive variant is not an option to select.

We now only load the active variant attributes to PoS.

opw-3713374

closes odoo/odoo#166491

X-original-commit: 3e31cc41d75f67242ac85b9eba6c7ab478f09303
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>