[FIX] l10n_it_edi: handle "/" and "NA" as empty VAT
Steps to reproduce:
- Install Accounting, l10n_it_edi and Contacts
- Switch to an Italian company (e.g. IT Company)
- Go to Contacts
- Create an EU contact with "/" or "NA" as VAT
(e.g. a German contact with a full address)
- Create an invoice:
  * Customer: [the created contact]
  * Product: [any]
- Confirm the invoice
- Process the electronic invoice
- Check the generated electronic invoice

Issue:
When a customer has not VAT number, a special value is used in the
XML (i.e. 0000000).
However, "/" and "NA" that are equivalent to an empty VAT, do not
have the special value.
Instead, the "IdFiscaleIVA" section is empty in the XML.

Cause:
"/" and "NA" are handled as a normal VAT and are truncated because
the 2 first characters are not digits.

opw-3889051

closes odoo/odoo#166492

X-original-commit: d8650fc8874a61a57a2334d898e86eace4a7e563
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>