[FIX] base: phone number can span two lines in contact widget
Issue
----

Phone number can span two lines with some fonts for reports that use the contact widget (e.g. Delivery Slip).

Steps
-----

- Use raleway font for the Delivery Slip document, or at least the `Warhouse Address` section. You can do this
  by going to the `report_delivery_document` view and adding a `style="font-family: 'Raleway', sans-serif;"` to
  the div surrounding the `Warehouse Address` section.
- Change the phone number of the current company to `+33 3 28 44 55 01`.
- Create a delivery (without lines) and print a Delivery Slip without validating.
- The phone number spans two lines.

Cause
-----

When the phone number is long and takes up space in some fonts, the phone number is line-wrapped, which
is ok for other text but not for phone numbers.

opw-3834061

closes odoo/odoo#166336

Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>