[FIX] base_import_module: fix the tree view
In the tree view of industries, when you click on one, you get an error
that it can't fetch a record with an id -1.
Also when we try to select multiple industries and then click on a
button (activate, upgrade, print), it also raises an error due to the
id.

We therefore add a js_class on the tree view, calling the `more_info`
method on click toi redirect the user to the form view. Also we removed
the selector option from tree view as no button is working for
industries.

Task - 3834095

closes odoo/odoo#166470

Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>