[FIX] hr_holidays : Set env user as self if not employee in context
Steps to reproduce:
	- Install Time off module
	- Create a public holiday with Working Hours 'calendar_id'
	- Go to env user Time off dashboard

Current behavior before PR:
The public holidays that has 'calendar_id' assigned to it
will not be shown in Legend when checking the dashboard
of the env user. This is happening because we assign the
self 'hr.employee' with the employee_id in context
https://github.com/odoo/odoo/blob/c6142d9f1432cfabedcaa7264af9e5bc8e538c20/addons/hr_holidays/models/hr_employee.py#L313
and in '_get_contextual_employee' we just check if employee_id
exists in ctx or not
https://github.com/odoo/odoo/blob/17.0/addons/hr_holidays/models/hr_employee.py#L388:L392
but we don't check the value so if ctx is like
{'employee_id': None} it will return None

Desired behavior after PR is merged:
We are now checking if the 'employee_id' in ctx
has a valid value as if it is not we need to return
the employee object of the env user.

opw-3863947

closes odoo/odoo#166548

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>