[FIX] l10n_uy_demo: res partner emails
we add email to demo partners this way be able to activate the
send and print functionality to send by email.

If this one is not set then we are not able to re use this
partners for testing purpose creatimng new invoices that we
want to send and print.

closes odoo/odoo#166543

Signed-off-by: Josse Colpaert <jco@odoo.com>