[FIX] crm, phone_validation: hoist regexp phone indexes to mixin
The index definition that supports the implementation of
`_search_phone_mobile_search` defined in the `PhoneMixin` is created in
an `_auto_init` in `crm_lead` of the `crm` module.
This means that other models that use the mixin (for ex: `res_partner`)
will not get the indexes for it's table, leading to Seq.Scans when
searching for a phone number on those models.

By hoisting the index definition to an `init` in the mixin, all models
that inherit from the mixin will have the supporting index.

task-3942852

Part-of: odoo/odoo#166698
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>