[FIX] spreadsheet: remove odoo links on public share
Before this fix, when a spreadsheet contained any odoo link (to an action, a menu or a XML Id),
when a public user opened the spreadsheet and clicked on the link, he had a nonsensical page
because the parser and resolver of the link is never loaded for the public user (he would not have
access to the menu data anyway).

After this fix, we replace the odoo links by their label, so the link itself is removed.

closes odoo/odoo#167352

Task: 3925333
X-original-commit: 9ff67fdcea202d0f5df8a8f783bdfa2c4d03b658
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>