[IMP] account, l10n_it_edi: Allow flat discount from negative line
SDI does not accept negative lines. But people use it to define
flat discount on the invoice.
We reuse the logic from l10n_mx_edi, so we move it to account.
It will now try to create a flat discount for other lines that
have the same tax.

task-3943357

closes odoo/odoo#168939

Related: odoo/enterprise#64379
Signed-off-by: Laurent Smet (las) <las@odoo.com>