[FIX] bus: fix websocket rate limiting tests with noble
Ubuntu Noble (24.04) now provides python3-websocket 1.7.0.
The websocket rate limiting tests asserts that an exception occurs when
the code 1013 is returned because this code was not supported in
python3-websocket 1.2.3 provided by Ubuntu Jammy (22.04).

With this commit, presence of the status code in the lib is verified and
the test is adapted accordingly.

X-original-commit: 100626bc91b52ac356319722f29a1e303f20d79f
Part-of: odoo/odoo#167826