[FIX] account: chart template when no country
Before this commit, when creating a company and directly installing a chart
template without putting a country first, the installation of the chart template
was giving a traceback.
By updating the country of the company by the one of the chart template, the
issues will not happen again.

closes odoo/odoo#166896

Task: 3945833
Related: odoo/enterprise#64991
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>