[REV] website_sale_stock: cart quantity getter
This reverts commit 8c5996f2e3e5799acd80ebf01ab966d3f1922187.
This commit included changes that triggered `sale_get_order` calls
in routes/requests that did not use it previously.

Since those routes were using an environment sudoed with the superuser,
the call to `sale_get_order` wrongly changed the customer to OdooBot.

This will be addressed in a cleaner way in a separate PR, but for now,
we only revert the wrong commit.

opw-3949391

closes odoo/odoo#166913

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>