[FIX] web: kanban: record displayed twice after being d&d
In a grouped kanban view, drag and drop a record from a group which
contains a lot of records and has the "Load more" displayed in the
bottom (i.e. which contains more records than the limit). Before
this commit, the drag&dropped record was displayed twice: once
where it was dropped (which is fine), and once from where it was
dragged (which is wrong).

This happened because we didn't correctly synchronize the reload
of the column (which must be done because there're more records
than the limit, and we "removed" one record from the column) and
the update on the moved record (s.t. it belongs to the correct
column). As a consequence, we reloaded the column before updating
the record, so it was still part of the original column.

This commit fixes the issue by splitting the logic of the
_removeRecords function: before, it altered the records locally
(to filter out removed records) and updated the count, and, if
necessary, it reloaded the list. Now, _removeRecords is only
responsible to update the list locally, which must be done directly
as we want the user to get a direct feedback of the move. In the
flows where a reload might be necessary, we do it afterwards. In
the case of a moveRecord, we can then do it after the update on the
record.

This commit fixes another issue, as a side-effect: aggregates are
now properly updated in grouped list views after deleting records
(see second opw below).

opw-3891269
opw-3959610

closes odoo/odoo#171411

Signed-off-by: Simon Genin (ges) <ges@odoo.com>
Co-authored-by: Aaron Bohy <aab@odoo.com>