[FIX] stock: remove company check on quant location_id
Steps to reproduce:
- Have two companies selected (i.e. San Francisco & Chicago)
- Have 'Storage Locations' option enabled
- Create a product then 'Update Quantity'
- Create a quant and try to select the stock of the other company

Issue:
Only the locations in the first company will be shown. Moreover, in the
case of tracked products, if the last quant created was one to a
'transit' location without company (e.g. 'Inter-company transit'), then
only internal/transit location without companies will be shown.

The issue comes from the `check_company` clause on the `location_id`
field. This will make a check against the `company_id` of the record,
which itself is a related field from the location

This doesn't make much sense, as for checking if the company of a
location is correct, we check if the company of the location is within
the company of the location.

This also introduces the issue mentionned above, as `check_company` will
enforce to have ONLY that company & no-company, or just no-company
depending on the company of the default location found.

closes odoo/odoo#167233

Signed-off-by: William Henrotin (whe) <whe@odoo.com>