[FIX] web: position properly graph legend tooltips
Since [1] the legends tooltips are misplaced.

A similar issue has been fixed for the main tooltips of the graph view.
See [2]

Since [3] the Chart.js lib has been updated and the event received by
hovering the legend is now wrapped in an object. To access the native
event we also need to access the `native` key of that object.

[1]: c8ca9da7bcee2c122a9d6cf8cda89f02823ba42d
[2]: c1f08c60b7272e5a30951c53b67427e2f205b722
[3]: 7e3c1ecdb86110912b15722e600f9571692807ed

closes odoo/odoo#167971

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>