[FIX] mail: editing activity from record will pass the proper context
Issue: When creating an activity linked to a record like a lead, if we
edit this activity, for example from To-Do to Meeting, we lose the
context from the Lead and have a bad context based on the activity
instead, resulting in an improper link in the calendar.

Steps to reproduce:

1. Install CRM (this will also install the rest of the required modules).
2. Create or use an existing Lead.
3. Create a new To-Do activity.
4. Click on edit for this activity and change it to Meeting.
5. Now click on Open Calendar and place the activity.
6. Click on the calendar event we have just created, and we will see a
link to "Activity". Click on it.

Solution: To ensure the proper link to the Lead/Opportunity, we need to
pass the right context when editing the activity. This issue arose from
changes made in `activity_service.js` and the addition of the model
`mail.activity.schedule`, which separates the edit and create functions,
unlike the behavior we had in 16.0.

opw-3942711

closes odoo/odoo#167996

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>