[I18N] base: improve in Spanish/Español translations
In Spanish (MX) / Español (MX) Language

The Translation of "Valor desconocido '%s' para campo booleano '%%(campo)s'"
is done wrong in es_MX.po file of the base module.
﻿﻿
which leads to a traceback

Traceback : KeyError
'campo'

In the above-translated term, we mistakenly translated 'field' to 'campo'.
We should not translate 'field' because it contains dynamic values.

sentry-5454184338

closes odoo/odoo#168273

Signed-off-by: Adrien Widart (awt) <awt@odoo.com>