[FIX] l10n_in_*: handle zero distance on ewaybill
Before this Commit:
While creating the E-way bill, when sending a request with the distance set to
"Zero," the government portal updates the distance based on their own database
and sends it back in the response as an alert in the format " , Distance between
these two pincodes is \d+, ". However, the Odoo system did not update the
distance field with this value.

After this Commit:
The Odoo system now correctly parses the alert from the government portal
response and updates the distance field on the E-way bill with the computed
distance.

task-3961833

closes odoo/odoo#168419

Signed-off-by: Josse Colpaert <jco@odoo.com>