[IMP] stock: add test, double onchange lot_ids
Here is a little test to check the onchange for multiple unlinks.
The purpose is to document a bug and avoid future regressions.

Steps to reproduce:
 - Create a storable product SP tracked by serial number.
 - Register 3 serial numbers SN01, SN02 and SN03 in stocks for that
   product (this can for instance be achieved by creating and
   validating a receipt where you assigned the serial numbers).
 - Create a delivery order for these 3 units and mark as to do.
 - "Extend" the stock move of the picking to see the serial numbers,
   remove two of them of then and save.

Expected behavior:
 - The last serial number remains assigned to the stock move.

Current behavior:
 - The quantity set on the stock move is zero and hence saving the record
   removes all associated stock move lines and every remaining SN

Cause of the issue:
 - The "lot_ids" field of stock.move is a computed non-stored many2many
   field.

The first time we remove an SN, an onchange setting the "lot_ids" of the
stock move is performed and correctly updates the new record both in
terms of quantity and lots/stock.move.line.

The second time that we remove a serial number, the onchange tries to
modify twice the SN linked to our move: once by modifying the "lot_ids"
of the "move_ids_without_package" of the picking with an "update"
command, and once more by directly modifying the "lot_ids" of stock move
with two unlink (the previous one and the current one).

The first of those updates will incorrectly set "lot_ids" to an empty
value in the cache since, to not modify the "move_ids_without_package"
directly on "real" record, a "new" picking will be created to update the
cache [1].  During this call of the "new" method, the cache is updated
again [2] but this time with validate=False!  During the latter call,
the field lot_ids on the stock move of the "new" picking is updated
using the ids provided by the "convert_to_cache" method [3].  Now, the
method convert_to_cache of x2many fields clears all the ids of an unlink
operation in case validate=False [4, 5].

As a result, the field lot_ids on the stock move is set to be empty in
cache.  Soon after, method "_onchange_lot_ids" is triggered to compute
the quantity linked of products related to the stock.move [6].  Since
the lot_ids is empty in cache, this value is used and the quantity is
set to 0.  As such, the stock move line linked to the move will be
unlinked to the move when we hit "Save".

Note.  This flow use to work correctly before Odoo 17, since the web
client was using the command SET to perform the changes on field
lot_ids.  This behavior has been changed by
https://github.com/odoo/odoo/commit/e4b66668e0ff3d5444ef7fc7b79e6226c513e2c0

[1] https://github.com/odoo/odoo/blob/b11480e757ba35e0c616b84bf4dd801210de66f0/addons/web/models/models.py#L1030
[2] https://github.com/odoo/odoo/blob/c6ddaeaf2a6ae70df7f36a1cbc3a874d58ca0b88/odoo/models.py#L6436-L6437
[3] https://github.com/odoo/odoo/blob/c6ddaeaf2a6ae70df7f36a1cbc3a874d58ca0b88/odoo/models.py#L6003-L6004
[4] https://github.com/odoo/odoo/blob/c6ddaeaf2a6ae70df7f36a1cbc3a874d58ca0b88/odoo/fields.py#L4207
[5] https://github.com/odoo/odoo/blob/c6ddaeaf2a6ae70df7f36a1cbc3a874d58ca0b88/odoo/fields.py#L4220-L4221
[6] https://github.com/odoo/odoo/blob/c6ddaeaf2a6ae70df7f36a1cbc3a874d58ca0b88/addons/stock/models/stock_move.py#L1121-L1123

opw-3897055

Part-of: odoo/odoo#168735
Signed-off-by: Raphael Collet <rco@odoo.com>