[FIX] account_check_printing: checks to print on account journal dashboard
Description of the issue/feature this PR addresses:
There is a bug with traceback on accounting dashboard when is clicked "Checks to print" on bank journal with more than one outgoing payment with "Checks" Payment Method.

Steps to reproduce:

1) With a database in v16 onwards, install the account_check_printing module.

2) Go to "Accounting / Configuration / Accounting / Journals" and create a new journal bank with more than one outgoing payment with "Checks" Payment Method.

3) Go to "Accounting / Vendor / Payments" and create a new payment with journal created on step 1 and "Checks" payment method and confirm.

4) Go to "Accounting" and on "Accounting dashboard" click on "Check to print" on Journal created on step 1 and then a bug with traceback is raised.

Current behavior before PR:
There is a bug with traceback on accounting dashboard when is clicked "Checks to print" on bank journal with more than one outgoing payment with "Checks" Payment Method.

Desired behavior after PR is merged:
There is not a bug with traceback on accounting dashboard when is clicked "Checks to print" on bank journal with more than one outgoing payment with "Checks" Payment Method.

Video/Screenshot link:
https://drive.google.com/file/d/1r6A3uMkFv2nIDZ1HYJLgj6e18pNH1P67/view

Ticket Adhoc side: 75606
Task latam: 1217

closes odoo/odoo#171067

X-original-commit: 800f60f2fbfa3efa17bea8d3e23a23b7ec93ff96
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>