[FIX] resource: remove rounding factor
Issue:
------
The use case in which certain employees, who work part-time
over several days (and not part-time with full days with one
or more days off), want to take leaves distributed via allocations
given in days is unfortunately not taken into account.

For example:
- monday morning: duration = 0.4
- monday afternoon: duration = 0.4

allocation duration = round(0.8 * 16) / 16 = round(12.8) / 16
= 13 / 16 = 0.8125 ~= 0.81

Note:
It's strange to round up to the sixteenth of an 8-hour day.
It makes sense to have half-hour intervals, but it is does not
work correctly with calendars that differ from the
standard 8-hour calendar.

opw-3957459

closes odoo/odoo#168517

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>