[FIX] pos_restaurant: load pos data with noupdate
When demo data is loaded by user from UI, the same records are
created as is done with loading demo data, ideally they should
have same attributes but loading through python function doesn't
result in same because noupdate flag is not set. And when ORM
tries to remove these records it ends up causing an error.

Part-of: odoo/odoo#168726
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>