[FIX] hr_fleet: don't assign unassigned cars
Since following PR: https://github.com/odoo/odoo/pull/157057

We are trying to keep the driver of the car when we change the user of
an employee.

The main issue is that when we create an employee, the work_contact_id
does not exist yet as the function _sync_user is called before the
creattion of work_contact_id.

So when search for cars having the driver or futur driver similar to
work_contact_id, it returns all cars without driver or futur driver, and
it assign the user_partner on it.

We also avoid to writ the partner on the vehicle if it is already the
correct one which is set.

Part-of: odoo/odoo#168908
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>