[PERF] base,website: avoid assets invalidation in multiwebsite
When 2 different users navigate a different website while being
connected, it sets force_website_id on their session.
Getting the web.assets_backend when going to /web would then depend on the
theme installed on the forced website. The 2 users will invalidate the
assets backend of each other, invalidating the cache.

Generating the backend assets should not depend on the website as used in
get_current_website but it does due to tours from themes being in
website.assets_editor which is included in web.assets_backend.

Similar issue to this: https://github.com/odoo/odoo/pull/118526 but for
another corner case.

To avoid this, we use get_current_website to add the website_id in the
context so it is coherent everywhere. This will create multiple
web.assets_backend (one per website and only if they have different
themes) and will prevent the unwanted unlink of the previous assets.

closes odoo/odoo#169028

Signed-off-by: Benoit Socias (bso) <bso@odoo.com>