[FIX] web: prevent crash when DnD property without group
Steps to reproduce
==================

- Install crm
- Add a single property
- Drag and drop it to the second group

=> TypeError: Cannot read properties of undefined (reading 'elements')

Cause of the issue
==================

A properties field is composed of a list of property stored in a JSON
object.

A property can either be a separator, or any other type, char, bool, ...

The properties field has a columns props.

If we either have no separator, or we have only one and it is at the
first position, we enter the split mode. This means that properties are
displayed across the available columns.

If we move a property in a group, it means we wan't that property to
stay in that group. When we're in split mode, we need to add the least
amount of new separators so that we get the expected layout.

In some cases, it wasn't possible to get the actual group where we dropped
a property.

This can happen for example when the target was inside an empty group
with a default invisible separator (it has no name).

In some cases also, the behavior wasn't simply what was expected.

opw-3961445

closes odoo/odoo#188742

X-original-commit: 5efbe09285f74a8ffb3031aad7b9bd414322d9cc
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>