[FIX] resource, hr_contract: update leave calendar on resource change
Versions
--------
- 15.0+

Steps
-----
1. Have an employee on a 40 hour/week work schedule;
2. for a past week, create a sick leave for monday & tuesday;
3. create 8 hour timesheets for the remaining weekdays;
4. switch the employee's work schedule to 35 hour/weeks;
4. navigate to the week with the leaves in Timesheets.

Issue
-----
The hours displayed next to the employee name shows -11:00, ignoring the
the timesheets created by the leaves.

Cause
-----
`resource.calendar.leave` records have a `calendar_id` field which is
inialized to the resource's calendar, and updates if the resource
changes, but not when the resource's calendar changes.

As a consequence, the `_work_intervals_batch` method used for this
calculation gets called on the employee's new calendar, which is no
longer related to the calendar associated with their leaves.

Solution
--------
Replace the `onchange_resource` method with a compute method which
updates the leave's field if the employee's calendar changes.

Add an override in hr_contract to avoid updating leaves that fell under
expired contracts.

opw-3693131

closes odoo/odoo#172346

X-original-commit: 017093b89a0aa806d0a50acedfac9b2636bbb83e
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>