[FIX] point_of_sale: send display background image with IOT
Currently, when using the customer display feature with a display connected
to an IoT box, the background image is not used. Instead the placeholder
image is used.

Steps to reproduce:
-------------------
* Install and configure an IoT box
* Go to **Point of Sale**
* Select the Shop, edit it
* Check `IoT Box`
* Configure `Customer Display`
* Now go to the settings of the Shop
* Enable `Customer Display` if not already done previously
* Upload a background image, save
* Open the Shop session
* Select the customer display

> Observation: The background image is not shown.

Why the fix:
------------
When using the customer display feature with the IoT box, the request is
done with a Public User:
https://github.com/odoo/odoo/blob/f7e3ada5ed290a6f418e4e1a2abddfb9f242270c/addons/web/controllers/binary.py#L171-L176

Since Public Users don't have reading access on the `pos.config` model, access
rights & rules are triggered:
https://github.com/odoo/odoo/blob/f7e3ada5ed290a6f418e4e1a2abddfb9f242270c/odoo/addons/base/models/ir_binary.py#L52-L58

Since we failed the try block, we end up loading the placehorder image by
default:
https://github.com/odoo/odoo/blob/f7e3ada5ed290a6f418e4e1a2abddfb9f242270c/addons/web/controllers/binary.py#L185

We want the field `iface_customer_facing_display_background_image_1920` to
be available to public user. In the case we are trying to access it with a
public user, we make the request in sudo. Sudo environment stops existing
once the request is made.

opw-3873765

closes odoo/odoo#169808

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>