[FIX] web, website: searchbar menu columns width and tooltips
This commit addresses the same issue as in https://github.com/odoo/odoo/pull/137214
which is overflowing of search bar menu items when their name is too long
while also adapting some width values and make it apply to the website
name column. It also adds a height value to the containsers so that the
bottom is no longer squished when the content overflows and finally adds
tooltips for long item names which will most likely be truncated.

opw-3963754

closes odoo/odoo#169958

X-original-commit: 7eb47d8d1657e23016028d7c0f2e968947f5cafb
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>
Signed-off-by: Julien Carion (juca) <juca@odoo.com>