[FIX] website_sale: fix ribbon bg-color on product
Steps to reproduce:
- Go to Shop page on website
- Enable edit mode
- Pick a product and create a new Badge/Ribbon
- Change the background color
- Bug -> the bg-color is not set.

Issue :
ribbon bg-color is not set on the product.

Cause :
The !important attribute has been duplicated in the CSS rule since [1]
and [2]. As a result, the background colour CSS rule for the ribbon is
broken due to having !important twice.

[1]: https://github.com/odoo/odoo/commit/c6f4929f65b899736c556f8f0bb7824883a6d893
[2]: https://github.com/odoo/odoo/commit/9ee115b58342b3e0dbc11081e7ff752c10f8bfa9

fix :
After removing the '!important' attribute from the create and write methods,
we only encounter the '!important' attribute once. As a result, the ribbon
background color is applied to the product.

opw-3964071

closes odoo/odoo#168265

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>