[FIX] stock(_account): 'picked' to False on Done stock.move(.line)
When unlocking a picking and adding a new line, the 'picked' field would be False for both the new stock.move and stock.move.line.
This happens because the 'picked' field default value is False, and is only recomputed if the state changes to 'done', and not if the state is 'done' from the start.

The main issue it creates is that no stock.valuation.layer would be created on stock_account.

OPW-3980565

closes odoo/odoo#169816

Signed-off-by: William Henrotin (whe) <whe@odoo.com>