[IMP] l10n_it_edi_withholding: Apply Pension Fund tax by line
Pension Fund tax was incorrectly applied at the invoice level instead of the line level. When importing it as a vendor bill, this caused incorrect amounts due to the tax being applied to all lines.

This commit exports TipoDato = "AswCassPre" for lines with the pension fund applied. During XML import, it checks which lines should have the "pension fund" tax and applies it accordingly.

task-3969730

closes odoo/odoo#170012

Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>