[FIX] website_sale: avoid converting some extra images to webp
Steps to reproduce the bug:
- Go on a product page.
- Enter in edit mode and click on "Add Extra Images".
- Click on "Add URL" and upload an external CORS protected image (e.g.
https://tinyjpg.com/images/social/website.jpg).
- Click on "Add" to use this image as an external image.

-> Error.

Since [3], images uploaded through the website builder that are neither
gif nor svg are converted to the webp format by default except if those
images are CORS protected. The goal of this commit is to adapt [4] to
follow the same logic and not convert uploaded product extra image to
webp if the image is CORS protected.

[3]: https://github.com/odoo/odoo/commit/0449fe85cb0e1d639a4e1aeba26e90906f79254d
[4]: https://github.com/odoo/odoo/commit/b284293c18b6bcdafe2bda989c5e0a084d1acbd1

opw-3959983

closes odoo/odoo#186370

X-original-commit: 204a5aff9886bf0db9d7df84b896cea027781d77
Signed-off-by: Outagant Mehdi (mou) <mou@odoo.com>
Signed-off-by: Colin Louis (loco) <loco@odoo.com>