[FIX] mail: speed up partner mention
Before this PR, searching for a partner using a mention ("@") could take up to 18 seconds on a channel with 200 members. As a result, the UI would freeze while waiting for the search function to return partners to display.

After this PR, trying to mention a partner on the same channel only takes 80ms.

Most of the time was consumed by the `sortPartnerSuggestions` function, specifically by the `discuss.members` compare function that looped over every channel member twice per comparison.

The rest of the time was consumed by owl's `reactive` (~1/3 of the time).

To solve this issue, this PR focuses on two points:
- Providing a set of member partner IDs to the compare functions to speed up membership tests.
- Removing unnecessary reactive callbacks by using `toRaw` (searching partners is not coupled to rendering).

closes odoo/odoo#170314

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>