[FIX] web: duplicate favorite crash with enter
This commit fixes a bug where there would be a traceback occuring when
the user would either create a favorite with a name that already exists
or one with an empty name by pressing enter instead of save. The crash
occurs because the keydown event is not passed to saveFavorite and in
these cases, saveFavorite will invoke stopPropagation on an undefined
event. Solution: pass the event to saveFavorite.

closes odoo/odoo#170551

X-original-commit: f91512be8e1a53694c7d6520f1b226ec59b57bab
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>