[FIX] account: fix _compute_max_tax_lock_date depends
The field `max_tax_lock_date` is not always computed correctly.
This is since the compute function does not depend on the `tax_lock_date` field.

It should not happen in production since the `tax_lock_date` and the
check of `max_tax_lock_date` should not happen in the same transaction.
But it can still cause issues in tests.

The missing dependency is added in this commit.

closes odoo/odoo#170475

Related: odoo/enterprise#65262
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>