[FIX] web_editor: preserve selection on applying color to table
Before this commit:

Applying color to the table cell does not preserve the selection. This
issue occurs because we are deselecting the table using `deselectTable()`
and making the selection collapse to the start.

After this commit:

Selection remains preserved when you apply color to cells in table.

task-3822527

X-original-commit: 5d3ba1efff0857ceff7b2202ae050c102d8a4f90
Part-of: odoo/odoo#188756
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>