[PERF] point_of_sale: speed attribut selection in pop-up
Improved response time dramatically:

Before the commit:
The page froze with no response.

Reproduce the error:
In POS, have a lot of products with several attributes and long descriptions.
Go to the information pop-up page of a product.
Choose a variant on the product information.

We perform the search using 'category_search_string' across all the product descriptions.
If the search string is too long, the page can freeze or the search can be very slow.

After the commit:
Responses are almost instantaneous or significantly faster.

We pass the product_tmpl_id in the query, allowing us to retrieve all products with the same product_tmpl_id and perform the match only within these products.

opw-3992954

closes odoo/odoo#170766

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>