[PERF] account: speed up _inverse_product_id
Currently the condition on `self.product_id or not self.account_id`
is inside the lambda passed to `_conditional_add_to_compute`.
`_conditional_add_to_compute` calls the lambda in a filtered
so this condition will be evaluated for all the lines in self.
As it only depends on self, i.e. not the current line, we can
move it outside of the implicit for-loop.

speedup

Customer database. Calling _inverse_product_id increasing the
cardinality of self.

| self size | Before PR | After PR |
|     10    |    5ms    |   5ms    |
|     100   |    14ms   |   11ms   |
|     1000  |    271ms  |   36ms   |
|     10000 |    28.7s  |   300ms  |
|     50000 |    16min  |   1.5s   |

opw-3928958

closes odoo/odoo#172997

X-original-commit: 31eef976f1d757e84c1edd5f9893573f3277a521
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Aurélien van Delft (avd) <avd@odoo.com>