[PERF] hr_recruitment: add index on `active`
Description:
The `web_read_group` done when loading the "All Applicants" view is slow
when having a large quantity of `hr.applicant`

Analysis:
The query done by the `web_read_group` will get all the `active`
`hr.applicant`, but the field isn't indexed, so even if only a small
portion of applicants are active, we are still scanning the whole table
to filter out the applicants.

Solution:
Index the `active` field.

Benchmark:
On a DB with 900k `hr_applicant`, where only 80k are `active`, the main
query for the `web_read_group` takes

|         | Before | After |
|---------|--------|-------|
| Timings | 320ms  | 80ms  |

Reference:
task-4011294

closes odoo/odoo#171000

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>