[FIX] payment_stripe: add webhook event
When the authorization of the payment expires that happens when you do not
capture manually tx in 7 days (by default). Stripe sends a webhook that
payment is canceled which was not captured before this PR, hence the
changes were not applied for the corresponding payment.
With this PR the cancel webhook is captured and  the corresponding
payment is canceled odoo-side.

opw-3957346

closes odoo/odoo#171988

X-original-commit: 1bca577426ad3a0868ea071cfa03102085fc0a5e
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: Valeriya Chuprina (vchu) <vchu@odoo.com>