[FIX] stock: hide placeholder on report when no name on delivery address
Problem: Placeholders were added in 17.0 to be used when switching to
Studio mode. However, this placeholder will appear on the printed
report if the `res.partner` has no name. Creating a `Delivery Address`
for a `res.partner` does not require a name due to its type.

Purpose: Hide this placeholder in cases were `Delivery Address` record
does not have a `name` set.

Steps to Reproduce:
1) Enable `Signatures` for deliveries
2) Create a new `Delivery Address` with no `name` for an existing
`res.partner` record
3) Create a SO -> confirm -> sign delivery
4) Print `Delivery Slip`

closes odoo/odoo#171305

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>