[IMP] stock: `Run Scheduler` without thread
The button `Run Scheduler`, calling this method
`procure_calculation`, used to require a long
processing time to complete.

This is no longer the case. The operation completes
in a few seconds nowadays.

There is therefore no longer a need of a thread
to handle this.

Anyway, thread are not the way to go for long-lasting
operation. Operations taking time should be handled in a cron.
So, if this operation would take time, it should be replaced
with a cron trigger, then.

opw-4172329

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>