[FIX] mail: tests - fix non-deterministic cross-tab test
Before this PR, the `Channel subscription is renewed when channel is
manually added` test was sometimes failing.

This test ensures the bus subscription is renewed when the user joins
a channel. To do this, the test relies on the `assert.step` API: the
channel is added, we wait for a tick, and then we assert that the step
was correctly registered. However, a tick is sometimes not enough.

This PR replaces the usage of QUnit's tep API with the web one that
waits for the step to be triggered, which is far more robust.

fixes runbot-61018

closes odoo/odoo#171401

Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>