[FIX] web_editor: line breaks at edges of block anchor nested within li
Commit [1] handled cases of pressing enter at the edge of an anchor, which is a
child of an unbreakable element. The commit inserted the `br`'s after anchors;
however, it missed the situation where the anchor tags are block elements nested
inside an unbreakable element inside a `li`. In this specific case, inserting
two `br` tags after a anchor block resulted in the creation of two new lines.
This commit handles that case by only inserting one `br` tag after the anchor,
rather than both.

[1]: df6f8dd

task-3631910

Part-of: odoo/odoo#171469
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>