[FIX] web: Discarding in an x2many form removes the unsaved changes.
In version 16.0, when a change is made to a record in an x2many
field, saving it without saving the parent form and then accessing the
record again and pressing discard will delete the changes previously
made to the record.

With these changes, the issue does not occur, and discarding returns
the record to the state it was last saved in.

In further version, there's no issue, so we only forward port the
test.

Closing #171703

closes odoo/odoo#172289

X-original-commit: f32a14025308a190dca705abe872b3c5d98bebb6
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>