[FIX] account: remove readonly constraint
This commit will change the readonly constraint on the journal when an account
move has been posted. It will readonly only when the sequence number is
different from False, an empty string or a slash and if it was posted before

l10n_ar have specific naming depending on the journal, when changing the
readonly some test were failing due to an inverse that had the wrong journal.
To avoid changing the behavior of the localisation in particular we added a
xpath that will change the readonly back to what it was.

closes odoo/odoo#180266

Task: 4028973
X-original-commit: a69fd77780e064fff22d650fc284d5612b172f5d
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>