[FIX] account: remove readonly constraint
This commit will change the readonly constraint on the journal when an account
move has been posted. It will readonly only when the sequence number is
different from False, an empty string or a slash and if it was posted before

l10n_ar have specific naming depending on the journal, when changing the
readonly some test were failing due to an inverse that had the wrong journal.
To avoid changing the behavior of the localisation in particular we added a
xpath that will change the readonly back to what it was.

closes odoo/odoo#180085

Task: 4028973
X-original-commit: 5110d7f45ae8b558b50314ec2918341d4b54d641
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>