[IMP] payment: Smarter retrieval of outstanding account when duplicating provider
When you duplicate a payment provider, a new payment method line is generated.
However, in some configurations, people specifies custom outstanding accounts on
payment method lines. When duplicating the payment provider, the outstanding account
is not copied and then, at the end of the payment flow, the payment raises an error
because an outstanding account is missing and the journal items can't be generated.

task_id: 4019380

closes odoo/odoo#172803

X-original-commit: ad71652eeb11d6205f4ecf918004360f21f5e413
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>