[FIX] spreadsheet: ensure correct padding for sign-in button
Since upgrading Bootstrap a few months ago, the `nav-link-padding-x`
and `nav-link-padding-y` classes now rely on CSS variables defined
within the `nav` class. To ensure these variables are correctly
applied, we need to wrap the `nav-item` in a nav class.

This change fixes the missing padding issue for the sign-in button.

closes odoo/odoo#171887

Task: 4032574
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>