[FIX] l10n_latam_invoice_document: WIP
Compute method should not overried the document type is the invoice has
been posted before.

If it does then an incosistency will occurr because te name, document
type and sequence will not match. A new sequence non-real sequence
will be used. Also the user it is not aware this is happending because
the field is readonly