[FIX] base: re-allow fetching user-restricted images in reports
This is an oversight in bd7fe63ab4d7efd09304d9fd0546a16270c398eb

The expiration date should follow the format defined in RFC 6265,
which is Day, DD-Mon-YYYY HH:MM:SS GMT

The expected format of the date for the expiration for
`format_time`, using `babel`, is
`EEE, dd-MMM-yyyy HH:mm:ss z`
not
`E, d-M-Y H:m:s z`

During the development of the above revision,
during the functional testing,
the images in the reports were correctly working as expected.
This is because the date parser was confused because of the wrong
date format, and was ok for the date during which the tests were done
(on April 23 2024).

Nevertheless, since a few days, beginning of July,
the expiration date parser no longer accepts the confusing format date,
and believe the expiration date is passed,
and hence wkhtmltopdf no longer forward the session id as expected,
and hence the images with a restricted access right
can no longer be downloaded, as the requests coming from
wkhtmltopdf are no longer authenticated with the user printing the
report.

However, there is actually no need to set an expiration date in the
first place, as this expiration date is meant for the client-side only,
and the cookie jar file is deleted right after the report is printed.

We therefore take the choice to simply remove the expiration date
to solve the issue.

A unit test is added to cover the case, so this can no longer happens.

opw-3985316
opw-3985758
opw-3996620
opw-4023078
opw-4023430
opw-4023619
opw-4024377
opw-4024378
opw-4024528
opw-4027104
opw-4027115
opw-4027146
opw-4027207
opw-4027316
opw-4027338
opw-4027355
opw-4027391
opw-4027423
opw-4027622
opw-4027922
opw-4028933
opw-4029223
opw-4029395
opw-4029582
opw-4030219
opw-4032424
opw-4032956
opw-4033108
opw-4033142
opw-4033182
opw-4033766
opw-4033837
opw-4034508
opw-4035079
opw-4035246

closes odoo/odoo#172021

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>