[FIX] stock: report availability from sublocations
Steps to reproduce:

- Enable multisteps route in the settings
- Create a storable product P and put 10 units in WH/Stock/Shelf
- Create and confirm an SO for 1 unit of P
- Go to the associated delivery and change the Source Location to
    WH/Stock/Shelf
- Back to the SO, click on the chart icon > view forecast
- Unreserve the 1 unit currently used by your picking

Issue:

While you have 10 units in WH/Stock/Shelf perfectly suitable to fulfill
the demand of the picking, instead of displaying a line allowing you to
reserve the 1 unit, you have a line telling you that the qty is
"not available" -1 unit associated with the SO.

If you go back to the SO: the chart icon is displayed red and the
details displayed in the availability widget shows a:
"No future availability".

Cause of the issue:

The chart icon is red because the JS detects a forecasted issue:
https://github.com/odoo/odoo/blob/758ced91f8cb220a003a49b01e047b507f8509d7/addons/sale_stock/static/src/widgets/qty_at_date_widget.xml#L8
This issue comes from the fact that the demand is not expected to be
fulfilled since the `free_qty_today` of the SOL is smaller
than its `qty_to_deliver`:
https://github.com/odoo/odoo/blob/758ced91f8cb220a003a49b01e047b507f8509d7/addons/sale_stock/static/src/widgets/qty_at_date_widget.js#L45
https://github.com/odoo/odoo/blob/758ced91f8cb220a003a49b01e047b507f8509d7/addons/sale_stock/static/src/widgets/qty_at_date_widget.js#L54-L55
The reason for this forecasted issue is that the `free_qty_today`
is computed to be -1 rather than 1 here:
https://github.com/odoo/odoo/blob/758ced91f8cb220a003a49b01e047b507f8509d7/addons/sale_stock/models/sale_order_line.py#L80-L82
since the `move.forecast_availability` is itself -1 rather than 1.
Note that if the source of the move was still WH/Stock rather than
WH/Stock/Shelf both of these value would be set to +1 and the issue
would not appear. However, the `forecast_availability` of moves is not
correctly computed if the location_id of the move is a strict
sublocation of the warehouse. To be more precise, the forecasted
availability is set to a negative quantity here:
https://github.com/odoo/odoo/blob/758ced91f8cb220a003a49b01e047b507f8509d7/addons/stock/models/stock_move.py#L470-L472
since the report line computed in the `_get_forecast_availability_outgoing`
specifies that 'replenishment_filled' is False and should not be:
https://github.com/odoo/odoo/blob/758ced91f8cb220a003a49b01e047b507f8509d7/addons/stock/models/stock_move.py#L2228
We are finally at the core of the problem, the replenishement is not
filled because in the `_get_report_lines`, the `currents` dict used
to compute both the reserved and the on hand quantity only updates the
quantity of the warehouse if the location belongs to the warehouse:
https://github.com/odoo/odoo/blob/758ced91f8cb220a003a49b01e047b507f8509d7/addons/stock/report/stock_forecasted.py#L327-L330
As such, 0 units are considered to be available in these locations and
nothing can be taken from stock for these moves:
https://github.com/odoo/odoo/blob/758ced91f8cb220a003a49b01e047b507f8509d7/addons/stock/report/stock_forecasted.py#L236-L242
which of course results in the impossibility to fulfill the replenishment

opw-3979953

closes odoo/odoo#172106

Signed-off-by: William Henrotin (whe) <whe@odoo.com>