[FIX] mail: avoid calling pyToJsLocale with undefined
Calling pyToJsLocale with undefined returns an empty string.
Intl.ListFormat expects a valid BCP 47 tag or undefined.
When the user's language is not defined, passing it to pyToJsLocale will
return an empty string, which can't be used as the argument of
Intl.ListFormat.

This commit ensures that the value passed to Intl.ListFormat when there
is no language is undefined instead of an empty string.

Follow-up of https://github.com/odoo/odoo/pull/171837

closes odoo/odoo#172222

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>