[FIX] point_of_sale: discount product search breaks when using name
Using the name field in the domain for a discount product would cause
the search to break because name was not being loaded into the
Javascript through this function _loader_params_product_product.

Changed the types of errors that are thrown instead being shown a
notification popup in order to avoid the clunky Javascript error in the
case of a typeerror from not having the field loaded.

opw-4027900

closes odoo/odoo#172372

Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>