[FIX] website: allow restricted editor with modification rights to edit
For records such as blogs, no specific rights exists. Customizations
could decide to give access to specific records based on their own
rules. They are limited by the fact that the check for edition is not
overridable.

This commit makes it possible to implement specific checks for
restricted editors to edit records from the website builder by reusing
the overridable `_check_user_can_modify` method introduced in [1].

Steps to reproduce:
- Install `website_blog`.
- Setup a restricted editor user.
- Override `website._check_user_can_modify` to allow access for a given
user to a given blog post record.
- Edit that blog post with that user from the website builder.

=> Blog post could not be edited.

[1]: https://github.com/odoo/odoo/commit/fa20b5f2fed9e683c420798dce506a86fd5e511d

Part-of: odoo/odoo#172385
Signed-off-by: Jérémy Kersten <jke@odoo.com>