[FIX] account_edi_ubl_cii: map SE endpoint to company_registry
In Sweden, Peppol endpoint is required to be exactly 10 digits - no other characters are allowed.
Currently, we're mapping SE peppol endpoint to be the VAT number, which contains the country code prefix and a 12-digit number.
As a result, a lot of our invoices are being rejected by the Swedish access points.
Let's map it to the company registry number and add a check in constrains.

See:
https://docs.peppol.eu/poacc/billing/3.0/codelist/eas/

no task, customer feedback

closes odoo/odoo#172756

Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>