[FIX] point_of_sale: missing discount note on pos invoice
Steps to reproduce:
-------------------
* Activate discount on lines on user parameters
* Change the pricelist config to show discount to the customers
* Open a PoS session with the modified pricelist
* Create an order and apply some discount on the lines
* Invoice and pay the order
> Observation: On the invoice pdf you should have a line saying
"Price discount from X -> X" but it's not there

Why the fix:
------------
Before the fix we were comparing the `price_unit` on the line and the
`lst_price` of the product. But if a discount was applied on the line
the price unit is not affected, and so the line was never shown.
To fix this we compare the `total_price` of the line with the
`lst_price` multiplied by the quantity on the line.

opw-4019107

closes odoo/odoo#173687

X-original-commit: 3bf7739f9eaf33100edee731d9c97df8b635eaf8
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>