[FIX] google_calendar: duplicated event 'all day'
Before this commit, recurrent events created with the 'All Day' option
were duplicating the first event of the recurrence in Google side. This
was happening because we wrongly synchronized the single event with Google
before the synchronization of its recurrence (which already synchronize the
events in Google side itself).

After this commit, the recurrence is created normally using the 'All Day'
option. We achieve that by skipping the useless synchronization we were
doing in the single event: only the recurrence must be synchronized in this
specific flow.

task-3768121

closes odoo/odoo#177073

X-original-commit: 6ea41963440a68793aec52ca0dc1e0c0b647f4a9
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>