[FIX] hr_expense: correct tour test
Issue:
The tour for hr_expense was not up-to-date and was not being called via a Python test.

Due to not being called, an error was not being caught: the field is_editable
is used in a search in get_expenses_to_submit, the function used in the button
'Create Report'.

Fix:
- This commit corrects that by updating the tour, adding checks for the button 'Create
Report' and creating a python test file which runs the the tour.
- Filter expenses for report creations using is_editable, instead of using the field
in the search.

Part of task-3981049

Part-of: odoo/odoo#172442
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maira Salazar <srma@odoo.com>