[FIX] mrp: assign backorder MOs
**Current behavior:**
If a production spawns some backorder which has available
component quants in a package, the backorder will not have that
quantity reserved automatically.

**Expected behavior:**
This available quantity should be reserved by rule (based on the
manufacturing picking type's reservation configuration)
automatically.

**Steps to reproduce:**
1. Enable packages and 2-step manufacturing, use the default
    `at_confirm` option for the manufacturing method's
    `reservation_method`

2. Create an MO for 10 of some storable product that has 10 of
     some storable product as a component -> confirm the MO

3. In barcode, open the picking generated for the MO and
     manually add 1 of the components -> put it in a pack

4. Repeat step 3 an additional time

5. Validate the picking -> create the backorder

6. Open the initial manufacturing order in barcode

7. Consume one of the created packaged quants of the component,
     ensuring to also add the +1 quantity to the final product
     barcode line

8. Validate the MO -> create the backorder

9. In the backend, open the backorder to see that there is 0
     quantity reserved

**Cause of the issue:**
In this scenario, the backorder's raw move does not gain any
availability until after the original order is marked done,
after which there is never any logic for the reservation to
occur.

**Fix:**
During the backorder generation and the mark done for the
original order, attempt to assign any backorder productions at
the instant when some component quantity becomes available.

opw-3989977

closes odoo/odoo#172955

Related: odoo/enterprise#66496
Signed-off-by: Vincent Ethan <etvi@odoo.com>