[FIX] hr_holidays: take time into account during deserialization
Steps to reproduce:
- Install Time Off
- Set your time-zone to UTC+1
- Modify your working schedules for Monday Morning to start at 00:30
- Create a leave for Monday and Tuesday in the Time Off app

Issues:
The calendar show 3 days for the leave. This is due to the field
`all_day`, since it's set we will deserialize the record with the
function `deserializeDate`.

https://github.com/odoo/odoo/blob/44a1b163481b0b781028ce337d72fbb0c8730475/addons/web/static/src/views/calendar/calendar_model.js#L503-L509

This function doesn't take into account the time in the record, which
means that it will just deserialize the date which is Sunday in our
case because of the timezone conversion to UTC.

Because of this the frontend thinks that the leave takes place from
Sunday to Tuesday.

We can remove `all_day` thanks to this FIX which enable `allDaySlot`
options by default which allow us to preserve the original behaviour of
the calendar view in hr_holidays while correctly deserializing the
datetime.

https://github.com/odoo/odoo/commit/3ccc6bfa86618b0ca13c13960c15b60f8f75f3ce

opw-4043180

closes odoo/odoo#173147

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>