[FIX] delivery: don't return error when rate can't be fetched
Before this commit
==================
Some contracts in SendCloud do not have rates. This caused an annoying popup
when adding it to the Sales Order.

After this commit
=================
This commit ensures red banner is displayed with the same message instead of
the annoying popup when adding it to the Sales Order.

closes odoo/odoo#173791

Taskid: 4012182
Related: odoo/enterprise#66976
Signed-off-by: William Henrotin (whe) <whe@odoo.com>