[FIX] l10n_in: change the logger error to warning
Currently, a logger error is occurring when the user tries to change the
Chart of Account to India with Indian company having no state.

Clearly, it's related to demo data and it's not breaking the flow.

So we can change the logger error to a warning to reduce the noise in sentry

sentry-5610804086

closes odoo/odoo#173485

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>