[FIX] web: reset property values after changing type
This commit fixes an issue where the property values were not reset after
creating a property, setting its value, using the pager to access another
record and changing the property's type. When going back to the initial
record, the old value would still be there and it could trigger a
traceback depending on the type change (e.g. from string to integer).

The issue is fixed by using a record observer to trigger
_saveInitialPropertiesValues on record id change in the properties field
component. This allows to keep coherent values inside the initialValues
and to generate a new property name properly when the property's
definition changes. In case of a change in the property's name, the
value of the property for all other records using it will be properly
reset to default.

OPW-4015949

closes odoo/odoo#173505

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>