[FIX] account: set default taxes on existing product also without demo
Previously the code setting the default taxes on product was called only if the database was in demo, which makes zero sense. Creating a new company and assigning it its accounting package was not configuring correctly the existing products with the default taxes.

We now assign default taxes on all existing products when creating a company, as long as it had at least already a tax set (some data products, e.g TIPS in point_of_sale, are supposed to be used without any tax)

ticket 4037904

Part-of: odoo/odoo#173803
Signed-off-by: William André (wan) <wan@odoo.com>