[FIX] website_event_sale: check seats availability before payment
Steps to reproduce:
- Install `website_event_sale`
- Go to an event and for one ticket set maximum as 1
- Open event page and purchase a ticket, but don't complete the payment
- Do the same in private navigation
- Click on pay on the two page

Issues:
If you check with stripe payment is processed meaning that the two
people are charged however one of them can't get a ticket as there isn't
enough tickets available. Fix is to check before sending payment to
provider and return error if there's not enough seats. Concurrency issue
is not fully solved but would require much more precise timing, and a
proper solution wouldn't be stable.

opw-3772093

closes odoo/odoo#185845

X-original-commit: 2ef5e4f240e94037665ec00066072ba5988a05bc
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>
Signed-off-by: Mattis Megevand (mmeg) <mmeg@odoo.com>