[FIX] base: prevent putting filter and defaut out of reach
This fix add a verification check to ir.filter and ir.default.
This aims at preventing user to put such record out of their own reach

This behaviour is usually valid in Odoo, however in this case,
it is not normal that the user cannot delete the filter and default
that he created himself.

closes odoo/odoo#176724

X-original-commit: d8f9cae6acde267d549fab1692409dc445edd380
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>
Signed-off-by: Florian Vranckx (flvr) <flvr@odoo.com>