[FIX] web: remove flicker when changing datetime
Before this commit, when selecting a new date in a date(time)_field,
the value would first be reset before performing the "onchange" call,
resulting in a flicker (selected value -> null -> new value after
onchange).

Now, the value is not reset and is simply left as selected when
performing the onchange call, as to not trigger any flicker.

Part of task 3471424

closes odoo/odoo#174326

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>