[FIX] web: prevent crash in web_read_group with date groupby and limit
Steps to reproduce
==================

- Edit a kanban view so that the default_group_by is a date field
- Go to that view
- Open studio

=> Granularity not set on a date(time) field

Cause of the issue
==================

`web_read_group` is called with a limit of 1

First a call to `read_group` is made

Then in order to return the total number of groups when the limit is reached,
a call to _read_group is made.

Inside `read_group`, there is a compatibility layer that notably add a
default month granularity to date(time) fields.

This isn't the case inside `_read_group`, which causes a crash

Solution
========

Extract the compatibility layer related to the groupby from `read_group`
and use that for both methods

opw-4051657

closes odoo/odoo#174396

Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>