[FIX] stock: enforce deterministic order on freed reservation
Steps to reproduce:
- Create a product
- Update its quantity and set 4 in Stock
- Create two MO that consume each 2 units of that product with the same
scheduled date
- Create a scrap order with 1 quantity of that product and validate it

Issue:
The unreserve will be done on the first MO.

When going through the candidate move lines in the
`_free_reservation()`, we sort them depending on their picking, their
date if they have a move, or their id if they don't.

The issue is that when their related move have an equal date, then the
order becomes underministic. Now, always sort by their id as a last
resort, so we have a consistent order no matter what.

opw-4046240

closes odoo/odoo#176399

X-original-commit: 6898e010d9261dcd9f787b7ba4a22bd2644355d2
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>