[FIX] web: added `fullscreen` in size props of `Dialog` component
Previously, setting the `size` prop to `fullscreen` in the `Dialog` component
caused an error in debug mode, as the `fullscreen` utility class from Bootstrap
was not included in the props.

This commit resolves the issue by allowing  the `fullscreen`
option in the `size` props, ensuring compatibility with Bootstrap's utility classes.

closes odoo/odoo#179627

Closes: #174588
X-original-commit: 043d3eb14b156c06bbf1b9665f4a72473a71ac09
Signed-off-by: Pierre Paridans (app) <app@odoo.com>