[FIX] web: added `fullscreen` in size props of `Dialog` component
Previously, setting the `size` prop to `fullscreen` in the `Dialog` component
caused an error in debug mode, as the `fullscreen` utility class from Bootstrap
was not included in the props.

This commit resolves the issue by allowing  the `fullscreen`
option in the `size` props, ensuring compatibility with Bootstrap's utility classes.

Closes: #174588
Signed-off-by: Pierre Paridans (app) <app@odoo.com>