[FIX] website_sale_stock: add product name in payment error
__Current behavior before commit:__
When you check out your cart with a product not in the stock anymore you have an error message. There is no product name in this message so if you have multiple products in your cart you don't know which one to remove.

__Description of the fix:__
Add the product name in the error message, so it's easy to identify which product to remove of the cart.
The warning method is linked to sale.order.line so we can access to the product directly.

__Steps to reproduce the issue:__
- add products to your cart
- remove one or many products of the stock
- try to validate your payment

opw-4016059 (upgrade issues)

closes odoo/odoo#177043

X-original-commit: 2e6639bef50aec13e3eaaa3cc7f4ccb59aa9abd5
Related: odoo/enterprise#68546
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>