[IMP] base: enhance error message during pdf merge
Before this commit, the user would get a UserError when multiple attachments
are merged for export and one of them is corrupted. The error message did not
contain information helping the user identify the problematic records.

After this commit, the user will get a RedirectWarning stating the number of
corrupted files and a link to a form view (if single) or a list view (if
multiple) of problematic record(s).

Example workflow:
1. Go to Accounting > Vendors > Bills.
2. Create a bill with a corrupted attachment.
3. Select the corrupted bill amongst other bills with valid attachments.
4. Click on Print > Original Bills.
5. The error is: "Odoo is unable to merge the generated PDFs."

Cause:
Errors raised by instantiating a PdfFileReader of a corrupted file were
handled with a UserError with a static error message.

Solution:
In _merge_pdfs method signature, add an context option to define
the way errors are handled. This way the default behavior stays the same if called
outside `render_qweb_pdf`.

In _render_qweb_pdf, a custom handler is defined to keep track of the
corrupted streams and log them in the RedirectWarning.

opw-4067992

closes odoo/odoo#174701

Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>
Co-authored-by: mcou <mcou@odoo.com>