[FIX] sale_mrp: compute valuation on all bom products
Suppose the following structure:
----------------------------------------------
BoM of Main kit:
  - BoM Type: Kit
  - Quantity: 4
  - Components:
    * 1 x Sub kit A
    * 1 x Sub kit B

BoM of Subkit A:
  - BoM Type: Kit
  - Quantity: 1
  - Components:
    * 2 x Component A (Cost: $10, Storable)

BoM of Subkit B:
  - BoM Type: Kit
  - Quantity: 1
  - Components:
    * 2 x Component B (Cost: $6, Storable)
----------------------------------------------

When creating a sale order for 1 Main Kit, delivering the components
and posting the invoice, the Cost Of Goods Sold computed by the
_stock_account_get_anglo_saxon_price_unit method was ignoring the COGS
for the Subkit B.
This is due to the commit https://github.com/odoo/odoo/commit/31e1352df686d8a23628ade83d321929c49d6f4e
which fetches the BOMs from the stock moves linked to the Sale Order. As
no component is present directly in the bom lines of the Main Kit BOM,
the bom was omitted.

Now, if the product's bom is not present in the stock move's bom, we'll
fetch one the old way.

opw-4033293

closes odoo/odoo#181598

X-original-commit: f005a58fc6a0693fb474fbcee7fb8b6c3643ad4e
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>