[IMP] web: make company selector inheritable
Currently, it is not possible to inherit the company selector widget.
To modify the behavior of the widget, developers need to create a new
one instead of inheriting it, which is not ideal. This commit introduces
changes that allow inheriting the company selector widget and also makes
it easier to replace or extend the CompanySelector class within the
SwitchCompanyMenu.

closes odoo/odoo#184008

X-original-commit: d3e537f889f8e0bb3c5230f8c32c41aeb97444b7
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>