[FIX] survey: chart issues
Issue 1:
Previously, the chart was being overlapped by the navigation on smaller screen
sizes. This commit fixes the issue by limiting the max-width for navigations and
also hiding the button text for smaller screens.

Issue 2:
Previously, after a certain amount of data, the bar colors defaulted to black as
the dataIndex exceeded the colors array. This commit fixes the issue by using
the modulo operation on dataIndex with the colors array to repeat the colors,
and also added additional colors referenced from reporting.

Task-4089537

closes odoo/odoo#183679

X-original-commit: 06b47345e7c6279c530a5fe7b6b98ddd75c353ea
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Aashish Thakur (aath) <aath@odoo.com>