[FIX] point_of_sale: remove POS order attachments on sync error
Before this commit, if an order failed to synchronize due to a
concurrent update error, the order would be captured in one process
and saved as an attachment in another process. This fix ensures that
the attachment is removed when capturing an order that already exists
in the database, thereby eliminating unnecessary POS order attachments.

opw-4091844

closes odoo/odoo#175668

X-original-commit: 4d9611924d753e314efdcc571c2bbe875b69e602
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>