[FIX] sale_project: allow linking task based on invoice/shipping partner
Versions
--------
- 17.0+

Steps
-----
1. Have industry_fsm_sale from enterprise installed;
2. create a SO using a different `partner_id` for the shipping address;
3. add a product that uses timesheets without creating a task;
3. confirm the sale order;
4. in Field Service, create a new task;
5. set Customer to `partner_id` used for shipping address;
6. select the relevant Sales Order Item;
7. add a product in catalog.

Issue
-----
After adding a product to the field service task, a new SO is created
instead of adding it to the existing SO (as would be the case if the
task was created by the SOL).

Cause
-----
When computing a task's `sale_order_id`, it checks whether the task's
`partner_id` matches that of the sale order. FSM tasks generally use the
`partner_shipping_id` instead, so if this one is different,
`sale_order_id` is set to `False`.

Solution
--------
Aside from a sale order's `partner_id`, also check its
`partner_invoice_id` and `partner_shipping_id` when deciding whether the
task should get linked to the SO.

opw-3901885

closes odoo/odoo#175658

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>