[FIX] hr_holidays : Delete time-off timesheets when archiving employee
Steps to reproduce:
	- Install Timesheet and Time-off modules
	- Let the time-off validation creates a timesheet
	- Create a time off for an employee and validate it
	- Archive this employee

Current behavior before PR:
After archiving an employee you will still have the timesheets
related to his upcoming leaves even when you are closing all of
his activities.
This is happening because when archiving an employee we are just
updating the state of his hr.leave to refuse without unlinking the timesheets
https://github.com/odoo/odoo/blob/15.0/addons/hr_holidays/wizard/hr_departure_wizard.py#L23

Desired behavior after PR is merged:
While archiving an employee we are calling the 'action_refuse' method
where it will update the state of the employee's leaves and also will
unlink the timesheet related to those leaves.

opw-4070684

closes odoo/odoo#176334

X-original-commit: 5d61f36674f2aec86f6fc50d5fea1ab622050176
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>