[FIX] website_blog: prevent "List" layout from overflowing on mobile
Steps to reproduce:
- Go to the "/blog" page.
- In edit mode, set the "Layout" option to "List" and save.
- Toggle the mobile preview or reduce the screen size at around 580px or
  below.
=> There is a horizontal scrollbar.

The scrollbar appears because the row containing the blog posts has a
bigger width than its container, which makes it overflow from it. It
happens because of the `o_container_small` class set on the container,
which has a rule forcing its horizontal padding to 0, where it was
supposed to compensate the row margins.

This commit therefore fixes this issue by forcing the row horizontal
margins to 0 for screen sizes under the `MD` screen breakpoint, which is
approximately where the blogs layout becomes mobile.

This fix also has the advantage of improving the "List" mobile layout by
making it look similar to the "Grid" layout one, that is, with a bit of
space around the posts, instead of being glued to the viewport.

opw-4052853

Part-of: odoo/odoo#176421
Signed-off-by: Arthur Detroux (ard) <ard@odoo.com>