[FIX] sms: fix sms notification rendering with urls in tags.
A problem was left to be fixed in the previous PR:
What was to be considered the url and label was sometimes
incorrectly computed with html_keep_url.

As it is used in other places, we simply don't use it anymore, but
extract the needed part from it: how a link is safely rendered.

We also clean obsolete test parts and documentation.

Follow-up of #140102

Task-3502174

closes odoo/odoo#184290

X-original-commit: 9b5eed37707ccc8aaf4990154f74c31424ec4117
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>