[FIX] lunch: prevent error when deleting all Location records
When all Lunch location records are deleted, the LunchLocation JS component fails to build, raising an error when accessing the Lunch App.

Steps to Reproduce:
- Install the `lunch` module
- Go to Lunch > Configuration > Locations
- Delete all records
- Access the Lunch App again

Traceback: OwlError
Invalid props for component 'AutoComplete': 'value' is not a string

This error occurs at [1], where the given props are empty.

This commit will resolve the above error by displaying a text message instead of trying to build a Many2XAutocomplete with empty values.

opw-4104561

[1]- https://github.com/odoo/odoo/blob/503c0d6f8fad09f5de678d8692fc283feac1b061/addons/lunch/static/src/components/lunch_dashboard.xml#L61

Part-of: odoo/odoo#176135
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>