[FIX] l10n_it_edi: fix access error on purchase invoice
On the purchase invoice form, the Italian localization adds a button
that allows sending the tax EDI to the Italian administration.

This depends on the field l10n_it_edi_is_self_invoice that is a computed
field that will need access to the model account.report.line.

In order to reproduce the issue, change the access rights of Marc Demo
to Accounting: Billing, go to Accounting > Vendors > Bills and hit New.

This commit fixes the issue by adding a sudo at the right place.

closes odoo/odoo#182959

Source: post-Odoo.com migration feedback
X-original-commit: a14d74db4805db93874271fe262ceca4c0bfc61a
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Paul Morelle (pmo) <pmo@odoo.com>