[FIX] l10n_{bf,bj,cd,ci,cm,ga,km,ml,ne,sn,tg}: Fix Carryover for syscohada countries
For some syscohada countries some tax report where using carryover.
The carryover on the tax reports from these countries were not working.
The cause was one, the label of the expression used to target the applied carryover was not prefixed with _carryover_.
And two, is that it had no expression dedicated for carryover with label _applied_carryover and this was causing the report to miss the info tag on the line for the carryover.

task-4110461

Part-of: odoo/odoo#176489
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>