[FIX] web_editor: hide tags option on blogs (keep on posts)
When a blog cover is edited, a tag option is displayed while there is no
`tag_ids` inside the `blog.blog` model because the
`o_wblog_post_page_cover` appears on both the `blog.blog` covers and on
the `blog.post` covers.

This commit hides the blog tags option if the model of the cover is
not `blog.post`.

Steps to reproduce:
- Go to blogs.
- Select the "Travel" blog.
- Edit.
- Select the blog's cover.

=> A tag option could be specified.

opw-4107748

closes odoo/odoo#177077

X-original-commit: 10ef728a70f9d64db9e01b46d8f4893070a49bcf
Signed-off-by: Arthur Detroux (ard) <ard@odoo.com>
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>