[ADD] l10n_ar: invoice with document type B for foreign customer
Description of the issue/feature this PR addresses:
It is needed to be able to create Factura B for a foreign customer.

Legal reference:
Here is some useful documentation ﻿:

https://servicioscf.afip.gob.ar/publico/abc/ABCpaso2.aspx?id=8555970
https://www.afip.gob.ar/EspaciosdeDialogoInstitucional/documentos/Acta-15-Espacio-de-Dialogo-AFIP-Camaras-Empresariales-14-11.pdf

Current behavior before PR:
It is not possible to create Factura B for a foreign customer.

Desired behavior after PR is merged:
It is possible to create Factura B for a foreign customer.

Task Adhoc side: 30008
Task latam side: 913

closes odoo/odoo#185611

X-original-commit: d70cb30f63b2753bef02ef0f2ee99249ae4e0e4f
Related: odoo/enterprise#72948
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>