[FIX] sale_stock: compute sol reserved qty wizzard from all moves
Issue: the reserved quantity appearing on the SOL wizard is not well
behaved with respect to 2+ steps deliveries.

Steps to reproduce:

- Enable Multi-step routes in the settings
- Inventory > Configuration > Warehouse Management > Warehouses
- Enable delivery in 2 steps
- Create a storable product and put 1 unit in stock
- Create and confirm an SO for 1 unit
- Click on the chart icon next to the delivered qty

> the reserved qty is 0 + "No future availability" even though 1 unit
is reserved from stock

Cause of the issue:

The reserved qty displayed on the next to the delivered qty on the SOL
is the `qty_available_today` computed field of the SOL model. This field
is computed by summing the qties of the stock moves linked to the SOL:
https://github.com/odoo/odoo/blob/817186b7b896c9a415bd947baf189bf7f1bde321/addons/sale_stock/models/sale_order_line.py#L69
https://github.com/odoo/odoo/blob/817186b7b896c9a415bd947baf189bf7f1bde321/addons/sale_stock/models/sale_order_line.py#L80-L81

However, when you are not delivering in 1 step, the only delivery move
linkedto the SOL will be the move which destination is the customer
location. To be more precise, confirming the SO in a delivery in two
steps a stock move from the Output to the customer will be created and
linked to the SOL. During the action confirm of this move a procurment
will be run to generate a move from stock to the Output but the SOL will
not be referenced anymore so that the SOL will not be linked to it.

Note:

The behavior is different in saas-17.2 where each of the delivering move
is linked to the SOL so that the probably need to be adapted.

opw-3981935

closes odoo/odoo#176756

Signed-off-by: William Henrotin (whe) <whe@odoo.com>