[FIX] mass_mailing: revert backgorund filter in cover template change
Commit that introduced the issue https://github.com/odoo/odoo/pull/128108/commits/33a04876cad58f413455bdc406c1ad4c70d73e1b

Steps to reproduce the issue:
=============================
- Create a new mass mailing
- Add cover snippet
- Test the email
- Background image doesn't appear

Origin of the issue:
====================
After the chnage of the mentioned commit, we add a div with a black
background filter. It displays correclty in mass_mailing view since it
has opacity 0.5 but after the `inline` process of the template we
convert rgba colors to hex color with `rgbToHex` because some email
engines doesn't support rgba. As a result, we will have a black div that
covers the background image in the sent email.

Solution:
=========
Revert the change.

task-4070400

closes odoo/odoo#177174

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>