[FIX] website_forum: limit the number of pages displayed in the pager
When having a lot of forum posts, the pager can display up to 10 pages,
because the specified scope is the set to the same value as the number
of posts per page, which is 10. This causes layout issues in mobile view
in that case, because the pager overflows from the page.

This commit fixes that by setting the scope to 5 instead (like the Event
pager).

opw-4050389

closes odoo/odoo#177275

X-original-commit: cf8b0947ca364198bf564d6409e6209e0be87c2b
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>
Signed-off-by: Soukéina Bojabza (sobo) <sobo@odoo.com>