[FIX] website_forum: limit the number of pages displayed in the pager
When having a lot of forum posts, the pager can display up to 10 pages,
because the specified scope is the set to the same value as the number
of posts per page, which is 10. This causes layout issues in mobile view
in that case, because the pager overflows from the page.

This commit fixes that by setting the scope to 5 instead (like the Event
pager).

opw-4050389

closes odoo/odoo#177436

X-original-commit: 227a538469480d511cb6f8bc2d82530b579f3b3a
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>
Signed-off-by: Soukéina Bojabza (sobo) <sobo@odoo.com>