[FIX] purchase_stock: use the last purchase date for the supplier info
Problem: When viewing the replenishment information for a product, the last purchase date is set to the date from the oldest purchase order. The last_purchase_date field is set using the purchase order with the oldest date_order field. The last_purchase_date field should be set using the purchase order with the most recent date_order field.

Purpose: By ordering the purchase orders by date_order desc we will get the purchase order with the most recent date_order field. This will cause the supplier’s last purchase date to be the most recent one instead of the oldest one.

Steps to Reproduce on Runbot:

1. Create a storable product and create a reordering rule for this product
2. Create a purchase order for this product and set the Order Deadline to a date in the past and then confirm it
3. Create a purchase order for this product and set the Order Deadline to the current date and the confirm it
4. Navigate to this product’s replenishment information
5. Observe that the Last Purchase date is the earlier of the two dates from the purchase orders

opw-4113810

closes odoo/odoo#180518

X-original-commit: b0684bd80f31a808521cf5229b9ac8a621e39afc
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Richard Prosser (prri) <prri@odoo.com>