[IMP] mail: Include success, info, warning & danger type color schemes
in simple notification handler

This PR allows users to send simple notifications with 'success',
'info', 'warning', and 'danger' border color schemes. If the warning or type
parameter is not provided, it will default to the 'danger' type, as it
did before this PR, so it will not affect any previously written bus
notification call.

closes odoo/odoo#177323

Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>