[FIX] point_of_sale: handle inaccessible paid orders
Before this commit, when a user without access to certain orders
attempted to load paid orders, the process would fail and result in an
error. This was particularly problematic if a POS order was linked to a
sale order that the user did not have permission to access, causing the
entire loading process to halt.

This commit addresses the issue by filtering out paid orders that the
user cannot access.

opw-4108044

closes odoo/odoo#192702

X-original-commit: 070aefbd1f389c5cbea4481e94e339923bc46d04
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>