[FIX] sale_loyalty: fetch discounts & loyalty from parent company
Problem: When a user is on a branch, they are unable to view the
parent company's discounts and are unable to apply it to the sales
orders.

Purpose: The branch should have access to the discounts & loyalty
programs from the parent company.

Steps to Reproduce on Runbot:
1. Install sale_loyalty
2. Create a branch
3. On the branch company, observe that no discounts & loyalty
programs from the parennt company are displayed
4. Attempt to apply a coupon code from a discount belonging to
the parent company on a sales order and receive a ValidationError

opw-4079487

closes odoo/odoo#178135

Signed-off-by: Morgane Demesmaeker <edm@odoo.com>