[FIX] l10n_it_edi_ndd: community bug
The test "test_account_move_payment_method" was failing in community only because
the payment_state is paid if only l10n_it is installed but is in_payment if the
enterprise module is installed.

We add the paid status in the if statement to solve the issue.

Also adding a domain on search for matching number to get only move line that
are from company with an account_fiscal_country_id equals to "IT".

no task-id

closes odoo/odoo#178235

Signed-off-by: John Laterre (jol) <jol@odoo.com>