[FIX] hr_timesheet: prevent the creation of fsm task without partner
This commit's purpose is to prevent the creation of a task with quick
create with the 'withTaskHour' widget if the project is an fsm one. This
is done to prevent the creation of fsm task without a partner set (the
field is required).

enterprise follow-up :
https://github.com/odoo/enterprise/pull/68933/commits/eabc30f6f28f9387c66e9280c87d12763614ad6d

task - 3922511
version 17.0 - master

closes odoo/odoo#178342

Related: odoo/enterprise#68933
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>