[FIX] pos_mercado_pago: fix rejected payments
There is an issue where the Mercado Pago terminal
rejects a card payment, but Odoo accepts the
payment. This is caused by a quirk in the Mercado
Pago API, where the payment intent status is
"FINISHED" despite the payment status being
"rejected".

This PR fixes the issue by checking the payment
status explicitly by calling the API.

task-4097968

closes odoo/odoo#178386

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>