[FIX] sale: salesman unable to create global discount
If nobody gave a global discount (fixed amount/on SO) in a given company before,
the first user to give a global discount through the dedicated wizard
would create the dedicated Discount product for that company.

Nevertheless, it was not done in sudo, so the user would get a traceback if
he didn't have admin rights (for company update) or manager rights (for the
product creation).

opw-4048403

closes odoo/odoo#178361

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>