[FIX] account: fix argument order in function calls in tests
Currently `_compute_terms` is called with the arguments out of order
in some payment term related tests.
It does not cause any real problems for the tests since the
computation works anyway.
But the numbers are a bit weird (sign is 1000 and taxed amount bigger than untaxed amount).

After this commit the order of the arguments is corrected.

X-original-commit: eab5a6bb68e4c89d328855df69a4f837bce2ef41
Part-of: odoo/odoo#178344
Signed-off-by: Laurent Smet (las) <las@odoo.com>