[FIX] web_editor: remove dark mode style from convert inline
Steps to reproduce the issue:
=============================
- Switch to dark mode
- Go to any record with a chatter
- Open mail composer in log note
- Add a table
- Save
- Switch to light mode
- The table is still dark

Origin of the issue:
====================
Since in dark mode we add some styles to improve the display, those
styles are getting applied in convert_inline.

Solution:
=========
Skip the stylesheets for table that have color only for now since
skipping all styles can produce more issues.

We skip also the background-color and border color for light mode too
because they will not display the same in dark mode.

opw-4137267

closes odoo/odoo#179928

X-original-commit: 16ab970fa33119a4fc1a9c5dd249ca467ff3b2c8
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>