[FIX] web: display small and SVG product images correctly
Problem:
If the product image size is smaller than `$o-form-picture-size` or is an SVG without intrinsic size, it uses the original image size or displays as 0px in the case of SVGs without intrinsic size. The desired behavior is for the image to always display at `$o-form-picture-size`.

Steps to reproduce:

- Add a small or SVG image without intrinsic size to a product.
- The image will not display correctly.

opw-4119433

closes odoo/odoo#187642

X-original-commit: 0d5f272aa85671bae3a599a9791ad9b5bb7ad343
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>
Signed-off-by: Walid Sahli (wasa) <wasa@odoo.com>