[FIX] point_of_sale: Keep browser session debug mode on PoS opening
Prior to this commit:
If you were in debug=1 or debug=assets prior to opening the PoS,
it would be lost when opening the pos session from the backend.
This can be particularly problematic on mobile app as the user can
not modify the url manually.
Being in debug mode on the PoS is useful to have the debug window
and troubleshoot with the asset code

After this commit:
The PoS is opened with the same debug mode than the browser session

closes odoo/odoo#178569

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>