[FIX] account: Catching TaxClosingNonPostedDependingMovesError when posting a batch of moves
When we tried to confirm a batch of moves during the tax closing in a multi company setup with branches, an error was raised because the dependant closing moves should be posted before the parent move.

This error was only catched in the `action_post` method but not on the modal that validates the batch.

Original Commit: https://github.com/odoo/enterprise/pull/65645

opw-4148070

closes odoo/odoo#178601

Related: odoo/enterprise#69206
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>