[FIX] hr_recruitment: make sure `partner_name` is editable
Before this commit, the `partner_name` field defined in `hr.applicant`
is a related field with a custom inverse method. The problem is the
related will by default set the field as readonly even if there is a
custom inverse method and so the inverse method will never be used. It
means during the applicant generation once a user will apply to a job
in the website. The `partner_name` field will be displayed in the
description field of the applicant as it will be considered as a custom
field.
The solution could be to just add `readonly=False` to the field
definition but unfortunately it is not sufficient, because in the case,
the related will erase the custom inverse method to put its own inverse
method. Thus, that field has to be a computed field with a inverse
method to be sure the custom inverse method will be taken into account.

This commit updates the definition of `partner_name` field into a
non-stored computed field to be sure the inverse method defined be used.

Part-of: odoo/odoo#181446
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>