[FIX] web: list: do not close group on click on disabled pager
Have a grouped list view with a pager inside a group. On a slow
network, click on pager next. The pager gets disabled during the
update. Click again on next or previous. Before this commit, this
closed the opened group, because as the buttons are disabled, the
click isn't handled and thus bubble up to the group row.

This commit adds an empty handler on the div wrapping the Pager,
s.t. all clicks inside that div are always stopped and never close
the group.

Task~4150362

closes odoo/odoo#178977

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>