[FIX] website_sale_loyalty: Avoid collect sale orders confirmed in backend
The method sale_get_order() returns a confirmed order in case we confirm it
at backend. That produces that in cart() method could modify the loyalty
values in a confirmed order. This should never happen so, with this fix, the
method avoids to delete the information of the loyalty fields in non draft
orders.

closes odoo/odoo#176718

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>