[FIX] mail: Fallback to mail.thread for _process_attachments_for_post
Using composer in the mass mail mode is allowed on non mail.thread models.
This fix covers the case where one of the required methods (_process_attachments_for_post)
isn't available on the model that we try to send mail on.

Reproduce
---
- install website_slides
- "Add Context Action" on the "Elearning: Add Attendees to Course" template
- open an eLearning course -> Attendees list view, select one, click on Send Mail action
- fill required replyto -> send -> BUG: traceback

The action_send_mail is triggered on the mail.compose.message model.
During the process, it tries to call the `_process_attachments_for_post`
slide.channel.partner model doesn't have this method, hence an error

opw-4062613

closes odoo/odoo#178999

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>