[FIX] mail: link-preview on public channel invitations
before commit:

When link-preview is done on public channel invitation links it posts a guest
joined the channel​ message in the public channel.

after commit:

now we are ignoring public channel invitations links for link-preview, so it
does not post guest joined the channel

backported PR: #178932 which handle's ignore link-preview.

task-4083161

closes odoo/odoo#184483

X-original-commit: 3e353ba1d188ff0c1292b70c202e6116b81284e3
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>