[FIX] l10n_es_pos: generate normal invoice in es PoS
In a Spanish company if you don't have a certificate set on the
company, the PoS would generate a proforma invoice instead of a normal
invoice.

Steps to reproduce:
-------------------
* Install l10n_es_pos
* Remove the l10n_es_edi_facturae_certificate_id from the company
* Open PoS
* Make an order and invoice it
> Observation: The invoice generated is a proforma invoice instead of a
normal invoice

Why the fix:
------------
When generating the invoice data in `_generate_pdf_and_send_invoice` the
wizard would have `l10n_es_edi_facturae_checkbox_xml` checked by default
This leads to the generation of a specific invoice that will fail
because there are no certificate setup on the company. This will then
fallback to the proforma invoice.

opw-4074779

closes odoo/odoo#179107

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>