[IMP] account: adding filters
On the invoice tree view we have a possible search on multiple field at once.
While this is a good feature for the user as it makes a very practical search,
for large databases this is not performant at all, as we're searching on
multiple fields at once, some of them not being indexed. To allow for more
efficient searches, we added some specific search.

closes odoo/odoo#179125

Task: 4154313
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>