[FIX] html_editor: inserting an unsplittable element inside an empty list item
Before this commit:

When an unsplittable block element was inserted into an empty list item (li), it
was incorrectly placed directly inside the ul, as a sibling of the list item.

After this commit:

Now, when an unsplittable block element is inserted into an empty li, it is
correctly inserted within the list item itself, rather than inside the ul.

task-3550599

X-original-commit: f620a94
Part-of: odoo/odoo#181451
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>