[IMP] account: recipient bank selection
Partners can have multiple Bank accounts in an active database, some being even
added automatically on Bank reconciliation. But usually, only some specific
accounts that are used for outgoing payments will be marked as "trusted".

Currently, when you select a Vendor on a draft Vendor Bill, the first account of
the partner is taken, whether or not the account is trusted. This Pr will give
priority to the accounts that are trusted.

closes odoo/odoo#179365

Task: 4166740
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>