[FIX] website_event: allow editing of short description on Events page
Steps to reproduce the issue:

1. Navigate to the Events page.
2. Activate the Web Editor.
3. Attempt to edit the Event's short description.

Issue: The short description of the event was previously not editable
directly from the Events page. the reason behind that is they have used
`t-out` instead of `t-field` in template. `t-out` lacks the necessary
attributes to enable field updates.

Solution: This commit adds the necessary attributes to the short
description element, making it editable from the Event page.

Updated the tour functionality to test the latest changes of Event's
short description.

task-4014153

closes odoo/odoo#188772

X-original-commit: 6a1538d044cd38eb6a4917f0599d6256df1c80ae
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>
Signed-off-by: Vivek Panchal (panv) <panv@odoo.com>