[FIX] sale_stock: update DO demand from partially delivered SO
Steps to reproduce:

- Create and confirm an SO for 5 units of a consumable product.
- Validate the associated delivery.
- Return and validate the return.
- Return and validate the return of the return.
- Change the demand of the SO from 5 to 8 units.

> A delivery order for 8 units is created instead of 3.

Cause of the issue:

Changing the demand on the SO will trigger the
` _action_launch_stock_rule` to determine which procurement to trigger
in order to make the delivery. During this call the qty to deliver will
be computed by comparing the quantity handled by incoming and outgoing
moves  linked to the sol. However, the return of the return of an
outgoing move will not be flagged as an outgoing move since it does not
have a `location_final_id` since it was created as a return:
https://github.com/odoo/odoo/blob/4a050ec0efb929f7605856f490025f7921d39672/addons/sale_stock/models/sale_order_line.py#L266-L267
https://github.com/odoo/odoo/blob/4a050ec0efb929f7605856f490025f7921d39672/addons/sale_stock/models/sale_order_line.py#L288-L292

opw-4105394

closes odoo/odoo#181437

X-original-commit: cd116b5d3c81808cc44f15b68cf8279c8a00b3e7
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>