[IMP] core,base_import: allow importing properties fields
Currently, we cannot import properties fields, which makes them
less useful than others fields. This commit extend the import
feature for these fields.

We can now import every properties fields independently. Note, that
because each properties is linked to a definition via the parent record,
when we try to import properties of a different parent, it will be
ignore and we cannot change the properties definition with the import
mechanism.

task-4027103

closes odoo/odoo#174366

Related: odoo/enterprise#69287
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>