[IMP] web: list selection box on mobile
Ths List view allows to select one or multiple records. While doing so,
an indicator is displayed, showing how much records are selected and
providing actions to perform on them.

On regular computer's screen, the central part of the ControlPanel is
used to display this indicator and its related buttons, hiding the
SearchBar. On smaller screens, all those visual elements take a lot of
space, easily triggering an horizontal scrolling and taking an
additional line in the ControlPanel, which push the records a bit down
on the screen.

This commit reuses the CogMenu paradigm to group all the actions/buttons
in a single element. To avoid the records being pushed down, the
selection indicator is transformed into a floating "Toast" as an overlay
to the Navbar (similar to what was previously done to the Pager).

task-3336242

closes odoo/odoo#179877

Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>