[FIX] calendar: uninvited admins edit non-private events
This commit allows uninvited admins to edit non-private events
(public, confidential). It was a needed change since administrators
might have to change the event information of these types of events
on day to day. A re-work of the compute methods will be made on master
followed by the addition of the proper ACL rules for handling calendar
events.

task-4182649

closes odoo/odoo#180051

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>