[FIX] hw_drivers: get server url error in access point mode
If the IoT Box is in access point mode, it should mean that it has no internet connection.
Thus, we avoid returning the linked db url to avoid fetching it.

We used to check the access point mode with a `subprocess` call to `hostapd` service. We now check the IP,
as it is set to `10.11.12.1` when in AP mode.

closes odoo/odoo#180159

Task: 4182947
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>