[FIX] hr_holidays : cache issues
There were some cache inconsistencies when computing remaining leaves
on the dashboard or in the accrual scheduled actions due to commit
7ef0ad2.

The cache invalidation is necessary to avoid issues when the call to
process_accrual is made twice in a row, but the cache invalidation
was too broad.

This commit aims at only invalidating the part of the cache that
were causing issues when computing the remaining leaves.

opw-4167557
opw-4167582
opw-4175815
opw-4149134
opw-4256833

closes odoo/odoo#185677

X-original-commit: 79e0b0f6085b5e6c295a9bd8abc791700d536c15
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>