[FIX] barcodes: remove usage of jQuery in barcode service
On chrome mobile, when going in the pos self order and trying to do a
search on the products you would get a traceback. This was happening
because jQuery was not defined.

Steps to reproduce:
-------------------
* Install pos_self_order and barcodes
* Open the PoS on a chrome mobile device (use browserstack)
* Try to do a search on the products page
> Observation: You get a traceback

Why the fix:
------------
jQuery is not always defined, so we remove the use of jQuery.

opw-4107823

closes odoo/odoo#180163

Signed-off-by: William Henrotin (whe) <whe@odoo.com>