[FIX] base_automation: show `filter_domain` in debug mode
Versions
--------
- 17.0+

Steps
-----
1. Enter debug mode;
2. create an Automation Rule;
3. select Task as model;
4. set trigger to Stage is set to New;
5. set domain to a specific customer;
6. add send email as action;
7. create a task.

Issue
-----
Email is sent after task creation, regardless of the customer.

Cause
-----
The triggers added to 0a744accc2aa automatically compute the
`filter_domain` value, and hide the field in view. With debug mode
enabled, the `filter_pre_domain` field is still visible & editable.

The newly added triggers are applied on both create & update, while
`filter_pre_domain` is only applied on update. This leads to confusion
when clients add a domain which appears to be ignored, as the selected
trigger is immediately hit on creation.

Solution
--------
1. Specify in the help string that `filter_pre_domain` is ignored on
   creation.
2. When entering debug mode, also show the `filter_domain` field,
   allowing users to further modify the domain computed by the selected
   trigger, and helping to distinguish itself from `filter_pre_domain`.

opw-3928082

closes odoo/odoo#180209

Signed-off-by: Levi Siuzdak <sile@odoo.com>