[FIX] l10n_it_edi: prevent generation of XML if document type is not set
Issue:
When creating a vendor bill for a non EU partner, setting the tax to
"22% G RC (Goods)" and generating the XML file, checking the generated
XML file gives the error: "Tipo Documento must not be empty."

Steps to reproduce:
- create a vendor bill with a non EU partner
- add a product with tax "22% G RC (Goods)"
- post the bill
- click Send Tax Integration
- check the generated XML file using an online checker like
www.fatturacheck.it
- it will show the error "Tipo Documento must not be empty."

Solution:
- The document type is usually computed. However, in this specific case,
it cannot be computed because it does not meet any of the conditions in
the computation logic. Additionally, there is no safeguard to prevent
the generation of the XML file if the document type is not set.
- A check has been added to prevent XML generation when the document
type is not set, ensuring the document type is always correctly provided
before XML generation.

opw-4182531

closes odoo/odoo#180364

Signed-off-by: Kawtar Drissi El Bouzaidi (kdeb) <kdeb@odoo.com>