[IMP] stock: allow custom domain in SML X2many field
The custom stock.move.line's x2many field computes a domain
automatically for the SML that can be selected from the stock.move
records (e.g. when trying to find a specific lot, etc. to fulfill a
demand).

This commit makes it possible to include the domain defined on the field
in the final constructed domain, making it possible to somewhat
customize this behaviour without the need for a js override.

Task-4116000

closes odoo/odoo#180421

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>