[FIX] sale_mrp: Incorrect move kit qty after change on SOL
We can have a user that has enough rights to generate sales orders,
but not enough right to check a BoM.
When creating a new sale order for a kit, everything works fine.

Before this commit, if we updated the quantity on the sale
order line, the quantity on the moves would become incorrect.

After this commit, the BoM is always found in order to correctly
recompute the quantity on the moves.

closes odoo/odoo#186567

X-original-commit: 738df038de0c94d34e1b6c2bb858c235833c6f52
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Pedro Nogueira (pno) <pno@odoo.com>