[FIX] website: allow `s_quadrant` to have the grid mode options
In commit [1], the "Quadrant" snippet has been added. Given its look,
this snippet is only allowed to be in grid mode (like "Masonry"), as
the normal mode would make no sense for it. As such, the "Layout" option
allowing both grid and normal modes has been disabled for this snippet.

However, the option for grid-only snippets has not been activated, which
results in the "Add Element" and "Spacing (Y, X)" grid options being
unavailable. We can therefore not directly add new elements in the grid
and we cannot control the grid gaps, which could have been useful to
customize the snippet more.

This commit adds the grid only "Layout" options to this snippet.

[1]: https://github.com/odoo/odoo/commit/efaec2c7d7575961751b264e5c7b3dc84f4c64c9

Related to task-4105045

Part-of: odoo/odoo#180515
Related: odoo/design-themes#915
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>