[IMP] website: extract "Vertical Alignment" option to be t-callable
This commit extracts the "Vertical Alignment" option in order for it
to be added with t-calls. This was mostly needed to allow having the
option without the `o_we_sublevel_1` class on it, which made it
impossible to have it as a main option.

This class is now added only if the `indent` variable is set to `True`
when t-calling it.

Related to task-4094408

Part-of: odoo/odoo#180515
Related: odoo/design-themes#915
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>