[FIX] website: remove flex display from `s_pricelist_cafe` snippet
In commit [1], the "Pricelist Cafe" snippet has been added. The columns
of this snippet have their `display` property set to `flex`, in order to
allow using the `justify-content-*` classes to align their content. With
this, they also have the "Vertical Alignment" option, allowing to change
this alignment.

However, using `display: flex` is not a good thing, because it can cause
issues with images. Indeed, with flex, the children tries to fill the
horizontal space, making the images have the width of the column instead
of their real size.

This commit removes "flex" from this snippet columns, and therefore also
removes the added "Vertical Alignment" option. In order to keep the same
visual with the image centered, this snippet will now use the good old
"Vertical Alignment" option which is applied on rows, to align the price
columns with each other.

This commit also fixes the string of the snippet by adding a capital
letter, to be consistent with the other snippets.

Steps to reproduce:
- Drop the "Pricelist Cafe" snippet (from the "Content" category).
- Drop an "Image" inner snippet in one of the columns and upload a small
image.
=> The image is really big, while it should be smaller.

[1]: https://github.com/odoo/odoo/commit/2c0dd5447121e5e032b65c7a1eeb7aa8404acbe6

Related to task-4094408

Part-of: odoo/odoo#180515
Related: odoo/design-themes#915
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>