[FIX] hr_holidays: make `_compute_number_of_days` consistent
Steps to reproduce:
Make a regular allocation for a time off type that can be taken
in hours for an employee that has a working schedule other than
standard 40h/week, and click save.

The number of hours is updated to an amount different to what was
entered.

The issue is that https://github.com/odoo/odoo/pull/163711 changes
the computation of number of hours to only work with the standard
`HOURS_PER_DAY` without considering the employee's schedule,
meanwhile `_compute_number_of_days` still takes the employee's
schedule into account, causing an inconsistency leading to the
bug.

This commit, makes the computation of days and hours consistent,
so at least the amount entered is kept.

opw-4190998

closes odoo/odoo#183929

X-original-commit: b35e3ca8fb7e149f0587325ab7934e57e0b30891
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>