[FIX] web: misplaced search panel tooltip
Since https://github.com/odoo/odoo/pull/121963, there are tooltips on
search panel items when the user hovers them. However, it is not very
convenient since the user needs to hover precisely the span which
contains the item name for the tooltip to pop up. This commit simply
moves up the data-tooltip attribute in the elements hierarchy so that
it will show for the whole button instead of only the title. This also
fixes a bug with the tooltip position in the charts of account search
panel as a side effect.

task-3917084

closes odoo/odoo#181880

X-original-commit: f3c3c7dd019e3cee260d0609d8ad428519bf0a22
Related: odoo/enterprise#70924
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>
Signed-off-by: Julien Carion (juca) <juca@odoo.com>