[FIX] web: misplaced search panel tooltip
Since https://github.com/odoo/odoo/pull/121963, there are tooltips on
search panel items when the user hovers them. However, it is not very
convenient since the user needs to hover precisely the span which
contains the item name for the tooltip to pop up. This commit simply
moves up the data-tooltip attribute in the elements hierarchy so that
it will show for the whole button instead of only the title. This also
fixes a bug with the tooltip position in the charts of account search
panel as a side effect.

task-3917084

closes odoo/odoo#181608

X-original-commit: 32f357741e4de6ba2d335b5fccfcf3938026d00f
Related: odoo/enterprise#70795
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>
Signed-off-by: Julien Carion (juca) <juca@odoo.com>