[IMP] html_editor, mail: enable embedded components
Embedded components are Owl components mounted inside the HtmlField value
managed by the editor. Their content is dynamic and only metadata useful for
their rendering is saved in the HtmlField value. This feature allows high
flexibility in the content that can be inserted in the editor (i.e. can
dynamically render iframes which don't need to be saved, since they would be
sanitized for security reasons).

Embedded components is the renewed implementation replacing knowledge Behaviors,
and have the same purpose (mount Owl components).

This is the introductory commit for embedded components in the HtmlField. It
adds an `embedded_components` option which is true if unspecified in standard
HtmlField. It is forcibly set to `false` in HtmlMailField since embedded
components can only be rendered inside Odoo.

This commit adds the ability to mount embedded components in the HtmlViewer.
This will be useful i.e. in knowledge locked mode (readonly), in the history
dialog, in website_knowledge (public view of an article), and in the knowledge
template dialog.

task-3672747

Part-of: odoo/odoo#173997
Related: odoo/enterprise#67083
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>