[FIX] html_editor: fix server value comparison (html field)
ISSUES:

Server-side, HTML is stored as a string which can have a different format than
what the current browser returns through outerHTML or innerHTML, notably because
of HTML entities. This can cause issues in the HtmlField when values extracted
from the DOM are compared to the value given by the server, since the content
will be re-rendered even though unnecessary, causing flickering and unnecessary
calls.

`stripHistoryIds` was using a regex to do replacement in a html value, and that
regex could match values in text nodes, while it was intended to only match
element attributes.

In `updateValue` of the HtmlField, the value passed to the Record instance was
always stripped of its history ids, and therefore the stale detection mechanism
of the Collaboration plugin could never work.

FIXES:

This commit introduces a `normalizeHTML` function which will take html as a
string and return a normalized html string which can always be compared with
other strings normalized through the same method.

`stripHistoryIds` now works on a DOM tree, and not on a string, to avoid
misfires.

`updateValue` always sends raw values to the Record. Other post processing are
handled by the HtmlField internally.

task-3672747

Part-of: odoo/odoo#173997
Related: odoo/enterprise#67083
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>