[IMP] html_editor: add the file embedded component
Move the file_behavior from knowledge to html_editor as an embedded component.
This simple component allows to display an attachment inside the HtmlField and
notably offers access to the FileViewer. This makes it convenient to integrate a
file next to an explanation.

task-3672747

Part-of: odoo/odoo#173997
Related: odoo/enterprise#67083
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>