[IMP] website: fine-tune `theme_nano` color palette
This commit fine-tunes the `primary` and `dark` color of the `theme_nano`.

As part of the `theme_nano` redesign, the color palette has been adjusted
to a `high-contrast` style. The colors remains in the same tones (orange
and dark grey) but are more saturated, to emphasis the content of the
theme.

task-4178081
part of task-4177975

closes odoo/odoo#181369

Related: odoo/design-themes#935
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>