[IMP] mail: respective thread icon in suggestion list
Before this commit:

Channels and threads shared the same icon in the suggestion list and mentioned
messages, making it difficult to visually differentiate between them.

After this commit:

This commit aims to display separate icons for threads and channels in the
suggestion list and mentioned messages, improving clarity and  differentiation
between the two.

Task - 4203558

closes odoo/odoo#182280

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>