[FIX] mail: outlook 365 warning not hiding mail
This issue was already fixed in c9e94bc78, but it's not working properly.
The explanation on that commit is correct, however the resulting CSS from that
SCSS is not a valid one:

```css
.o_Message_content *:not(li):nottable(li div)[align="left"][width="100%"] {
  float: none;
}
```

The result is actually weird, which may be due to an issue in the SCSS compiler.

In any case, this commit fixes the SCSS, so that it's compiled like this:

```css
.o_Message_content table[align="left"][width="100%"] {
  float: none;
}
```

Which actually makes more sense, IMHO.

closes odoo/odoo#181099

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>