[FIX] web: allow drag and drop elements in mass_mailing
Issue:
======
We can't move elements in mass mailing.

Steps to reproduce the issue:
=============================
- Create a new mass mailing with a template that have some blocks
- Try to drag one of the blocks to another position
- When you click on the button to drag, the right sidebar refreches

Origin of the issue:
====================
The issue was first introduced by [1].

The problem when we have the wysiwyg inside an iframe
`document.activeElement` will be equal to the iframe element and the
`ev.target` is the button of dragging. In reality the button is inside
the iframe but the content of the iframe is like this `<iframe>
document <iframe>` so contains returns false.

Solution:
=========
When the `activeElement` is an iframe we search inside its
`contentDocument`.

opw-4196067

[1]: https://github.com/odoo/odoo/commit/ef97eaf9876bcf1712d4a6bfb91307875b0034ac

closes odoo/odoo#181143

Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>