[FIX] calendar: copy attachment raw data when sending mail
Steps to reproduce:
- Install Website and Appointments modules
- Add a sample attachment to the email template “Appointment: Attendee Invitation”
- Go to the website and book an appointment
- Check the email in the Emails technical menu

Issues:
The attachment size is 0, this is due to `copy_data` not being
implemented in `ir_attachment.py`. Because of this we are not copying
the raw attribute of the attachments, which is why we end up with an
empty file.

In order to solve this we just use the copy function which will copy the
raw attribute.

opw-4082806

closes odoo/odoo#181125

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>