[FIX] website: replace background-position of s_parallax style attribute
The `new_page_template_services_s_parallax` template wrongly adds the
`background-position` of the `style` attribute.

This commit fixes this definition so that the old `background-position`
is replaced.

Steps to reproduce:
- Add a new page
- In the Service tab, pick a template that contains a parallax block (at
the time of writing, this is the case of the second one in the first
column)
- Inspect the parallax block

=> The `background-position` was defined twice.

This was found with the tests added in [1].

[1]: https://github.com/odoo/design-themes/pull/933

task-4206845

closes odoo/odoo#181194

Related: odoo/design-themes#933
Signed-off-by: Alice Gaudon (agau) <agau@odoo.com>