[FIX] hr_presence: fix traceback in employee list view
Steps:
- Install the hr_presence module
- open employee list view
- Click on the `Create a Time Off` from Presence Control Menu

Description of the issue/feature this PR addresses:
In the HR module, When selecting multiple employees from the list view and
clicking `Create a Time Off` in the Presence Control menu, a traceback error
occurs.

Cause:
The error is caused by opening the hr.leave form view, which contains the
employee_id field that allows the selection of only one employee. Since multiple
employees are selected, a singleton error is triggered.

Fix:
This PR resolves the issue by updating the wizard of the model
`hr.leave.generate.multi.wizard`, enabling the creation of time off for multiple
employees.

task-4207392

closes odoo/odoo#181199

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>