[FIX] account: audit_trail activated without any companies with it
To reproduce:
Go to Accounting/Invoicing, Customers and create a new Customer.
Delete it.
You get an error talking about audit trail while you don't have
it activated.

The issue is that we block based on customer/supplier rank when
there is no company on the partner.

We now check if we should block the deletion of a message of a
partner if this partner has at least one journal entry with a
company that has audit trail activated.

closes odoo/odoo#187455

X-original-commit: 1133c79914a26ef1539f61065c03038755170b60
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>