[FIX] account: fix traceback when removing lock date with exception
In commit d8bf86fec7b9a0a3f6e0aef2c4c124bc81bfb898 we allowed removing
lock dates with an exception.
But when computing the user lock date based on (parent) company lock
dates and user exceptions (`_get_user_lock_date`) we still assume
that it is not possible.

This is fixed in this commit.

Reproduce traceback
  1. Set Sale Lock Date to 2024/08/31
  2. Remove Sale Lock Date with an exception for me forever
  3. Create an invoice and try to set the invoice date to a date
     before 2024/08/31

closes odoo/odoo#181210

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>