[FIX] website: remove flex display from `s_carousel_intro` snippet
Like the other commits of this PR, this commit removes the fact that the
columns are in `display: flex` in the new `s_carousel_intro` snippet
(added in commit [1]). It then adapts the snippet in order to look as
close as possible as how it was before, while making sure no content
overflows out of the snippet.

[1]: https://github.com/odoo/odoo/commit/454d743ed631c068b7b0bc72e4fd81b225a5ac52

Related to task-4077628

closes odoo/odoo#180515

Related: odoo/design-themes#915
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>