[FIX] hr: redirect to public employee if no employee rights
Steps
-----
- Install Employees app.
- Have a user with "Settings" Administration rights and no "Employees"
rights.
- Open an employee profile, open "Work Information" tab and click on an
employee in the organizational tree.
** Access error for `message_main_attachment_id` field **

Cause
-----
We are checking if the user has access rights on the employee
https://github.com/odoo/odoo/blob/c58253e6c9c7a92ced59beaf296427311e756237/addons/hr/models/hr_employee.py#L294
and redirecting to the private employee if that's the case.

However, this does't guarantee that the user has the `hr.group_hr_user`
group necessary to access some fields that were added with
`groups="hr.group_hr_user"`
https://github.com/odoo/odoo/blob/c58253e6c9c7a92ced59beaf296427311e756237/addons/hr/models/hr_employee.py#L122

opw-4096073

closes odoo/odoo#183600

X-original-commit: d0828eecf60f7c8622d6875b8651eb663bc7d695
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>