[FIX] website: fix multiple buttons vertical spacing on small devices
This commit introduces a fix for the snippets that include at least two
buttons in their layout.

Prior to this commit, the buttons were simply defined within the `<p>`
tag, due to a lack of testing corner case scenarios. This implementation
was actually wrong, because as soon as you add `n+1` button within the
editor, it'll apply a `mb-2` on each button to space them correctly on
smaller devices.

To fix this issue, we simply replicate the behaviour of the editor by
adding the class on the snippets that match this scenario.

task-4210852
part of task-4077427

closes odoo/odoo#181371

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>