[FIX] web: reaction blocked on mobile view
Before this commit:
In mobile view, when the 'Add a reaction' action is clicked, the EmojiPicker
opens, but the native keyboard also appears due to the focus being on the
input field. This causes the keyboard to block the view of the EmojiPicker.

After this commit:
Now, in mobile view, the input search bar will not be automatically focused
when the EmojiPicker opens, preventing the native keyboard from appearing
and obstructing the EmojiPicker.

task-4208499

closes odoo/odoo#181341

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>