[FIX] website: fix Facebook snippet
The goal of this commit is to solve three bugs introduced by [1].

The first one can be observed thanks to the `/website/demo/snippets`
page:
- Go on the `/website/demo/snippets` page.
- Try to enter in edit mode.

-> The system never enters in edit mode.

The problem is that the Facebook snippet on this page does not have the
`href` data attribute. Due to it, the public widget is started without
creating `this.resizeObserver`. When the user tries to enter in edit
mode, the `FacebookPageWidget` public widget is destroyed and the system
tries to disconnect `this.resizeObserver`. As it is not defined, the
system fails in the `destroy()` method.

Steps to reproduce the second bug:
- Add an "Image-Text" snippet on the website.
- Add a "Text-Image" snippet under the "Image-Text" one.
- Save and edit.
- Drop a "Facebook" snippet on the "Image-Text" snippet.
- Move the "Facebook" snippet on the "Text-Image" snippet.
- Click on the undo button.

-> Two iframes are on the Facebook snippet.

The problem is that since [1], the editor observer is not deactivated
when removing the iframe in the `destroy()` method of the
`FacebookPageWidget` widget.

Finally, this commit also adds back the `frameborder` and
`allowTransparency` attributes on the iframe (removed due to [1]) and
sets the width and height in the style attribute. Although the
`frameborder` and `allowTransparency` attributes are deprecated, this
commit adds them back in case a potential customization relies on it.

[1]: https://github.com/odoo/odoo/commit/707dd91d1e83342a328e50b5ad38e6ec1064ec25

task-4220972

closes odoo/odoo#182134

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>