[FIX] project: task description generation
The task description of a task generated after receiving an email from an email alias
should be the body of the email (from the message thread). We differentiate this
case from the case where the task is generated in another way (e.g. manually or
triggered by another module), in which we should not populate the task description.

related-https://github.com/odoo/odoo/pull/108360

task-4207145

closes odoo/odoo#181806

X-original-commit: 7b2095cd94d1cffe6a9b738d1d9e677e9eb4eac3
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>