[FIX] {sale_,}stock: see delivered quantities for interco sales
Currently, we only check delivered quantities in a Sale Order based on
the `usage` of the destination location of the related delivery. This
means that in the case of Inter-company transactions, we won't consider
them as deliveries, as its delivery location will be 'Inter-Company
Transit', which itself is a 'transit' location.

Task-4207132

closes odoo/odoo#181922

X-original-commit: 444dfdfe2eae3bf3ecfba524af22f9ad5297ef24
Related: odoo/enterprise#70960
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>