[FIX] stock: reserve on inter-company transit location
It used to be forbidden to reserve on inter-company transit, as this
could raise issues with transfered lots that would belong to a company
being processed in another company.

Since now lots can be made company-less for the purpose of being
transferable between companies, it would make sense to enable the
reservation again, and allow to draw from the inter-company location as
if it was a normal transit location.

Task-4207078

closes odoo/odoo#181545

Related: odoo/enterprise#70737
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>