[FIX] account: allow more flexible tax search again
After [this refactoring] a regression occurred on the search for taxes
on invoices/bills/journal entries. Before, there was a custom search
function that allowed to search with a term like "21M" and get the taxes
"21% M ...". After the refactoring this was broken.

This commit fixes that again by intercepting the `_search` call on
`account.tax` when we search for the `name` with `ilike` or `like`.

[this refactoring]: https://github.com/odoo/odoo/commit/7fc8794655840961dab1bd80a4bfad3b30953448

closes odoo/odoo#182020

X-original-commit: 1afac96c43d97209d2d66836e750380363e6e8a7
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>