[FIX] account, l10n_ar: account: fiscal pos differents bill/delivery addresses
We did not compute fiscal position correctly regarding
billing/delivery addresses combinations.
Before this commit, we used the delivery to compute the
fiscal position, except for invoices with no delivery or in the same
country as the company, but it is not sufficient for several use cases.
With this commit we check multiple combinations of
EU/extra-EU partners with/without VAT.

In this fw-port we had to override `_get_fpos_ranking_functions` to
handle the forced value of `delivery_country` and `vat_valid`

Change has been made regarding this spreadsheet:
https://docs.google.com/spreadsheets/d/1PDlXpsgRiGvYvrUKguJqWPVl-g-7H6YqJiWoRNgWCYI/edit?gid=0#gid=0

opw-4072691

X-original-commit: e3acf5937cd7a65087a0178fadd666dead2b99db
Part-of: odoo/odoo#183789
Related: odoo/enterprise#72019
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>