[FIX] microsoft_calendar: consider system parameter on event fetch
Before:
All event from last year to the next 2 years were fetched. This might
cause timeout depending on the amount of events.

After this commit:
Reuse the system parameter to allow limiting to a set value

Note: This does not need to be done on google as the fetch limit is
already present, see:
https://github.com/odoo/odoo/pull/66250/files#diff-f1bbd37c3355f798d3f2d89ccc00c778aa4faa4c6709b6d66cc9650ad0d553b6

opw-4077113

closes odoo/odoo#181613

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>