[FIX] mrp: Fix Manufacture from BoM
The Manufacture button added in the BoM Overview prepares a new
manufacturing order using the bom.
However this has a drawback of changing the state to 'in progress'.
This is not the case when creating a new manufacturing order
and then selecting the same bom.
The origin is within '_onchange_producing / _set_qty_producing'
and the way the record is initialized:
- from 'Manufacturing orders / New' all fields have their default values,
- from 'BoM Overview / Manufacture' the bom_id and related fields are set.

closes odoo/odoo#181614

Signed-off-by: William Henrotin (whe) <whe@odoo.com>