[FIX] mail: livechat visitor should not see their country flag
Before this commit, when a visitor posts a message in livechat,
the header was showing their country flag.

Steps to reproduce:
- log in as Mitchell Admin as available operator
- log in as Joel Willis (portal user, on another browser context)
- open a livechat as Joel and send a message
=> Country of joel "USA" is visible in header

This happens because it was always showing `anonymous_country`
when available, which is the visitor country.

This commit fixes the issue by limiting the showing of the country
flag to the operator of livechat conversation.

closes odoo/odoo#181684

Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>