[FIX] mail: defaults batch size to 50 in composer
Following odoo/odoo@4c36d2d747b5f801917d74e9f0f83770e54fa3eb, the default `batch_size` for generating mails was
reduced from `500` to `50`. Rationale is that inserting content is
quite fast and it is better to commit often and smaller batches to
avoid too much memory usage notably.

But one instance seems to have been missed, during the generation of
mails in the wizard, which is used by the mass-mailing cron. To address
this, we are just changing `self.batch_size` in the wizard from `500`
to `50`. As the variable may have been overridden in custom code we
keep the static variable.

Task-3084943: Event: Improve communication scheduler scalability

Part-of: odoo/odoo#177334
Related: odoo/enterprise#68658
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: pivi-odoo <pivi@odoo.com>