[FIX] account_edi_ubl_cii,*: Line/Description full value
*: l10n_ro_edi

When a user generate a UBL XML that inherits from ubl 2.0, the
Item/Description xml node will only contain the information about the
product name, without any description. This commit aims at restoring the
behavior before saas-17.4 and gives the full description on that XML
value.

To make sure it keeps this new behavior, a test was modified in l10n_ro_edi
(the ticket origin) to make sure the Description display the `line.name`

The change was previously introduced in
https://github.com/odoo/odoo/commit/4f325ef620263c27e095eb49026a677ac617a0ee

Similar fix to l10n_es_edi_facturae for reference:
https://github.com/odoo/odoo/pull/180315

opw-4213014

closes odoo/odoo#181877

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>