[IMP] web: clarify m2m edit_tags option behavior
This commit slightly changes the behavior of the m2m edit_tags option so
that it will only be effective in form view (like tag color edition) and
also so that it will override color edition when color_field is set. It
also adds visual feedback with the pointer style when the edit_tags
option is set.

task-3961350

closes odoo/odoo#181799

Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>