[FIX] spreadsheet: Speed up tests bypassing canvas drawing
Following the same optimization as in [1], we can bypass the canvas
drawing of spreadsheet in the tests (as long as) we don't need it.
This revision adds a generic patch that will disable the major part of
the grid drawing of a spreadsheet while letting it enabled in debug mode
for.. debug reasons.

Currently, the time spent in drawGrid is around 12~14% (depending on the
screen size) when executing spreadsheet edition related tests (see
task). This revision makes it drop to 0.5%.

[1]: https://github.com/odoo/o-spreadsheet/commit/d3036999989bb0d611ff8c3a05c7e0c2aa60dd49

task-4214049

closes odoo/odoo#181931

X-original-commit: e93e4ea1abe5464b4e68e97d1b95b27149557730
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>