[FIX] mail: message subject more distinct from body text content
Before this commit, when a message has a subject different from
chatter name e.g. by renaming the record while some messages have
been posted prior to rename, the subject was visible in prefix
of message text content.

The showing of subject is good, but currently the style is almost
the same as the text content, and this is put inline before the
text content, so the subject looks like actual text content.

This commit improves the style so the subject is clearly different
from the text content of the message, when the subject is displayed
on UI.

closes odoo/odoo#181895

X-original-commit: 58ec7b05f688ba7897734e56145a23ccf672a770
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>