[FIX] web: kanban: color cards even if field not in arch
Kanban cards can be colored (their left border) dynamically by
using the root attribute `highlight_color="color"`, where `color`
is the name of the integer field to use as color field. However,
before this commit, this required the given `color` to be also
defined in the arch, otherwise it wouldn't have been fetched, and
no card would be colored.

With this commit, we automatically fetch the highlight_color field.

This fixes a bug in planning where colors weren't displayed.

Task~4215979

closes odoo/odoo#181830

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>