[FIX] point_of_sale: static uuid tracebacks in demo orders
In this commit:
=================
Fix traceback when deleting demo records in POS.
When deleting some demo orders in the POS, a traceback was
raised because the lines of the orders did not had an order
linked to it. This was because they did not have an uuid.
We now have default uuid values for orders, order lines and
pos payments.

closes odoo/odoo#181858

Task: 4212901
Related: odoo/enterprise#70925
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>