[FIX] pos_restaurant: retain customer count information
Currently, if you change the number of customers in the restaurant, the
final order will show a 0 as the number of guests.

Steps to reproduce:
-------------------
* Go to the **Point of sale** App
* Open restaurant session
* Select a table, add items on the order
* Change the guest count
* Send order to kitchen
> Observation: Guest count resets to 1
* Pay the order
* Select the order outside the session
> Observation: Guest count is 0

Why the fix:
------------
After the refactoring https://github.com/odoo/odoo/commit/2a5f1abf2e98ee09fa7a912b87d71879b5ff260b,
we use the same field structure as in python. Therefore, when we load/create
a pos order we use `this.customerCount` is undefined since in python the field
is called `customer_count`.
https://github.com/odoo/odoo/blob/ae0b903b753decab1592df09a1e3cf1baa55f566/addons/pos_restaurant/static/src/overrides/models/pos_order.js#L8

By changing `this.customerCount` for `this.customer_count` we not only retreive
the correct field when loading an order in the session but we also send the
correct information to the backend.

opw-4182891

closes odoo/odoo#182590

X-original-commit: 7f165dd84b10f1d830e7c2bae8b79436d2f040b6
Related: odoo/enterprise#71362
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>