[IMP] l10n_eu_oss: OSS on EU branch
The aim of this commit is to allow non EU countries to have an EU branch
with an EU VAT number that can have both the taxes for that specific
country and the OSS taxes mapped from that country to the other EU
countries.

For that branch to be able to see the OSS tax report, we have to be able
to set the fiscal_country_id as an EU one.

It also needs to be able to instantiate the EU tax from that country.
To allow that, we need a foreign VAT fiscal position. In order to do
that, we had to remove the constraint that prevent to have both a
foreign vat fpos and the fiscal_country_id to be set on the same
country and same vat number.

task: None

Global context:
TSB, LZZ, THBE and me (BIB) had a very long meeting to find out what
would be the better the solution to allow a UK company to use the
OSS tax system with Belgium as its entry point.

The UK company has a wharehouse in Belgium and thus needs to report
anything coming out of that wharehouse to the on the Belgian VAT report
and the "Belgian" OSS tax report.
It also needs to invoice any Belgian with the Belgian taxes and VAT (and
thus can't use OSS for that one).
Some of their product might be purely Belgian and thus only have a
Belgian tax set on them. We thus need to support both the mapping from
the motherland UK and from Belgium.

The following are required:
- Creating the Belgian VAT taxes.
- Creating the OSS taxes as they would be for a Belgian company.
- Create the fiscal position mapping from Belgian tax to EU taxes.
- Create the fiscal position mapping from UK to EU taxes.

The main goal of the task is to:
- be able to instantiate OSS on a specific branch: we go from the
  current one up and we instantiate OSS to the first one with a TAX ID.
- we should be able to use the classic OSS tag as usually.
- be able to report those numbers in the OSS tax report.

Related: odoo/enterprise#71006
Related: odoo/enterprise#71320
X-original-commit: 1858eb5bc938bad57bb52775f94a0ef8c3a47a97
Part-of: odoo/odoo#182508
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>