[FIX] mail: reduce chances to accidentally start a call in mobile
Button to close conversation and start a call are too small and
close to each other. This means attempt to close the conversation
can lead to accidentally starting a call, which scare all members
of the conversation.

This commit increases the size of these buttons and show borders
so that it is less likely to click on start a call when the intent
is to close the conversation.

closes odoo/odoo#182099

X-original-commit: 251ef3edcc7e0894683acb7c69bfa437a024110f
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>