[FIX] spreadsheet_dashboard_pos_hr: Fix dashboard dependency
We recently added some model dependencies to determine if a dashboard
should display its sample data if the database did not have enough
information worth showing.

The pos_hr dashboard was depending on sale orders by mistake. It's
supposed to depend on `pos.order` and to a certain extent
`report.pos.order`.

Task-4220370

closes odoo/odoo#182169

Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>