[FIX] hr_timesheet: avoid calling several times same rpc in list view
Before this commit, due to the commit a5e27d8f, when the user goes to list
view of timesheets, he could be disconnect due to a many rpc calls made
to get the projects for which we only allow to create and edit task
(because some fields could be required).

This commit moves most of changes made in a5e27d8f in enterprise since
only with hr_timesheet module installed, we will never get a project
for which we don't want to allow the quick create on task_id field.

Impacted versions: 17.0+

task-4221621

closes odoo/odoo#182303

X-original-commit: a5e27d8f
Related: odoo/enterprise#71183
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>