[FIX] portal: portal users can't create api keys
This issue occurs only in frontend debug (which I initially missed but
odony caught), the problem is that the `InputConfirmationDialog` used
for the creation dialog has a required `onInput` callback in its props
but none is provided by the "new api key" button case. Props are (I
assume) only validated in debug mode, so that's the only situation in
which the error appears... and blocks users.

While `onInput` seems like a useful hooks for input validation and
stuff, it doesn't seem necessary to mandate it, and the code
specifically checks if it's unchecked (and ignores it in that case).

OPW-4194305

closes odoo/odoo#182285

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>