[FIX] web: kanban with sample data: no flicker when quick create
Have a grouped kanban view with existing groups but no records s.t.
sample records are displayed (e.g. in CRM pipeline with a default
filter, or in Project in a new project). Click on the "+" icon in
a column to quick create a record. The first time, it's fine.
Click on the menu again to relaunch the action (do not reload the
webclient), and do the same: there's a flickering as sample records
briefly appear "as real records" (i.e. they're not displayed as
ghosts), before completely desappearing. This is even more obvious
on a slow network.

The difference between the first time and the others is that the
form view used in the quick create must be loaded the first time,
and is in cache afterwards.

When we click on the "+" icon, the following happens:
 - we remove sample records from the groups and we enable the
   quick create in a column => triggers a rendering of the
   KanbanRenderer
 - in the same tick, we toggle the useSampleModel flag on the model
   => triggers a rendering of the Controller

After its rendering, the Controller no longer has the classname
`o_view_sample_data` which ensures that sample records are
displayed as ghosts. After its rendering, the Renderer no longer
contains sample records.

The flickering occurs when the rendering of the Renderer is async
(the one of the Controller being always sync, as it doesn't wait
for his children to be re-rendered, as their props didn't change).
Indeed, in that case, there's a small timeframe during which the
controller no longer has the classname `o_view_sample_data` but
the renderer still contains sample records.

Normally, the rendering of the Renderer should always be sync.
Indeed, we triggered the loadViews in its onWillStart but we
didn't wait for the rpc to return (we have a `isLoaded` flag, and
we have an empty rendering while `isLoaded` is false). However,
when the loadViews is already in the cache, the promise is resolve
in the next microTick, and we directly render the component with
the state `isLoaded` true, i.e. with the KanbanQuickCreateController.
But that component is always async, as it loads the form view data
(onchange) in its onWillStart, and must wait for it. As a
consequence, in that case, the whole rendering of the Renderer is
delayed.

To fix the issue, this commit simply ensures that the rendering of
the KanbanRecordQuickCreate is **always** sync, by toggling the
`isLoaded` flag in onMounted instead of onWillStart. That way,
the Renderer is rendered without the sample records in the same
animationFrame as the Controller, and only then we toggle the
KanbanQuickCreateController.

Closes #181743

Task~4196741

closes odoo/odoo#182753

X-original-commit: ae41fc7286566b720f07207491d779f1335b9544
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>