[FIX] point_of_sale: prevent singleton error when opening rescue sessions
When a point of sale has more than one rescue session and the user tries to open
the rescue sessions, a traceback will appear.

Steps to reproduce the error:
- Make 2 or more rescue sessions for one point of sale
- Click on the outstanding rescue session

Error: A traceback appears:
"ValueError: Expected singleton: pos.session(22, 21)"

https://github.com/odoo/odoo/blob/7638f1bdd6be52554ac4266291941ef0953edfd9/addons/point_of_sale/models/pos_config.py#L669
Here, session_ids have multiple records, so when it tries to access the id,
It will lead to the above traceback.

sentry-5916215733

closes odoo/odoo#183095

X-original-commit: 893f1fc3988f105904d6ae93dda9ec96ba27d47c
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Renilkumar Kajavadra (reka) <reka@odoo.com>