[FIX] project_stock_account: compute delivery aal into the materials costs section
This commit's purpose is to put the aal generated from deliveries into
the 'materials' section under the costs section in the project
profitability panel. Currently, those aal are considered as random aal
and put under the 'other costs' section. This does not sees very
appropirate.

In order to be able to differentiate aal created on the fly and the aal
generated from picking, a new option is added on the already existing
selection field 'category' in order to avoid a heavy search and
computation on the move.line model.

version 18.0 - master
task - 4180276

closes odoo/odoo#184121

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>