[FIX] mail: DiscussSidebar hover doesn't flicker on mouse-hovering
Mouse-hovering items when discuss sidebar is compact result in some
style flickering on im status icon.

This happens because the items are button and they need `bg-inherit`.
Due to CSS specificity, however, the classname is not enough,
so we need `!important` in CSS, which is what this commit does to
fix the issue.

closes odoo/odoo#182394

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>