[FIX] pos_sale: traceback while pressing backspace
Steps to reproduce :
- Install the point_of_sale and sales module.
- Set downpayment on the shop and open session.
- Press quotation/orders action button.
- Open any quotation try apply percentage downpayment.
- Try to press backspace twice.

Issue :
The number buffer would be "" at starting but when we again
press backspace it will become null as mentioned from number
buffer services which was directly passed to parseFloat
function.

Cause :
The parse float function applies startsWith function which
won't be there if value is null.

Fix :
Instead of passing buffer directly now just send buffer to
parse if it exist else send "" instead of null.

closes odoo/odoo#188852

X-original-commit: 9235f908e00c4bea138c318b5ed8d2079a8aeb0a
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Jainam Shah (sjai) <sjai@odoo.com>