[FIX] point_of_sale: ensure flexible record loading order
Before this commit, the `loadData` function required records to be
connected in a specific order, prioritizing models that are
dependencies for others. This posed a problem, for instance, when
attempting to load `pos.order` before `account.move`, as the latter
might be missing, preventing a correct linkage between `pos.order` and
`account.move`. This commit eliminates the dependency on loading order.
Now, if a record cannot be connected due to a missing dependency, it is
temporarily stored and later connected once the missing record becomes
available.

opw-4229303

closes odoo/odoo#182528

X-original-commit: a91da6a5e835c11f59a4847026ffc79ba97969c5
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>