[FIX] web: include postgresql server testing in health route
Prior to this commit, the health route is responding with a 200 status
even when the database server is down.

In deployment scenario (CI/CD), this route (/web/health) is (can be)
used in order to get the deployement state. This is actually the only
one that can be used in order to check it without an active session or
credentials or master password.

Including the optional db_server_status arg will allow integrating the
postgresql server state without breaking any eventual existing overrides.

closes odoo/odoo#183265

X-original-commit: 39a60b8a5a9642f949cde0edaf8132759012655f
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>