[FIX] account: amount_currency update in bank statement line sync
When changing the move associated with the bank statement line we
i.e. update the `amount_currency` field on the bank statement line.
Currently we just set it to the `amount_currency` value of the
suspense line.
But the `amount_currency` on the bank statement line expresses the
liquidity (total amount) and not the suspense (residual amount).
Thus this may not be correct in case we also have other lines on the
move: We should have (liquidity = suspense + other). But we do not
really know what the other lines are (they may have a different currency).
So to be safe we just do not update the `amount_currency` in case
there are other lines.

(Note: We cannot just use the liquidity line
It is in journal currency / company currency but the `amount_currency`
of the bank statement line may be in a foreign currency that
is neither the company nor the journal currency.)

closes odoo/odoo#182589

Related: odoo/enterprise#71243
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>