[FIX] web_editor, html_editor: remove ufeff characters when copying
Issue:
======
clipboard data has ufeff characters in it.

Steps to reproduce the issue:
=============================
- Create a new note/todo
- Add a link
- Copy it
- Visualise the data copied
- It contains ufeff characters of the link

Solution:
=========
for web_editor:
We remove the characters fron the text and html we put in the clipboard
data.

for html_editor:
We dispatch a CLEAN command before selecting the content and we handle
it by link_selection_plugin and we normalize after after cloning the
content

opw-4029722

closes odoo/odoo#184519

X-original-commit: f1edf4f771d42548d0bf0571e547c5bb7b9d374b
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>