[FIX] l10n_fr_hr_holidays: adjust full day leaves hours
Steps
-----
- Install l10n_fr_hr_holidays
- Use french company and be in a timezone different than UTC
- Set your user's employee to use a different working hours
schedule than the french company
- Take a day of paid time off

-> Two days of paid time off appear on the Dashboard.

Cause
-----
For full days leaves, the hours are set to min and max to avoid
calculation issues
https://github.com/odoo/odoo/blob/24e100a5acf7531787a646ff1473c4b027d1ad12/addons/l10n_fr_hr_holidays/models/hr_leave.py#L39-L40
But is not converted to UTC time, leading to the leave being displayed
over several days if not created by a user in UTC time.

Change
-----
For full days leaves, adjust the start and end time of the leave in
the same way that is currently done for half days.
This also makes the leave's hours closer to reality (no longer covering
a full day).

opw-4203972

closes odoo/odoo#186288

X-original-commit: 50f9bd9d351d83ca4e4ae88438210d7c9a208154
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>