[FIX] hw_drivers: fix browser refresh
The DisplayDriver class uses the `xdotool`
command to send keypresses to the browser.
However for Chromium, it is using the wrong
process name causing the keystrokes not to
be received.

The symptom of this is the Pairing Code does
not disappear after 5 minutes as intended,
leaving an invalid code on the screen.

This PR fixes the issue by using the correct
process name for Chromium.

closes odoo/odoo#182745

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>