[FIX] hw_drivers: remove usages of `_` translation helper
On the IoT box, translations are not supported as
there is no language context and no po/pot files.
However, there were still some usages of the `_`
function which will now log a warning in version
18.0+.

In this PR we simply remove the usages of `_`, as
it was not translating anyway and removes the
warnings.

closes odoo/odoo#182767

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>