[FIX] web: model: do not generate update command on virtual ids
This commit fixes an issue that requires a very specific scenario
to be reproduced:

Go to Settings > Technical > Server Actions
Add a Server Action on the "Contact" model with type "Execute Existing Actions"
Add a child action on the "Contact" model and set a domain using the
"country" field
Save the dialog and save the record.

Before this commit, it crashed. Indeed, we generated an update
command (1) with a virtualId as record id. This happened because
of the ReferenceField that was used in the dialog. When the record
was saved, it was also reloaded (via web_save), and the
`update_related_model_id` field was unset (the business logic unsets
it, I don't know why). This field being the `modelField` for the
reference field (`resource_ref`), the latter was unset by the
ReferenceField (via useRecordObserver), which called record.update.
This produced an update command on a record that had been saved
meanwhile but was still referenced as a new record (hence the
virtual_id).

This commit fixes the issue by not reloading the record with
web_save, as we're closing the dialog anyway, and that record is
loaded for the x2many list view anyway. Moreover, we also ensure
that if an update occurs on a record that was initially new but
that has been saved meanwhile, the command is associated with the
correct resId, not the virtualId.

Closes #178541

closes odoo/odoo#182929

Signed-off-by: Francois Georis (fge) <fge@odoo.com>