[FIX] l10n_de: Don't raise if account code not changed
In #171244, we prevented users from modifying account codes on German
CoAs, as part of changes for GoBD compliance.

However, when the user exports and re-imports an entire CoA in Excel,
`write` is still called on the account code, even if it doesn't change.

With this commit, we don't raise if the code doesn't actually change.

Back-port of #181234

task-4192627

closes odoo/odoo#182821

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>