[FIX] l10n_it_edi_doi: add condition to delcaration info on reports
Currently we always try to add a string informing about the used
declaration of intent to the invoice / sale order PDF.
I.e. it also happens when no declaration of intent is selected.

After this commit the string is only added in case there is
actually a declaration of intent selected.

opw-4237368

closes odoo/odoo#182913

Signed-off-by: John Laterre (jol) <jol@odoo.com>