[IMP] delivery: display shipping pickup point name
Issue: When a sale order made from the website with a pickup point
(for instance using a sendcloud delivery method) is confirmed, a new
partner is created to mix the data of the partner making the order and
the address of the pickup point. If the partner is named Bob, this new
partner will be named Bob, Bob. It woul dbe much clearer if he was named:
Bob, pickup point name.

opw-4181787

closes odoo/odoo#183164

X-original-commit: 7f51dd982b3252f824f679b28f5f7944b593f730
Related: odoo/enterprise#71674
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>