[FIX] sale_project: copy SO doesn't copy project
Steps:
Activate analytic accounting.
Create a SO with a service that creates a project.
Confirm. The project that has just been created is now set on the SO.
Duplicate.

Issue:
The project is duplicated too.
Yet is doesn't correspond to the the one that
will be created upon confirmation.
Also, the old analytic distribution is copied on SOLs.

Solution:
Simply mark the field as copy=False not to duplicate it upon SO duplication.
Also, reset the analytic distribution where it corresponds to the old
project.

task-4207182

closes odoo/odoo#182978

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>