[IMP] account: better error message when using inconsistent payment account on sale/purchase invoice
Before this commit, when a payable account was used on a customer invoice, or a receivable account on a vendor bill, an error message was raised saying "Any journal item on a receivable/payable account must have a due date and vice versa.". This was unclear, and caused for example confusion when importing the accounting history. Therefore, we now handle this case with a new error message.

opw-4196597

closes odoo/odoo#183795

X-original-commit: 850d34296d0c9bcba85b29040a31a4aa85c79a4f
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>