[FIX] http_routing: don't consider path prefixed by _ as a lang during match
Prior to this commit, when a URL started with an underscore (_) and didn’t
match any existing controller, Odoo attempted to resolve the first part of the
URL as a language code.

If the first segment started with an underscore and was not a valid language
code, Odoo would split at the underscore and attempt to match the empty string
(''), causing it to default to the first available language.

For example, accessing https://www.odoo.com/__NULL__ on odoo.com would result
in the language being incorrectly set to Arabic, and a redirection to
https://www.odoo.com/ar.

closes odoo/odoo#184217

X-original-commit: 674299f59ba6cca151eb83bf1de10e50fd42e7ca
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>
Signed-off-by: Jérémy Kersten <jke@odoo.com>