[FIX] mail: allow saving template in mass mode
In the new composer design [1] saving a template is done by first
saving the composer, then opening the "template creation" view
on the same composer.

The second step thus requires the composer to be properly saved first.

This is easily checked as `Record.save` returns false on failure.
In that case we do nothing and let the framework point out
why the record could not be saved.

[1]: 6c4526ec3b9509e0a6b99503d8d888417b2695e0

task-4246399

closes odoo/odoo#183108

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>