[FIX] hr_payroll: fix paperformat in document layout
In the document layout configurator, the paper format A4 would not be there if
the module l10n_be_hr_payroll is installed.

This is because in Belgium, some reports need A4 but in the configurator, we
only get paperformat not linked to any report.

The solution is to remove the domain for this field in the configurator.

Since there will be more paper formats, I renamed the newly created formats in
the Belgian payroll to differentiate them.

closes odoo/odoo#183134

Related: odoo/enterprise#71649
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>