[PERF] point_of_sale: optimize product addition to order
Before this commit, loading 20,000 products into the PoS and adding a
product to an order was slow due to multiple writes to the IndexedDb.
This commit optimizes the process by not saving ongoing order's
products in the IndexedDb, as they can be loaded as needed. Missing
products are loaded when loading the orderlines.

opw-4200540

closes odoo/odoo#183338

X-original-commit: 67abdb74c350b8b237718e44bb3c8d138462336d
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>