[FIX] account: ignore non-product lines on sale/purchase orders when creating accrued revenue entries
Problem:
When using the generate accrued revenue entries action on a sale/purchase order where the first order line is a section or note leads to an invalid account move being created. The move contains a line with not account set.

Purpose of this PR:
To make sure only product lines are taken into account when generating these entries

How to reproduce:
1) Create and confirm a sale order where the first order line is a section or note and the next contains a product
2) Run the Accrued Revenue Entry contextual action
3) Fill the Accrual Account and Amount fields with arbitrary data and hit Create Entry
4) Error appears, a move line is generated with no account

closes odoo/odoo#185665

X-original-commit: b3b1fe6a78f9e5b6f1d993b5aa2fed11e33c793e
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Emeka Ubakanma (emub) <emub@odoo.com>