[FIX] l10n_it_edi: document type misconfigured
When sending a document to the government, we have to put a certain document
type. The document type are computed depending on what's on the lines of the
move, the type of the move, etc

The mapping was wrongly done. For the moment when doing an invoice with a
service, it will be mapped as a TD03 which was wrong. It should still be TD01.
We misinterpreted the professional fees by just checking if there is more
service than goods product but now we will add a tag that can be used on account
for that.

closes odoo/odoo#183337

Task: 4244462
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>