[FIX] pos_self_order: prevent multiple payment page renders
Before this commit:
===================
The payment page was being rendered multiple times due to the `onWillStart` method
triggering the `startPayment` method multiple times. This led to multiple requests
being sent to the Razorpay terminal for the same order.

After this commit:
==================
Replaced `onWillStart` with `onMounted` to initiate the `startPayment` method.
This prevents duplicate requests to the Razorpay terminal for the same order,
ensuring the payment process is handled efficiently.

Task- 4254682

closes odoo/odoo#183518

Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>