[FIX] crm: prevent crash with many2many tasks
Steps to reproduce
==================

- Go to CRM
- Open a lead
- Open studio
- Add a new notebook page
- Add a many2many field
- Select the task model
- Save and exit studio
- Switch to the new notebook page
- Add a new task

=> Record does not exist or has been deleted.

Cause of the issue
==================

The CrmFormRecord is also used for the Many2Many popup.
project.task also has a stage_id field, checkRainbowmanMessage is then
called, but with the id of the task

https://github.com/odoo/odoo/blob/54d6a19444bfa4f01d03117e4542a5244f274429/addons/crm/static/src/views/crm_form/crm_form.js#L44-L51

Before https://github.com/odoo/odoo/commit/218ad8456a06503dd508e7216edcffdc90b35cac
this behaviour was part of a custom form controller and was only called when
saving the main record.

After that, it is now part of a custom Model

Solution
========

It makes no sense to do this if we are not acting on a crm.lead record
=> We check the resModel before proceeding

opw-4101752

closes odoo/odoo#184012

X-original-commit: 1aa1e9fa0cff77f029bf44235930892f51352f15
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>