[PERF] mrp, *: add checks to _compute_days_to_order
In both mrp and purchase_stock, `_compute_days_to_order` does some computation
when the rule_ids' action is 'manufacture' and 'buy' respectively. This forces
the computation of rule_ids on all orderpoints in self. This commit adds
some prechecks and prefiltering on self to reduce the size
of the recordset that needs the value of rule_ids.

Part-of: odoo/odoo#183964
Signed-off-by: William Henrotin (whe) <whe@odoo.com>