[PERF] stock: use filtered instead of continue in compute functions
`filtered` sets the _prefetch_ids to the records that matches the condition.
Using `filtered` before looping instead of doing an if + continue in
a for loop reduces the size of the expanded recordset whose compute field
value needs to be recomputed. This reduces the overall number of operations
done for the given compute function.

Part-of: odoo/odoo#183833
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Aurélien van Delft (avd) <avd@odoo.com>