[FIX] web: add an empty value to the minute selection in DateTimePicker
Steps to reproduce
==================

- Open any list/kanban view
- Add a custom filter
- Select a datetime field, for example Created on

=> The default value is the current datetime
   If you click on the datetime field, the datetimepicker is opened but
   the minutes are set to 00.
   If you click on apply, 00 is not set.

Cause of the issue
==================

The DateTimeInput has a rounding props (by default 5).
If we open it with a value that has the minutes not rounded by that
amount, 00 will be displayed instead because a <select/> is used, and
only the rounded values are present.

Solution
========

Add an empty value to the select if needed

opw-4170906

closes odoo/odoo#185994

X-original-commit: fdc50a9328389325589a5bd7a2860a84002f9f3c
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>