[FIX] html_editor: consistent working of dropdown
Before this commit:

Clicking near the edge of a toolbar dropdown item would not work. This occurred
because the `user-select-none` class was applied to the child element rather
than the dropdown item itself, causing `onSelectionChange` to be triggered,
followed by `updateToolbarVisibility` closing the toolbar.

After this commit:

The `user-select-none` class is now applied directly to the dropdown item,
preventing `updateToolbarVisibility` from closing the toolbar when clicking near
the edges of dropdown items.

task-4241028

Part-of: odoo/odoo#183486
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>