[FIX] resource: resize html field when record saved/changed with pager
Before this commit, when the user uses the pager, the html field is not
resized and have a minimal height as defined in its class. Same issue
when the user saves the record.

This commit makes sure the height of the html field is recomputed when
the record is saved or when the user uses the pager.

Part-of: odoo/odoo#183396
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>