[FIX] website: prevent error when saving page properties without page title
When the user tries to save page properties without a Page Title,
a traceback will appear.

Steps to reproduce the error:
- Go to Website > Site > Properties > Remove the Page Title > Save & Close

Error: A traceback appears:
'TypeError: normalize() argument 2 must be str, not bool'

This error generates after this commit:
https://github.com/odoo/odoo/commit/58705803a9e7ef0b0f5ff9e8b82d52ad1321b64a

https://github.com/odoo/odoo/blob/4b99aa4e48158ddc7cea8910c0f610e62778bc9c/addons/website/models/website_page.py#L158
When, user tries to save page properties without a page title,
vals['name'] will be False,
So, It will lead to the above traceback.

sentry-5989403584

closes odoo/odoo#183550

Signed-off-by: Benoit Socias (bso) <bso@odoo.com>