[FIX] website: trap the focus inside popups
Trapping the focus in modal dialogs and refocusing the last focused
element when closing the dialog are requirements from the WCAG. It was
not done up to now for popup snippets in the website.
As popups without backdrop aren't dialogs (they don't prevent
interacting with the rest of the page), we don't trap the focus in
these.

Steps to reproduce (Chrome/Firefox):
- Drop a popup and save
- Wait for the popup to appear
- Use tab / shift + tab to navigate around the page
=> Without the fix, you're able to focus elements outside of the popup.

task-4256274

closes odoo/odoo#190183

X-original-commit: 89e2513f577e9455d5bfd933d7e479b295d45a26
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>