[FIX] pos_hr: handle non-existent employee across sessions
Before this commit, logging into the PoS with a user not present in
another session and then navigating back to the backend to open a
different session resulted in an error due to the missing employee.
This commit resolves the issue by incorporating the confid ID when
saving employee ID in sessionStorage, ensuring employee information
is accurately maintained.

opw-4255768

closes odoo/odoo#184543

X-original-commit: 7671d6cec2132d0052ab8425a5243f579ecc2f6b
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>