[FIX] html_editor: correctly position powerbuttons for RTL languages
Previously, the positioning of powerbuttons only considered the `dir` attribute
on the closest block element. However for true RTL languages like Arabic, the
`dir` attribute is applied to the editable container, not the individual nodes.
This commit makes sure that `dir` attribute of editable is taken into account
when the closest block does not have its own `dir` attribute.

task-4259040

closes odoo/odoo#183710

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>