[IMP] web_editor: autoscroll latest prompt in chatgpt
Behaviour before commit:

In chatGPT prompt dialog, scroll position is not set to the latest
message automatically after hitting 3-4 prompts. User has to scroll
to bottom manually.

After this commit:

This commit aims to make sure that the scroll postion is set
to the latest message after submitting prompt.

task-4256085

X-original-commit: 3c403505c72ccc4af4f44c8d84580a0a1f8d1e1b
Part-of: odoo/odoo#185901
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>