[FIX] l10n_din5008{*}: incorrect din5008 report addressing
The DIN 5008 layout was not properly addressing reports on invoices and delivery slips. On invoices/pro-forma, the reports were not addressed to the correct partners (commercial partner instead of invoice partner). In the delivery slips the reports were not addressed to the delivery partner. This led to functionally and legally incorrect reports.

task-4089521

closes odoo/odoo#190012

X-original-commit: eaf9e77650d70e241ae8261fad3c09a801493453
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: Anton Romanova (roan) <roan@odoo.com>