[IMP] l10n_ee: rework tax report
The tax report (KMD report) and annex to the tax report (KMD INF report)
for Estonia needed to be reworked to correct errors and fix rounding
issues.

In this commit, the 'balance_from_tags' label is added to the tax report,
allowing a better audit of the report lines. Furthermore, the tax tags
are edited to differentiate between tax and base tax lines.

These changes are necessary to improve auditing of the tax report lines.

task-3997203

closes odoo/odoo#190180

X-original-commit: c43174df64ecde09eb0503cdb9f75b98937fc253
Related: odoo/enterprise#75407
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Maira Salazar <srma@odoo.com>