[FIX] web: ARIA attributes of field widgets not ported correctly to OWL
When field widgets were ported to OWL in 48ef812a, a few ARIA attributes
were not adapted correctly. For instance, in the statusbar widget, the
title was ported as an `aria-label` attribute ("Current state" and "Not
active state"), which overrides the actual status name.

This commit fixes the above issue by removing the incorrect `aria-label`
attributes and fixing logic for other ones like `aria-checked`.

closes odoo/odoo#192180

X-original-commit: 355e035e9e156646b63ffb4eb6eb55c5bad8485d
Related: odoo/enterprise#76380
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>