[FIX] website_project:  prevent KeyError: 'task' on click Preview URL
Currently, an exception is generated when the user clicks on the preview
URL of the "Create a Task" form by following the steps:

- Go to any website  page and open the website editor
- Add form to website page
- click on the added form and select Action as "Create a Task"
- Click on the Preview URL button of the added form >> errro generated

error `KeyError: 'task'`

This is because when the user opens the direct URL, no task is available.

This commit will fix the above issue by accessing the key `task` when
it is available in view.

sentry-5983056185

closes odoo/odoo#183866

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>