[FIX] hr_fleet: rearrange fleet attachment upload buttons
Steps to reproduce:
- Fleet app > Any vehicle > Drivers History > Attachments
- In mobile view > Primary button 'Upload' is not accessible
- Primary button 'New' does not do anything

This happens because the controller cannot find a form view for
ir.attachment in fleet, which is the default view type for record
creation (barring quickcreate stuff). Also d-none class hides the
display of the 'Upload' button but the dropdown menu arrow is still
there which is rather confusing.

Seeing how both buttons serve the same purpose, there is no need to keep
'New' which has no corresponding view.

opw-4229756

closes odoo/odoo#186035

X-original-commit: 6aefdc0ec28cedf535040b07df92210a0e5dc23e
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>