[FIX] base: remove mogrify for _read_group_orderby
`_read_group_orderby` used `cr.mogrify`, `split(",")` and `split()[0]` for the
result coming from the `_order_to_sql` which may come from `_field_to_sql`
As a side effect, SQL code after a space/newline not adjacent to a comma will
be dropped after `split()[0]`. So the mogrified customized `_field_to_sql`
shouldn't have any space/newline not adjacent to a comma if the field is used
for `Model._order`.

closes odoo/odoo#183925

Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>