[FIX] website_sale: display zero-priced products with price extras
Versions
--------
- 16.0+

Steps
-----
1. Create a zero-priced product template;
2. add some attribute values;
3. configure price extras for the attributes;
4. configure eCommerce to prevent sale of zero-priced products;
5. go to eCommerce products page.

Issue
-----
Product is displayed as "Not Available For Sale," even though you can
click on it, select a price-extra attribute, add it to cart, and
purchase it.

Cause
-----
The template doesn't consider potential `price_extra` attributes when
displaying the not available message.

Solution
--------
Add a `t-elif` element to the template, checking if the product has any
non-zero price-extra attribute values, if so, don't display the not
available message, but also don't display the zero price.

opw-4225183

closes odoo/odoo#184052

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>