[FIX] account: portal use a custom template
In the Print & Send, it is possible to use a different report template
than the Odoo default one ('account.account_invoices').
This is not reflected in the portal preview.

Solution:
Which template has been used for an invoice is not an information that
is saved in Odoo so we can't do something 100% accurate.
When an invoice is generated with a different report than Odoo's one
we save that as a setting on the partner.

We will render the portal preview with that template if it is set, to
try to be closer to what should be shown.

task-no (approved by CHKL)

closes odoo/odoo#184150

Signed-off-by: Laurent Smet (las) <las@odoo.com>