[FIX] base: give has_group access to internal user
The code comment says non-internal users should not have access to
has_group if it's not for themselves. But the code checked the group
of the targetted user, not the current user.

Without the fix, the added test failed because:
- AssertError was not raised if non-internal accessed internal user
- AssertError was raised when internal accessed non-internal user

note: found when reviewing 18.0 forward-port of d0828eecf60f7c8622d6875b

opw-4096073

closes odoo/odoo#184600

X-original-commit: 3cc0044c0129c90f52df78d6b4e23c6a424736ca
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>