[FIX] mail: fix html_mail_field error on Safari
Issue
On Safari iOS, editing `html_mail` fields (such as "Email signature"
in user profile or "Subject" in Quotation send by mail) throws an
error: `@page is not a valid selector` at save.

Change
Add `@page` to the list of ignored selectors as it was done in commit
7b68c9eaf8ad94ecbcc3893bf6cc220b7e9b6636 as it is currently not
supported by all browsers.

opw-4256252

closes odoo/odoo#184316

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>