[FIX] web,mrp: workorder list view fixes
- changed the "start", "pause" and "done" button text to icons to be able to
set the button column size at a fixed value
- moved the unblock button in the dropdown
- made the previous block button toggle between block and unblock
- change the dropdown bullet to red when blocked

This commit also modifies the width of the json_popover as it was using
the base column width which was too big for the icon.

It also modifies the width of the mo_list_dropdown column to limit the empty space.
This implies changes in the web module:
in `column_width_hook.js`, the width of a widget is equal to it's `listViewWidth`,
like the fields. However, unlike the fields, the widget class is missing a validation
for that props.

closes odoo/odoo#184308

Task-id: 4279789
Related: odoo/enterprise#74396
Signed-off-by: Steve Van Essche <svs@odoo.com>