[FIX] pos_loyalty: correctly detect gift card and eWallet
When adding a rules in loyalty program that had a product set, if you
try to refund an order containing this product you would get an error

Steps to reproduce:
-------------------
* Create a loyalty program with a rules that has product_ids set to any
  product.
* Open PoS and make an order with the product set on the loyalty program
* Validate the order
* Try to refund the order
> Observation: You get an error saying you cannot refund giftcards or
  eWallets

Why the fix:
------------
This was happening because `trigger_product_ids` is related to
`rule_ids.product_ids`. And when checking if a product is a giftcard or
eWallet we first checked that `trigger_product_ids` was set.
Now we only check the type of the program linked to a product.

opw-4206226

closes odoo/odoo#187325

X-original-commit: 0b95547bc6e290e3e869e076d45329173a4add12
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>