[FIX] point_of_sale, pos_restaurant: always show change value
When using the restaurant, orders that were sent to the kitchen did not
show the change amount on the receipt.

Steps to reproduce:
-------------------
* Open restaurant
* Select any table, add items to the order
* Select **Order**
* Select **Payment**
* Select any payment method
* Pay more than the order
* Validate
> Observation: The receipt show a change of O.0

Why the fix:
------------
Before this commit: https://github.com/odoo/odoo/commit/1649ee1a757ab1a095fd92f7605ac2b004568352
the change could get computed on an onchange:
https://github.com/odoo/odoo/blob/662f63b630cae8604179174cdd67536f3300bb53/addons/point_of_sale/models/pos_order.py#L417-L427

By definiton, onchanges will only trigger from web form views. In our case,
the onchange was never triggered when synchronizing orders from the UI, even
though we would make a change to the payment_ids when adding the change line.
After the above mentioned commit, the code inside the onchange was put into
a separate function `_compute_prices()` which is still called in the onchange
but also in the `write` function.

When an order was previously synced from the UI, we call the write function.
https://github.com/odoo/odoo/blob/430656132044f8d675712d5b6cbfef807880d024/addons/point_of_sale/models/pos_order.py#L124

By going into `_compute_prices` in the write method, we compute the `amount_return`
before the change line was added to the order.
https://github.com/odoo/odoo/blob/430656132044f8d675712d5b6cbfef807880d024/addons/point_of_sale/models/pos_order.py#L128

Computing `amount_return` before adding the change line is problematic since
it is computed with regards to the negative payment lines.
https://github.com/odoo/odoo/blob/430656132044f8d675712d5b6cbfef807880d024/addons/point_of_sale/models/pos_order.py#L440

To solve our solution, we call `_compute_prices` after we add the change line.
`amount_return` represents the change given to a client. If we give `10$` back to
a client, the change payment line will have an amount of `-10$` but `amount_return`
should be positive 10. (Thus the change of `-`)

We also add a small fix for `amount_difference` which can have precision
error and thus show something like `-0.0`.

opw-4253292

closes odoo/odoo#184459

Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>