[REV] l10n_it_edi: prevent generation of XML if document type is not set"
This reverts commit fd1da69f8331dbc55be2d164fad381fca1622a67.

The l10n_it_document_type belongs to `l10n_it_edi_ndd` module.
This module exists to add this stored field that was missing.
When the error that this commit tried to fix happens, which is due
to the fact that the compute doesn't find a correct matching document, we
don't want to block the flow. The user can still set a document type himself
through Odoo (if he has the aforementioned module installed) or directly in
the XML.

closes odoo/odoo#184481

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>