[FIX] account: prevent singleton error while un-merging multiple account
Currently an error occurs when the user unmerges two or more charts of accounts
with the same company.

error:
`ValueError: Expected singleton: account.account(1249, 1252, 163, 164, 165, 754)`

This is because at [1] `self` used to display name here `self` contains
multiple records.

This commit will fix the above issue by looping records of self and using
`display_name` of account, which is needed.

[1]- https://github.com/odoo/odoo/blob/0bac60a7085c51ac8cb4f793b013dbea927d180c/addons/account/models/account_account.py#L1108

sentry-5985600163

closes odoo/odoo#184529

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>