[FIX] website_sale: make ProductConfiguratorDialog title translatable
On the website, the title of the ProductConfigatorDialog is not
translated into any language. This is because only "frontend modules"
translations are fetched in the context of the website. The title of the
dialog is defined in the "sale" module, which is not a frontend module,
so the translation is missing.

This commit solves the problem by redefining the string to translate in
the website_sale module.

Task-4182798

closes odoo/odoo#184569

X-original-commit: 4364073e8e8bbc194fa3e499df4f07afcd5721ea
Signed-off-by: Valentin Chevalier <vcr@odoo.com>
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>