[FIX] mail: click on 'cancel' or 'save' once when editing message
Before this commit, when editing a message, we had to click twice
on "Cancel" or "Save" to register the edit message action.

This happens because while clicking on these action for the 1st
time, composer intercepts a "focusout" event, which re-renders the
template and thus the "Cancel" & "Save" buttons too, thus cancelling
their action. The 2nd click doesn't trigger this focusout thus it
works.

This commit fixes the issue by moving code that generates the
"Cancel or Save" text in a sub-component. Doing so ensures that
the component is only rendered when needed, in this case when
Composer.mode or ui.isSmall change.

opw-4119283

closes odoo/odoo#184626

X-original-commit: 6419235bfc5e226bf8bf011e46004002890bd741
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>