[FIX] sale: prevent warning on quotation creation
Issue:
A warning popup appears incorrectly when a user-defined default is set for the "Quotation Template (Sales Order)" field. This warning is only relevant when the company is changed while creating a quotation.

Steps to reproduce the issue:

1- Navigate to Settings > Technical > Actions > User-defined Defaults.
2- Create a new User-defined Default for the field "Quotation Template (Sales Order)" with a default value (in JSON format) of 1 .
3- Go to the Sales module.
4- Create a new quotation.

A warning with the title "Warning for the change of your quotation's company" is displayed, despite the company not being changed.
The onchange('company_id') method is triggered by the user-defined default for the quotation template.

opw-4244961

closes odoo/odoo#184713

Signed-off-by: Soukaina Tyes (soty) <soty@odoo.com>