[FIX] website{_slides}: scroll animation in fullscreen
Current behaviour:
When you make content of type article in eLearning,
then add an animation on scroll on an element,
when you put the article in fullscreen,
the animations do not trigger.

Steps to reproduce:
1. Go to eLearning
2. Select a course with articles
3. Click on Go to website
4. Go on the article content
5. Open the editor
6. Drag and drop the columns block at the end
7. Select "feature one" and add an animation on scroll
8. After saving, scroll, the animation works
9. Click on "Fullscreen"
10. Scroll again, animation doesn't trigger

Cause of the issue:
In website_slide, fullscreen mode, the right scrollable
element is not #wrapwrap, but .o_wslide_fs_article_content

Note:
If the animation is not complete, but the page is scrolled
at maximum, animation intensity can be lowered in the
editor panel, so that the animation can finish earlier.

opw-4151770

closes odoo/odoo#188156

X-original-commit: 02acb080746be0f88075eafb51a17291abb0451b
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Signed-off-by: Antoine Demany (ande) <ande@odoo.com>