[FIX] sale{_project}: Move method back to sale
This fixes issues that arose since 238a41e3

By moving the `_prepare_analytic_account_data` method
from sale to sale_project, any call to this method
in sale_expense would break as sale_expense doesn't depend on project.
Because this method is crucial in the re-invoicing flows we need it
in sale.

closes odoo/odoo#184743

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>