[FIX] l10n_in_withholding: tds tax on partner type not payment type
While adding the `TDS` taxes on payment for receipt from vendor and payment to
customer, the type of tax i.e purchase or sale should be dependent on partner
type i.e supplier or customer.

For the vendor tds taxes should be of purchase type and for the customer it
should be of sale type.

task-4262953

closes odoo/odoo#184823

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>