[FIX] website: allow empty default value for select form fields
When model selection field is used in a form, its default value is
always one of the available values. Because of this, if the field is
required, it does not force the user to pick a value, but it provides a
possibly incorrect default value.

This commit makes an empty value available for selection if no default
value is chosen.
The behavior is also made available on custom fields, if such an empty
value had been manually created, it will be replaced by this
implementation.

Steps to reproduce:
- Install website_form_project
- Add a form in a page
- Select the "Create a Task" action
- Add the "Analytic Account" field to the form
- Make the field required

=> There was no way to have no default value by default.

opw-4268164

closes odoo/odoo#185495

X-original-commit: a3812be90017a921df3cfeb4af48a04a75caa315
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>