[FIX] l10n_es_{edi_facturae,pos}: fix computation of the default selection of Facturae
Before first commit of this PR https://github.com/odoo/odoo/pull/184822, we were erronously
not using Facturae as default EDI formats on Spanish customers.
Adding back the feature revealed we failed the FW of https://github.com/odoo/odoo/pull/181435, who
thought it was fixed in 18.0 already.
It was not, this commit reinstore the logic and refine it a bit: we avoid to set Facturae
by default if no certificate ever existed. That mean that :
- if a certificate is there but expired, we try to use Facturae which will raise
- if a certificate is there and valid, we use Facturae
- if there is no certificate at all, we don't use Facturae

task-no

closes odoo/odoo#184822

Signed-off-by: Laurent Smet (las) <las@odoo.com>