[FIX] web: dynamic placeholder popover not closing
This commit fixes an issue where the dynamic placeholder popover would
not fully close upon pressing escape or clicking on the cross when a
relation field was selected. This is because its previous condition for
changing state instead of closing was if it had some field path selected
upon closing the model field selector popover. This would not handle
well the case where a relation is selected before closing. The fix is
therefore to specify inside calls to the close function of the model
field selector popover whether it closed because a final value was
selected or not.

task-4066184

closes odoo/odoo#184992

X-original-commit: e3aab423dc3ae8c71362b705d10857e54d112177
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>
Signed-off-by: Julien Carion (juca) <juca@odoo.com>