[FIX] http: remove usage of cgi
As part of the removing `Dead Batteries` pep 594, the usage of the `cgi`
module is deprecated and it will be removed the standard library in
python 3.13. See https://peps.python.org/pep-0594/#cgi

Here the cgi module was used to parse the `Content-Type` of a response
and separate the mime part from the parameters. But as the mime part was
only used to verify the beginning of the string this separation is not
nedeed.

While at it, this commit removes the unused zlib import.

closes odoo/odoo#184845

Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>