[FIX] payment_adyen: add idempotency key for customer-initiated payments
Payments initiated by the customer from the payment form could sometimes
lead to multiple charges if a webhook notification would arrive at the
same time as the payment request's response. This is due to the webhook
acquiring a lock in the database when updating the transaction's state,
followed by the processing of the payment request's response trying to
do the same and thus encountering a concurrent access error. The payment
request is thus retried and a new charge is created on Adyen side.

This commit fixes the issue by passing an idempotency key with the
payment request. If a payment request is inadvertently retried, Adyen
silently ignores it and returns the same response as for the initial
request. The response is processed again in Odoo and the customer is
smoothly redirected to the payment landing page.

closes odoo/odoo#185720

X-original-commit: 3c902087122f84d46f66c301ace0e7e70eb2c113
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>