[FIX] stock: add default value to Selection fields
Adds default value to a few required selection fields as currently a
value is apparently set in the form (as the select field will pick the
first value) but without having the value really set in the field,
meaning that on save, the field will appear as invalid.
Also, to be able to select the "almost default" value, you'd need to
select another value then pick the first value again to properly set it
in the field and be able to save the form.

To avoid those issue, we can add a default value to those fields
corresponding to the current "default" value in the form.

closes odoo/odoo#185269

X-original-commit: 6f833492d0092e1077c3f2996d75a89a897131b6
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>