[IMP] html_editor: remove editor command "ADD_EXTERNAL_STEP"
This command's purpose was to dispatch the event at each individual external
step, but it was only used for embedded components, and `onExternalHistorySteps`
resource (which handles every batch of collaborative steps) is enough for
embedded components, and it is called less often. This commit therefore removes
`ADD_EXTERNAL_STEP` command.

Small fix to the table of content plugin, only `STEP_ADDED` has a
`stepCommonAncestor` payload property, therefore commands which don't have it
will use `this.editable` instead to search for new headings.

task-4281356

Part-of: odoo/odoo#184929
Related: odoo/enterprise#72681
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>