[FIX] analytic: prevent error when the user clicks view button in anyalytic
The `action_read_distribution_model` method was removed from
the following commit.

https://github.com/odoo/odoo/pull/182278/commits/777dea298fd74f2349ca22d05fe7f48596216dfe

Since the client's non-updated views still call it,
but get a traceback as it doesn't exist anymore.

Error:-
"AttributeError: The method 'action_read_distribution_model'
does not exist on the model 'account.analytic.distribution.model'"

We add it again to fix the issue.

sentry-5804153543

closes odoo/odoo#185018

Signed-off-by: Adrien Widart (awt) <awt@odoo.com>