[FIX] base: prevent traceback when editing the view
When the user edits the view and adds context with invalid syntax,
a traceback will appear.

Steps to reproduce the error:

- Go to Settings > Technical > Views > Open any view
- Add context = "[]" or context = "{a}" like this in the view
- Save

Error: A traceback appears:
"UnboundLocalError
cannot access local variable 'err' where it is not associated with a value"

https://github.com/odoo/odoo/blob/69b404c7109ff689381f56520aad758424ec01aa/odoo/addons/base/models/ir_ui_view.py#L413-L418
Here, the "err" variable is referenced before the assignment,
So, it will lead to the above traceback.

sentry-5993638522

closes odoo/odoo#185033

Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>