[FIX] web,*: align checkbox with label text in mobile and form view
* = [hr_expense, pos_loyalty, product, sale_loyalty , website_sale_loyalty]

Before this commit:
In the mobile and form view, when a checkbox field had a long label, the
checkbox would appear on top, with the label text displayed below.

Also, Long labels extend beyond the boundaries of the form view, appearing
outside of it.

After this commit:
Checkbox fields with long labels are now properly
aligned in the mobile and form views. The checkbox and its label text are
displayed correctly,

Task-4269578

closes odoo/odoo#187980

Related: odoo/enterprise#74561
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>