[FIX] purchase: remove HTML tags from tax description in PO report
When printing the purchase order, the tax values were displayed with HTML tags because the field is an HTML field, and it was passed to the `join` function, which treated it as plain text.
```<span t-esc="', '.join(map(lambda x: x.description or x.name, line.taxes_id))"/>```
To fix this, invoice_label is used instead of description.
This issue occurred after https://github.com/odoo/odoo/pull/160342

Steps to reproduce:
1. Create a purchase order PDF using taxes from a localization (e.g., Belgium).
2. Print the purchase order and observe the tax values displaying with HTML tags.

opw-4260595

closes odoo/odoo#185095

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>