[FIX] web: close google slide in shopfloor
When an iFrame and its parent document are from different origins,
contentDocument is null. See:
https://developer.mozilla.org/en-US/docs/Web/API/HTMLIFrameElement/contentDocument

This will cause an error when the iFrame contains a Google Slide, which is from a
different origin. Due to this, we need to check that there's a contentDocument
before trying to use its content.

opw-4240622

closes odoo/odoo#185107

Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>