[FIX] web: load_breadcrumbs controller should always have a return
In the case of a client action with a record id on the breadcrumb part
of the url (`myClient/22` for instance). The controller should return
the display name for the combination action + record id. And for the
action alone (a multi-record if it was the case of a window action), the
controller should return an error, to be removed from the breadcrumb.
This is because client actions don't have multi-records views.

Before this commit, the controller avoids returning anything, and an
error raised.

Now, the controller returns correctly an error, and it's correctly
removed from the breadcrumb.

closes odoo/odoo#185328

Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>