[FIX] maintenance: rework calendar_with_recurrence tours
Issue:

`test_dblclick_event_from_calendar` tour test fails on Saturday

Explanation:

In the tour, we test an event with daily recurrence. Since the Weekly Calendar starts on Sunday and ends on Saturday, if an event starts on Saturday, the second occurrence of the event is on the next week and can not be reached.

Fix:

Moving to next week for the occurrence test, limiting recurrence to 8 days to remove useless rendering.
Setting `drag_and_drop` tour in Monthly Calendar view to avoid issues caused by potential `drag_and_drop` behaviour changes. In Monthly Calendar view, some days from previous and following months are visible, with a maximum of 6 before (Feb 23, see Mar 2025) and 14 after (Mar 14, see Feb 2026). Initial date being set in the backend, it does not have to respect those limitations. Changes operated in the frontend are bound to these and must be set between 15 and 22 included with the current selector.

runbot-error-65494

closes odoo/odoo#185385

X-original-commit: 9f4f1d8e92d0dc68b5628194d9da56d346e708f6
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>
Signed-off-by: Stroobant Paul (stpa) <stpa@odoo.com>