[FIX] website_sale: make input placeholder translatable.
When you choose to add a free text value to a variant of the optional
products at checkout, you will encounter a placeholder that is not
translated: "Enter a customized value". This is because it's defined
only in `sale` module, which is not a frontend module. This commit adds
the untranslated text to the frontend module: website_sale.

Task-4229289

closes odoo/odoo#185789

X-original-commit: 0ce6c390183e80c75c3e41dfe4292f80af217174
Signed-off-by: Mohammed Basioni (basm) <basm@odoo.com>
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>