[FIX] web: move alert styling from bootstrap_overridden to report.scss
Changing bootstrap versions caused a change in how the color for the
alert-link class was being chosen which broke something with pdf reports.
A change was then made to fix pdf reports
but this ended up breaking alert-links elsewhere in Odoo and the light
mode colors for alert-links were being used in dark mode.

This change moves the styling to a report specific scss file to resolve
the issue in the backend.

Here is the aforementioned change: https://github.com/odoo/odoo/pull/178824

opw-4211334

closes odoo/odoo#185262

X-original-commit: db129a4ccef3f899ccaa28fabdeb6f4c9fe52393
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>