[FIX] spreadsheet: disable see records on manipulated pivot table
Steps to reproduce:

- insert a pivot with more rows than columns
- write in a cell =TRANSPOSE(ODOO.PIVOT.TABLE(1))
- right-click on the grand total value

=> boom

When computing which cell of the pivot table is clicked, we assume the matrix
comes directly from the ODOO.PIVOT.TABLE(...) function to compute the offsets
from the array formula. But it's completely wrong as the cell could at a
completely different place if the matrix is manipulated by other functions
before being outputted to the grid.

closes odoo/odoo#185534

Task: 4292134
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>