[FIX] sale, pos_sale: take into account pos downpayments on invoicing
Currently, if an order has a downpayment made from the POS, when we invoice
it, the unit price of the downpayment becomes 0 on the sale order. The same
flow using only sale will correctly show the downpayment price.

Steps to reproduce:
-------------------
* Create a quotation in the **Sale** app
* In the **Point of Sale** app, open shop session
* Create a downpayment for the order
* Go back in the **Sale** app
* Open sale order
> Observation: The downpayment line has a price unit set
* If needed, deliver the items
* Create an invoice (regular invoice)
* Confirm the invoice
* Go back to the sale order
> Observation: The downpayment line has a price unit of 0.0

Why the fix:
------------
The difference between the two flows (pos/sale) mainly resides in those
few lines:
https://github.com/odoo/odoo/blob/b3b1fe6a78f9e5b6f1d993b5aa2fed11e33c793e/addons/sale/models/account_move.py#L75-L85

`dpl.invoice_lines` will include 2 items when using the sale flow whereas
only one when making the downpayment in pos.

Downpayments from pos are not automatically invoiced and even if they are,
the model `pos.order.line` does not have the field `invoice_lines`.

However each downpayment **sale order line** that comes from the POS is linked
to its downpayment **pos order line** with the field `pos_order_line_ids`.

The only element left in `dpl.invoice_lines` will not be counted as it is
included in `real_invoice` (the current invoice).

The idea of the fix is to include the price unit of the downpayment made
in pos in this sum.

opw-4160111

closes odoo/odoo#186753

X-original-commit: f66c8a91f77cae894c9f5ebb4612832a75e8c20d
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>