[FIX] base: do not fetch uninitialized columns
When guessing the language at install (or upgrade) we cannot trigger a
pre-fetch of all fields of `res.company` because some of its columns may
be not initialized yet. One way to trigger this behaviour is to add a
stored field whose default value uses translations, while at the same
time adding any other stored field. If the translation is fetched in a
sudo environment we can get an error when the System partner has `NULL`
language.

Steps to reproduce:
1. On a cleanly initialized DB (only `base`) add any extra language
   (en_GB would do).
2. Open Users & Companies (show inactive users) open System user.
3. Open Related Partner.
3. In the form view of the System partner set the language to None.
4. Try to install via the command line a module with the fields as
   described above (see below for a code example).

We get an error like:
```
2024-10-30 08:48:40,831 41180 INFO test_16 odoo.modules.registry: module mymodule: creating or updating database tables
2024-10-30 08:48:40,839 41180 ERROR test_16 odoo.sql_db: bad query: SELECT "res_company"."id" AS "id", "res_company"."name" AS "name"
... <many more fields, redacted>  ... "res_company"."bar" AS "bar", "res_company"."foo" AS "foo" FROM "res_company" WHERE "res_company".id IN (1)
ERROR: column res_company.foo does not exist
LINE 1: ...e" AS "write_date", "res_company"."bar" AS "bar", "res_compa...
```

The solution we propose is to skip the prefetch of `res.company` fields
when we are in the process of guessing the language. In this case the
language is correctly picked from the System partner's company.

This issue has been observed in many upgrades to 18, but it's present in
older Odoo versions as well. The language field in `res.partner` is not
required, thus it is also possible to set the language to `NULL` via
code.

Example data observed in upgrades:
```
> select id,name,active,lang from res_partner where lang is null
+----+-----------------------+--------+--------+
| id | name                  | active | lang   |
|----+-----------------------+--------+--------|
| 4  | Public user           | False  | <null> |
| 5  | Default User Template | False  | <null> |
| 6  | Portal User Template  | False  | <null> |
| 2  | System                | False  | <null> |
+----+-----------------------+--------+--------+
```
There may be other standard flows that lead to such setup where System
partner has no language set.

Example inherit of `res.company` that triggers the issue:
```py
class A(models.Model):
    _inherit = "res.company"

    bar = fields.Many2one(
        comodel_name="ir.sequence",
        default=lambda self: self.env["ir.sequence"].sudo().create({"name": _("Dummy")}),
        required=True,
    )
    foo = fields.Integer()
```

Standard example of this code pattern that triggers the issue during
upgrades to 18:
https://github.com/odoo/odoo/blob/dd71d998/addons/account/models/company.py#L140-L152

closes odoo/odoo#185913

X-original-commit: c345a22aef0c23f8a6275ec84d672146559ae229
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>