[FIX] sale: make amount_to_invoice invisible in the views
When it was originally introduced on 'sale.order', the 'amount_to_invoice' field
was only intended to be used to trigger the credit limit warnings
on sale orders and invoices.

With time, it came to be used for other unintended purposes, which created
confusion regarding the computation method and ensuing fixes that did
more harm than good.

This is why, for the master version, it was decided to remove all unintended usaged of the field
in the views, and rename it to further clarify its meaning.

Here, in version 18, we are simply making the fields invisible.

task-4213628

closes odoo/odoo#185991

Signed-off-by: William André (wan) <wan@odoo.com>