[FIX] web: Hoot - fix improper fetch return types
Previously, mocked `fetch` responses' content types would rely on the
type of the body of the fetch call and not the reponse.

Now, it relies solely on the content type of the response/request or the
type of the returned value.

This commit also slightly refactors some internal values in Hoot mocks
for readability and performance.

Part-of: odoo/odoo#186403
Related: odoo/enterprise#73361
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>