[FIX] web: tests - mock server
This commit fixes several behaviors in the mock server:

- `pure` RPCs on the mock server would not throw errors and return the
error objects instead, which contradicts the precept of "pure" being
that the caller is fully responsible of the handling of the request
(i.e. if it should throw, it throws).

- Before this commit, route listeners on the mock server would match
against the start of a given route, meaning that listeners for "/a/b"
would also be called with "/a/b/c", while in most cases this is not what
we want (and that behavior can still be achieved with "a/b*"). This
commit ensure that the regex used to find route listeners is matched
against the whole route string.

This commit also refactors a few bits of the mock server to improve
readability and developer experience (e.g. allow lists of strings for
the `onRpc` helper). It also adapts some tests, either regarding some
of the changes above, or to enforce good practices across the test code
base.

Part-of: odoo/odoo#186403
Related: odoo/enterprise#73361
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>