[FIX] mrp: compute duration_expected
This is a non-deterministic bug.

Setup:
- install mrp, enable workorders and serial numbers
- create a storable product A with serial tracking
- create a BOM with 1 operation for product A

Steps:
- create a MO for 2 units of product A -> the duration expected on the WO
should be 120 minutes.
- confirm then plan the MO
- produce all -> you'll see a wizard for batch production (2nd wizard)
- generate serials then produce or prepare MO (doesn't matter which)

Issue:
Sometimes the duration of the initial workorder is correctly computed
-> 60 minutes, sometimes it is not (left as is). This is a mix of 2 'bugs':

First, we have the `self.env['mrp.production'].browse` via a set().
Because it's a set, the productions received are not always in order, so
the loop just after may process the last MO first. This triggers the 2nd
'bug'.

Second, the first MO (that was planned) won't trigger `_compute_duration_expected` at
https://github.com/odoo/odoo/blob/014ed907a2fab280e9fa6dd93b23a6a852700796/addons/mrp/wizard/mrp_batch_produce.py#L116
when changing `qty_producing` because of `is_planned`:
https://github.com/odoo/odoo/blob/014ed907a2fab280e9fa6dd93b23a6a852700796/addons/mrp/models/mrp_production.py#L730-L731
instead, it will go through the compute later because the second MO (created in the split,
but not planned) will trigger the compute for both ids via `recompute` because the original
MO id was still in the pending transactions:
https://github.com/odoo/odoo/blob/014ed907a2fab280e9fa6dd93b23a6a852700796/odoo/fields.py#L1413
However, the first workorder will already have its qty_producing set to 1 so won't pass in
`_get_duration_expected`.

Fix:
Use `OrderedSet` instead of set along with `reversed()` so that the loop will process the last MO first.
This way the pending compute for the original MO will correctly compute the duration
because it's WO's qty_producing will still be zero.

closes odoo/odoo#185877

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>