[FIX] web: hide superuser menu if not admin
Before this commit, the "Become Superuser" item in the debug menu was visible even if the user
was not an admin.
This was due to an error in the way that item was defined:
it used a key "hide" that was not used by the DebugContext.
Instead, it should return a falsy value, as the DebugContext uses a simple filter on the items built by their factory function.

After this commit, the superuser menu is only visible to admins.

task-4260306

closes odoo/odoo#190179

X-original-commit: 937f128988f408e9553de289dfc11a8a169a4a6a
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Bhaumik Pittalia (bit) <bit@odoo.com>