[FIX] base_vat: turkey tax id numbers
The Turkish National Identification Number (NIN) consists of 11 digits in the
format NNNNNNNNNNN where the last two digits are check digits.

The Vergi Kimlik Numarasi (VKN) is a 10-digit number in the format NNNNNNNNNN.
The Vergi Kimlik Numaras is the Turkish tax identification number used for
businesses and certain foreign individuals.

closes odoo/odoo#187701

Task: 4282883
X-original-commit: a2816c953a9444fd396670c7bfa0a697e1b3ee2d
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>