[FIX] base_vat: turkey tax id numbers
The Turkish National Identification Number (NIN) consists of 11 digits in the
format NNNNNNNNNNN where the last two digits are check digits.

The Vergi Kimlik Numarasi (VKN) is a 10-digit number in the format NNNNNNNNNN.
The Vergi Kimlik Numaras is the Turkish tax identification number used for
businesses and certain foreign individuals.

closes odoo/odoo#187228

Task: 4282883
X-original-commit: 7d6a1118a74faa84e58a1b21b7319cee98ae7d78
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>