[PERF] account: speed up `_fetch_duplicate_reference`
Description:
A warning is shown on the form view of invoices/bills, to warn when
a possible duplicate is present in the database. The query searching
for those duplicates contained disjonctions for matching conditions
for both 'in' moves and 'out' moves. Postgres doesn't plan well with
that disjonction (subplan scan with high-filter rate, non-indexable
due to the disjonction). Splitting that disjonction into 2 queries
(one for each type of in/out, lazily) leads to better plans and also
faster execution, since there is better segragation of the lookup
criteria (there is no need to check for a potential duplicate
customer invoice in the set of vendor bills for ex.).

Benchmark:
On next.odoo.com, opening an invoice form view,
the `_fetch_duplicate_reference`

|               | Before | After | Speed-up |
|---------------|--------|-------|----------|
| Timings (hot) | 1s     | 40ms  | 25x      |

closes odoo/odoo#185975

Signed-off-by: William André (wan) <wan@odoo.com>