[FIX] mail: no Threads and Invite People popover overflow
Before this commit, the "Threads" and "Invite People"
actions in Discuss app could overflow and have horizontal scroll.

This happens because the `ActionPanel` has some responsive width
when in a panel, but these panels could also be displayed in popover
in which this rule shouldn't apply.

This commit fixes the issue by fixing a constant width to these
action panels in popover, with similar width as the messaging menu.

Task-4292021

closes odoo/odoo#186102

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>