[FIX] spreadsheet: update chart data source on type change
Before this commit, when changing the chart type, the data source was
not updated. Which was a problem, because the data returned for
different chart types is different (eg. pie charts filter empty points).

That meant that when changing the chart type from line => pie we
displayed a pie with empty points, but when reloading the page the
data would change to the correct one.

closes odoo/odoo#187188

Task: 4306227
X-original-commit: 239fd1fb1111be6ea1816dcc94f7d9ceba9ae49d
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>