[FIX] spreadsheet: update chart data source on type change
Before this commit, when changing the chart type, the data source was
not updated. Which was a problem, because the data returned for
different chart types is different (eg. pie charts filter empty points).

That meant that when changing the chart type from line => pie we
displayed a pie with empty points, but when reloading the page the
data would change to the correct one.

closes odoo/odoo#186110

Task: 4306227
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>