[FIX] purchase_stock,stock_account: aml zero quantity case
**Current behavior:**
Manually adjusting a FIFO/AVCO product's cost creates a journal
items with non-zero quantity, despite there not being a logical
quantity associated with this entry.

**Expected behavior:**
AMLs generated for this move have 0 quantity.

**Steps to reproduce:**
*Having accounting, purchase_stock*
1. Create a storable product with `average` costing

2. Create a purchase order for some of the product with some
     arbitrary price unit, receive the product

3. In the product form, change the cost to a smaller value

4. In the journal items pivot view, select the `quantity` field
     from the 'Measures' dropdown in the top left, expand the
     stock valuation journal on the left axis, observe that the
     change in cost has created 2 balancing AMLs, each with
     quantity == 1

**Cause of the issue:**
This case was not handled and quantity is always set to 1 if
falsy.

**Fix:**
Create the AMLs created in the cost change handle with
quantity=0. While it does risk being overwritten to 1 if the
`display_type` of the line changes, the diff is minimal and
use-case is non-critical- which makes it the best solution here.

opw-4090620

closes odoo/odoo#180998

Signed-off-by: William Henrotin (whe) <whe@odoo.com>