[FIX] account: improve rounding methods UI strings
The rounding method 'HALF-UP' is confusing for regular users when they
see it in the UI. Replacing the user-facing term by the less specific
'Nearest' is clearer, since people naturally expect that to round halves
up.

Also the other UI strings were in all caps and have been changed to
capitalized strings.

opw-4266444

closes odoo/odoo#187306

X-original-commit: 8900a98e3ea5e4835b8377c9d0b202cfc3188faf
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>