[FIX] pos_loyalty: take into account discount proportions
Currently when multiple discounts are applied with loyalty rewards, the proportion
applied to each line is not computed and lead to incoherant values for the
next discount applied.

Steps to reproduce:
-------------------
* Create a promotion program rewarding 20% on the order for a minimum of
  5 products
* Create a promotion program rewarding 20% on a product category C1 for a
  minimum of 1 product
* Create 2 products of 1$, one of them with the catgory C1
* Open shop session
* Add 5 products (not C1)
* Add 1 product (with C1)
> Observation: -1.2$ in discount for the 6 products and +0.08$ for the
products in category C1

Why the fix:
------------
After this fix https://github.com/odoo/odoo/commit/85047eca7c0f28e3b174ba166c0d7a1d9965b6b8
all discounts were taken into account when added to `linesToDiscount`.

Since it was now counted we had remove this part to not count the discount twice:
```python
if (lineReward.discount_mode === 'percent') {
    const discount = lineReward.discount / 100;
    for (const line of discountedLines) {
        if (line.reward_id) {
            continue;
        }
        if (lineReward.discount_applicability === 'cheapest') {
            remainingAmountPerLine[line.cid] *= (1 - (discount / line.get_quantity()))
        } else {
            remainingAmountPerLine[line.cid] *= (1 - discount);
        }
    }
}
```

However when they were added to `linesToDiscount` the whole discount was taken
into account, and not just the portion that applies on the order line.

We revert part of the previous fix but we modify the added part to only apply
on rewards that have fixed amounts.

We also remove this part of the code:
https://github.com/odoo/odoo/blob/f2e49b31e501bc4c0a46f6e7319a8064ebe96ac2/addons/pos_loyalty/static/src/js/Loyalty.js#L1309-L1326

as it was found that `discountedAmounts` were alsways `{}`
and thus process was never doing anything.

opw-4284817

closes odoo/odoo#190181

X-original-commit: 159d3883b3588f866d6743c61a2d4ab573f7f70e
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>