[FIX] l10n_sa_edi: Use journal company for correct API URL
Issue:
When multiple companies are set up, if you are working in a company with
a different l10n_sa configuration and attempt to send an EDI document
for another company, the retrieved API URL may be incorrect depending
on the other company's configuration.

Steps to reproduce:
1. Create Company A and set `l10n_sa_api_mode` to production.
2. Create Company B and set `l10n_sa_api_mode` to sandbox.
3. Navigate to Company B (keeping Company A selected).
4. Create an invoice for Company A or attempt to send the EDI.

The API URL retrieved will be for Company B, while it should be for Company A.

New behavior:
The API URL is now correctly retrieved using the company linked to the
journal. Additionally, we now follow the same access as in lines 440 and
539 to fetch the SA environment.

opw-4253449

closes odoo/odoo#186266

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>