[FIX] stock: avoid move line reset when opening detailled ops
This reverts [1] commit b45e249f8f5f36cffdb1ff615953eac7bb45b995.

Steps to reproduce:

- Enable Multi-step routes
- Create a storable product: put 5 units on shelf 1 and 5 on shelf 2
- Create and "mark as todo" a delivery order for 3 units
- Click on the detailed operation and change the reservation from
  shelf 1 to shelf 2 > save (the subrecord not the record)
- Reopen the details operation
> The records get resets in front of you

Cause of the issue:

One of the main problems of [1] is that it modifies dirrectly the data's
of the props of the component used by the X2many dialog:
https://github.com/odoo/odoo/blob/7e01d83ac08f9b076e0cd7d7e644679d5264ff7d/addons/stock/static/src/fields/stock_move_line_x2_many_field.js#L48
As such, as soon as you mount that component, you will reset the data's
used by that same component. Worse, the component is doing this to
itself. Hence, if you did not perform a save before opening the details
operation you will reset the values to match the DB's data. Furthermore
this data change modifying the props of the component we are looking at
will force to reload the component with the newly hardcoded data (which
explains that you see your last change resetting in front of you).

In addition, the `quant_id` field of the stock.move.line model is a very
particular field as it is neither stored or computed but is used by the
create and the write for the stock.move.line data's to match some of the
quant informations:
https://github.com/odoo/odoo/blob/7e01d83ac08f9b076e0cd7d7e644679d5264ff7d/addons/stock/models/stock_move_line.py#L85
https://github.com/odoo/odoo/blob/7e01d83ac08f9b076e0cd7d7e644679d5264ff7d/addons/stock/models/stock_move_line.py#L310-L311
Therefore this field will never be set on a move. line that was not
modified in JS and is not saved yet. In order to take into account
the offset between the DB Data and the data's we are setting on the form
of stock move it is therefore necessary to perform an rpc to **guess**
what quant was used to create a given line. HOWEVER, as teached by the
JS framework formation:
```
The first rule of customizing Odoo in JS is: do it in python
```
So that you should not do that matching in JS but rather where it
belongs.

Finally, modifying the props of a componenet during the onMounted is a
super hacky way to use OWL as it is the first Hook for which OWL does
not destroy the componenet because of props inconsistency...

+ this commit already required an other fix: Commit ea4fca8faa4ee84693e645514085a65b92364842

opw-4294650

Part-of: odoo/odoo#186400
Signed-off-by: William Henrotin (whe) <whe@odoo.com>