[FIX] website: fix `rounded box` header missing `padding
This commit aims to fix a spacing issue within the `rounded box`
header template, which has no spacing in mobile, making the
first and last element stick to the border of the header.

To fix that behavior, we simply need to set a `padding` utility class
on Mobile as well.

task-3608124

closes odoo/odoo#186396

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>