[FIX] web: Hoot - systematically clear failed tests
Before this commit, failed tests would be registered in the local
storage, and would be cleared if either:
- the "run failed tests/suites" button was clicked
- the failed test passed in a subsequent run

This meant that clicking "Run failed tests" would potentially run tests
from several runs ago, which we may not be aware of and could be
bothersome.

Furthermore: aborted tests would be added to that list of failed tests,
while it is guaranteed that such tests would fail and it doesn't mean
much to consider them as such.

This commit:
- cleans up the failed tests list once a failed test has been registered
in a new run;
- does not add aborted tests to that list.

Part-of: odoo/odoo#186403
Related: odoo/enterprise#73361
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>