[PERF] sale_stock: add small cache for downpayment
Currently when checking if an account.move has downpayments we need
to call filtered on `self.invoice_line_ids.sale_line_ids`. This can
be pretty heavy, especially in
`_stock_account_get_anglo_sexon_price_unit` as this method will
be called for account.move.line linked to the invoice being posted.

This commit introduces a small method `_get_anglo_saxon_price_ctx`
that can be overriden to add data to the context that will be used
in `_stock_account_get_anglo_saxon_price_unit`. This is done
in `sale_stock` to avoid the computation of downpayments for
each account.move.line.

opw-4226593

closes odoo/odoo#186659

X-original-commit: e614841a0a36b27f5f80f23afaf45bf3ba5d6a47
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Aurélien van Delft (avd) <avd@odoo.com>