[FIX] web: adapt `FormStatusIndicator` alignment
Prior to this commit, the `FormStatusIndicator` buttons were not
properly aligned with the content of the breadcrumb in the control
panel.

This commit adjusts the alignment of `FormStatusIndicator` correctly
according to the content of the breadcrumb.

Also, to align the cog menu perfectly with the "save" and "cancel"
buttons, we've removed the border around the latter two.

task-3874495

closes odoo/odoo#186463

Signed-off-by: Pierre Paridans (app) <app@odoo.com>