[FIX] project_purchase: remove empty POs from the count of stat button
In project dashboard, the count of purchase orders in the stat button was taking empty purchase orders
into account. However, when clicking on it, only the POs containing lines were shown, which is not consistent.
We then remove the empty POs from the count, this is also to be consistent with the count of sales orders
which does not take the empty SOs into account.

task-4313971
version-18.0

closes odoo/odoo#186450

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>