[FIX] account: skip manual attachment in send&print wizard
The Send & Print dialog will not consider attachments that have the same name as removed ones.
Steps to reproduce:
- Create an invoice
- Confirm it
- Action menu > Print > Invoice, save the file (e.g. INV_2024_00001.pdf)
- Click Send & Print, wizard will open
- Select Email only
- Remove the generated attachment (INV_2024_00001.pdf)
- Add the INV_2024_00001.pdf attachment saved before
- Send the email

Issue: The invoice will not be attached to the email (you can check in the chatter).
This occurs because we filter out attachments by checking their names,
and the name is the same as the auto-generated one.

opw-4260237

closes odoo/odoo#187235

X-original-commit: 73a3893b71ede1767df7e36185263e29cb15888d
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>