[FIX] web: do not retain elements in tooltip service
Before this commit, elements on which tooltip were attached were used
in a `Map` as keys. That caused the elements to be retained even after
being detached from DOM. This commit changes the `Map` to a `WeakMap`
to not keep the element's reference and clears properties that kept
an element when the tooltip is closed.

closes odoo/odoo#187053

X-original-commit: 306b1a1b8f726dac6ac843a16176521dc02bc6a3
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>