[FIX] web: do not retain elements in tooltip service
Before this commit, elements on which tooltip were attached were used
in a `Map` as keys. That caused the elements to be retained even after
being detached from DOM. This commit changes the `Map` to a `WeakMap`
to not keep the element's reference and clears properties that kept
an element when the tooltip is closed.

closes odoo/odoo#186579

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>