[FIX] l10n_de: detach XML invoice
Steps to reproduce:
Install the app l10n_de and switch to a German company
Create an invoice
Send & Print the Invoice with XRechnung 
PDF and XML versions  are stored as attachments in the record (Chatter).
Try to delete them form the chatter. When deleting them, the unlink function only removes the link to account.move fields and saves the attachment with a new name: "document detached by user on date".
This works fine for the PDF version but not for XML one.

This is because `ir_attachment.res_field` is set to `ubl_cii_xml_file` for XML version of the attachment, but when filtering to detach them the filter looks for `ir_attachment.res_field` that are set to `ubl_cii_xml_id`.
opw-4273836

closes odoo/odoo#187597

X-original-commit: 6dd549f53d59eb3f8518082e6bd5b37a82d5670c
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Abdelaziz EL HAYYANY (azey) <azey@odoo.com>