[FIX] certificate: Handle subject parsing error
In version 42.0.0+ of the `cryptography` library, which we now use,
certain special characters in the subject/issuer fields of certificates
may cause parsing errors. While the certificates are recognized
correctly, attempting to access a subject or issuer attribute with
special characters sometimes throws an exception due to parsing issues.

Fix
---
This issue appears to be specific to recent versions of `cryptography`,
as the same certificate values can be successfully retrieved using
`openssl`. Currently, we only access the `commonName` attribute from the
subject field, which is not used in any critical way in our application.

Given this, and due to the blocking nature of this issue for customers
requiring specific EDI setups (e.g., in Mexico), we have opted to bypass
such cases for now by ignoring these parsing errors.

opw-4295754
opw-4293293

closes odoo/odoo#186633

Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>