[FIX] mail: im_status flickering
When a user disconnects from a device, the server assumes he is
disconnected until another device/browser says otherwise. However,
this can lead to small flickers. This PR fixes the issue by debouncing
the update of the im status field of the persona model. This way, there
is no flickering.

Steps to reproduce the issue:
- Open two browser windows with mitchell admin in the discuss app
(incognito + regular windows).
- Go to the chat with your self, where the im status can be seen.
- Reload one tab several times: you can sometimes see a flicker from
online to offline.

task-4236550

closes odoo/odoo#186764

X-original-commit: d3458bf9034ffeb2dd138377d9b6f60f38f6901c
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>