[FIX] mrp: don't write on invisible field
`uom_po_id` is invisible in the product view in 2 scenarios. the uom
group is not activated or the product cannot be bought.
As we are setting the default value anyway in those tests, it's better
to not give a particular value to avoid error about writing into
invisible field

runbot error : 65999

closes odoo/odoo#186792

X-original-commit: 65f5e734db0dfe4b8c742604958254bf1f7528d2
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>