[FIX] mail: correct title display in web notifications for group chats
Before this commit:

When a user posted a message in a group chat, the web notification displayed
`#False` as the title. This occurred because `record_name` was used as the title
but it was empty for channels with `channel_type = 'group'`, as the displayName
name is calculated on the UI side when displaying chat in Discuss.

After this commit:

The title of the web notification now correctly displays the group chat name,
ensuring recipients see accurate and complete group titles in notifications.

This commit also take guests into account in push notification
from new messages in channels as author. Group chat default name
also uses guest names for the conversation name.

closes odoo/odoo#188337

X-original-commit: d0e2c472ced48d491d12523ef72053f684a5a6b2
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>