[FIX] mail: correct title display in web notifications for group chats
Before this commit:

When a user posted a message in a group chat, the web notification displayed
`#False` as the title. This occurred because `record_name` was used as the title
but it was empty for channels with `channel_type = 'group'`, as the displayName
name is calculated on the UI side when displaying chat in Discuss.

After this commit:

The title of the web notification now correctly displays the group chat name,
ensuring recipients see accurate and complete group titles in notifications.

closes odoo/odoo#186756

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>