[FIX] web: Hoot - freezeTime stops performance.now
Before this commit, the inner counters of the `time.js` module would
effectively advance normally due to calls to `performance.now` even if
the time was supposedly "freezed" (call to `freezeTime`).

This caused some time values to be offset by a few milliseconds.

This commit fixes this issue by replacing calls to `performance.now` by
`0` when time is freezed.

Part-of: odoo/odoo#187034
Related: odoo/enterprise#74093
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>