[FIX] mail: update activity view after creating multiple activities
Before this commit:
When a user creates multiple scheduled activities and clicks the "Close" button
the newly created activities are not updated in the activity view.

After this commit:
When a user creates multiple scheduled activities and clicks the "Close" button,
the newly created activities should update and visible in the activity view.

Task-4057815

closes odoo/odoo#187095

X-original-commit: c968e3c5eabc1d41b92e6cefd1962f23c5951458
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>