[FIX] l10n_it_edi: do not mingle downpayment lines in negative lines dispath
The downpayment moves need to be reported in the FatturaCollegate and
also as real negative lines in the XML.

That worked in 16, but as we added the dispatch of the negative lines
on the line discounts of the positive lines, the negative downpayment
linked line got swallowed by the dispatch.  The fix is to not dispatch
downpayment linked lines.

opw-4274292

closes odoo/odoo#187457

X-original-commit: 919cc0f56b6242224e4fc0ce141dbc7a42dbea30
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>