[FIX] html_editor: setContent move selection
Before this commit, in the tests, when the setContent utils is used
to move the selection, the entire html is recreated. This implies a set
of unnecessary mutations in the dom.
This commit will therefore prevent the creation of all these mutations
by checking if the ‘content’ is different from that in ‘el’.

closes odoo/odoo#186817

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>