[FIX] web_editor, *: fix traceback when adding section background video
*: test_website

Steps to reproduce:
1. Drag and drop a text snippet.
2. Select background video and insert the video in the mediaDialog.
3. Click on the add button.

Issue:
A traceback occurs due to the tag name not being found for an element.
This issue was introduced in commit [1].

Solution:
This commit resolves the traceback by adding a `parentEl` condition in
the replace media on save function.

Note: this forward-ported version already includes the fix being added
by [2], which fixed the original tour having issues on runbot.

[1]: https://github.com/odoo/odoo/commit/36594d04a8909dd40ab6384f387c372ddae62345
[2]: https://github.com/odoo/odoo/pull/188214

closes odoo/odoo#188169

X-original-commit: 1460bb4df501f80cdb5ee411956bc4f2367bb9e4
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Co-authored-by: qsm-odoo <qsm@odoo.com>