[FIX] purchase: Correct display in purchase report
Changed the tax display in the purchase order template from
`t-esc` to `t-out` to ensure proper HTML rendering of tax labels.
This adjustment is necessary to display the tax descriptions
correctly without escaping HTML content. Before this commit writing a
tax description and then generating a purchase order report led to the
bad escaping of the tax description HTML content, which was causing
display issues in the rendered document.

This change addresses the need for accurate representation of tax
information in purchase orders, improving the clarity and correctness
of the document output.

opw-4313981

closes odoo/odoo#187177

Signed-off-by: Adrien Widart (awt) <awt@odoo.com>