[FIX] base: fix mexican address formatting
Issue:

When printing the PDF of a Purchase Order, if the company has multiple address, those appear on the same line.

Steps to reproduce:

- Create a Mexican Company with 2 address lines
- Create a Purchase Order
- Print the PDF of the Purchase Order

Cause:

The separator between the 2 address lines in the address format is set as a blank space.

Solution:

Replace the blank space separator with a new line.

closes odoo/odoo#188191

X-original-commit: 7ccef681bdcffab6dc25cf5ad0be8f3f598e55bc
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>