[FIX] l10n_eu_oss: don't ref inexistent tag
The aim of this commit is to prevent traceback when mapping local and
eu taxes to oss taxes.

Before the commit:
If the company chart template to use resolves to a CoA coming from a
module not installed, it will crash when trying to resolve the xml_id
for that specific localization.

After the commit:
If the l10n module isn't installed, it won't try to reference any
tax tag as the tax report line wouldn't be there anyway

Note:
1) This case is unlikely as if the customer has a vat number for a specific
country, it means they have a tax report to fill and thus must have
installed the related localization either as main CoA or as a foreing
fiscal position.
2) This was spotted through a runbot single l10n build error.

Nevertheless, it is something that is possible and thus should be
working smoothly.

runbot-100532

closes odoo/odoo#187052

X-original-commit: e9a36f21ff9675a1b11f19e3d402c6d6dcb0d50f
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>