[FIX] stock,repair: no longer break repair xpaths
Following #186831, the installation of the repair module would fail if
stock was previously installed without upgrading it first.

This is due to the new xpath defined in the repair module that reference
the changed form in the stock module. However, if stock wasn't upgraded,
then the new label / div don't exist yet, leading to an error as the
xpaths link to something that doesn't exist yet.

This means that this patch needs to work with two possible versions of
stock:
- The old one, with only <field> in the form
- The new one, with <field> and <label> in the form

Due to that singular situation, it made the declaration of xpath in
repair impossible to have a proper target, as both versions of stock
would require different targets and cannot co-exist.

To solve that situation, the choice was made to overwrite completely the
common ancestor of the two versions (i.e. the locations <group>) and
write it back in repair as if every previously declared xpath have been
applied to it.

closes odoo/odoo#186985

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>