[FIX] point_of_sale: fix IoT customer display SSL
Currently, when the IoT Box doesn't have a valid SSL certificate yet
when the user is trying to open the Point of Sale customer display it
triggers a request error related to invalid SSL certificate in a loop.
This is usually linked to the absence of the IoT subscribtion for this
particular database.
After this PR the user will see an error suggesting to check the SSL
certificate status which will help functional support to solve related
tickets more easily.

ticket-4175902

closes odoo/odoo#187248

X-original-commit: 067f2b95b07d18564304e47607ce472c81a9e333
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>