[FIX] spreadsheet: do not translate value of quarters in filters
Currently, when using a language that translates the abbreviation of quarters (Q) to anything else,
the formula `=filter.value(...)` would return the translated version, which then could not be used
on other formulas like `=odoo.balance`, this can be observed by opening the
Finance/accounting dashboard while in french and selecting a filter in quarter.

After this fix, the formula filter.value will always return Q1 to Q4 for quarter names.

closes odoo/odoo#187043

Task: 4274741
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>