[FIX] account: convert `quick_encoding_vals` to a Json field
Currently, if the `quick_encoding_vals` field becomes visible in a view
(e.g., by toggling "Show invisible elements" in Studio), the system
crashes. This happens because the field is currently a Binary field that
cannot be represented as a string.

Fix
---
The `quick_encoding_vals` field is an unstored computed field that outputs
a dictionary. To resolve the issue, we can safely convert it into a JSON
field.

opw-4241572

closes odoo/odoo#188259

X-original-commit: f85a8d10e3bde82b37642cd8568264fbeb0ac705
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>