[FIX] web: Hoot - properly display constructors
Some objects in JavaScript do not have an eplicit constructor carrying
the generic name of the object (e.g.: 'arguments').

Before this commit, this meant that sometimes a diff would appear
between 2 seemingly identical objects, while in fact their constructors
would differ (in the exemple of the 'arguments' object, the 2 would be
displayed as arrays).

This commit ensures that object constructors are clearly and explicitly
displayed in test reportings for all cases.

Part-of: odoo/odoo#187034
Related: odoo/enterprise#74093
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>