[IMP] point_of_sale: Do not display lot number that are not available
Only display lot/serial numbers if their forecasted quantity is available. Now when using a 2 step routes (inventory) we only show in PoS the serial numbers that are not already reserved for a stock move (like in sales).
To do this we return the `stock.lot` that have a related `stock.quant` which  is located in the `pos_config` default source location and have a quantity > 0.

closes odoo/odoo#190493

Task-id: 4320066
X-original-commit: 4b91652f2639c66c962d6830b3614e5325c5835b
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Manu Vaillant (manv) <manv@odoo.com>