[IMP] l10n_es_edi_tbai, account, sale: reverse from sale
When we have products on delivery on a sale order:
- we update delivered qties
- create invoice
- we diminish delivered qties
- create invoice -> should be a refund

We want this last refund to be linked with the
first invoice.  As there is only one, there is
no doubt that the two should be linked.

However, we only do this for the localizations where
the link is absolutely required, which is KE (enterprise)
and ES Ticketbai for now.

We also make sure in Ticketbai that the field is
sufficiently editable.

closes odoo/odoo#189617

X-original-commit: f3d3e0c612392f1a35b9c131d29b4f3bd70a9d36
Related: odoo/enterprise#75148
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>