[FIX] web: editable list's buttons not displayed on small screen
This commit fixes an issue where, in an editable list view on a small
screen, once trying to modify a record the "Save" and "Discard" buttons
doesn't appear. This is due to those buttons being put inside a
dropdown... but hidden on small screen.

Steps to reproduce (on a small screen):
- Install CRM module
- Go to CRM > Configuration > Pipeline > Tags
- Try to modify one of the list's record

=> The "New" button disappears, but the "Save" and "Discard" buttons are
not displayed

closes odoo/odoo#187160

Note: reported during the mobile tests conversion to HOOT in master.
X-original-commit: 4f325231f6f83481f18a736cfd9fdc3d32f4985c
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>