[FIX] web: Prevent traceback when users use shift click on several page
Steps:
    - Open any list view with more than 1 page
    - apply a limit if you don't have enough records
    - Toggle first record checkbox
    - Click on `o_pager_next`
    - Shift click on another record
    - Traceback

The problem arises when you change page after using range-selection with
the shift key.

Indeed, range-selection uses `this.lastCheckedRecord`, which is the last
record clicked on to start the selection, except that if you change page
(and if you use a filter), this element may no longer be in the current
page.

Currently, the code just checks that this element exists before attempting
to create a selection by range.

The fix consists in checking that this element exists in the list of
records before attempting anything.

opw-4284708

closes odoo/odoo#187244

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>