[FIX] l10n_fr_account: add ids to bank records in csv
Records in csv files also need an id column
to create a data entry, and to be able to load
the file in a mode other than init.

The file can be loaded on module installation
and all the records would be created without an
xmlid, but once the module is reloaded or upgraded
the file fails to load. It is a non blocking
error for the module but if someone is missing
the records then they will not be created, and
for the CI this is a failure.

closes odoo/odoo#187204

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>