[IMP] point_of_sale: Upgrade IoT to 2024-10-22 RPi OS
For the next IoT box image, we will upgrade to the latest
version of RPi OS Lite. This fixes an issue on bootup where
the display is not initalised (hangs waiting on `dev-dri-card0/start`).

Also removes the `wtmp` and `btmp` paths from `logrotate.conf`.
These paths were duplicated from existing logrotate config files,
causing the service to fail to start.

closes odoo/odoo#187689

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>