[FIX] pos_self_order: modify combo quantity
Currently if you try to modify the quantity of a combo product from the
cart in the kiosk, the price will not update.

Steps to reproduce:
-------------------
* Open a kiosk
* Add a combo product to the order
* Go to review the order
* Click on the + button
> Observation: The price of the order does not change.

Why the fix:
------------
In the backend we can see that the parent quantity was changed accordingly
but the child products quantity stayed at 1. We were not accessing the right
field when checking the child products.

opw-4283481

closes odoo/odoo#187304

X-original-commit: 5851539a892c19c374dc95f0cec384fd54a8f78f
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>