[FIX] project_account: display project updates using custom analytic plan
Steps to reproduce:
- Install Accounting > Settings > Enable 'Analytic Accounting'
- Project App > New > Project Settings > Edit Analytic plan
- Change Plan from 'Projects' to any other (except Internal)
- Manufacturing > Operations > Manufacturing Orders > New
- Product: 'Table Top'
- Miscellaneous tab > Analytic distribution: <Project account>
- Confirm > Produce All > Set quantities & Validate
- Project > ':' Menu > Project Updates (Dashboard in 18.0+)
- The cost of materials ($160) is not deducted in Project costs

Analytic distributions include different categories for each plan, this
is reflected in DB with a different column for each plan,
account_id/x_plan2_id/x_plan3_id/... which contain the project's
analytic account. In version 18.0 and above we allow each project to have
several analytic accounts (for the project, department, ...) but prior
to that we relied on plan_id to differentiate the project account category.

When checking for analytic lines to populate the project updates we
check the `account_id`, meaning only lines with projects having analytic
plan 'Projects' would pass that filter domain. We instead want to grab
the account_id from any of the plan columns.

opw-4282092

closes odoo/odoo#187275

Signed-off-by: Yoann Baron (bary) <bary@odoo.com>