[IMP] account: Fix dynamic PDF reports generation on send & print
Problem
---------
Additional attachments linked to an Email Template used in the Send &
Print are completely ignored. They should not, and it's especially
useful for Timesheets.

Steps
---------
1. Install Accouting, Timesheet and Sales
2. In debug, access mail templates
3. Go to the 'Invoice: Sending' record > Setting tab
4. In Dynamic Reports, add 'Timesheet'
5. Go to sale
6. Create an SO with 'Junior Architect' product for 10 units
7. Confirm the SO
8. Click the Timesheet widget on the SO form view
9. Register some hours for today (for example 3 hours)
10. On the SO Form view, click 'Create Invoice'
11. Select the option to invoice timesheeted hours and select from today
to today as the date range
12. Create the Invoice and post
13. Send & Print
-> The timesheet report appears in the send and print widget, when you
click send, the timesheet report is actually not sent.

Solution
---------
1. Add a placeholder in the send&print widget, the placeholders are
stored in the `mail_attachments_widget` used for such purposes. The
dynamic reports are defined with an new key: `dynamic_report` to
differenciate them from the `manual` attachments and the invoice PDF.
This new key eases the complexity to retreive and create the attachments
fom the placeholders list.
2. Add a method when sending the mails to generate the dynamic reports
only if the reports have not been removed in the S&P widget
(`skip != False`). Since the attachments placed in the mails are
a collection of elements from `mail_attachments_widget` that possess a
valid attachment ID. We add the relevant attachment data in the
mail_attachments_widget of the invoice.

task-4283972

closes odoo/odoo#188513

X-original-commit: b457fb7ddb9382a73eb8c9dc3ddeb04acfcd7582
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>