[FIX] account: autoinstall coa only if country matches
When new modules with CoAs are being installed, the matching CoA based
on country is autoinstalled if possible, but if multiple charts match
then we overwrite the autoinstall and use the one from the last module.
When the company has no country set, and the CoA being installed
doesn't set the country either we may get an error when installing, for
example,  without demo.
The only allowed coa when no country is set is the geenric_coa, used
and expected in many tests.
(shortened:)

This happens because we get an empty  here
https://github.com/odoo/odoo/blob/db3180ef6d5d028b73ee4bf0b0ba2df788eb3f34/addons/account/models/chart_template.py#L428-L431
That latter gives a  currency here
https://github.com/odoo/odoo/blob/db3180ef6d5d028b73ee4bf0b0ba2df788eb3f34/addons/account/models/chart_template.py#L442-L446

closes odoo/odoo#188093

X-original-commit: 84fb9cf722241ff120a159c59dabc113514fe124
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>
Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>
Co-authored-by: Alvaro Fuentes <afu@odoo.com>