[FIX] stock: parse warehouse context key on use
Steps to reproduce:

- Inventory > Configuration > Warehouse Management > Warehouses
- Create a second warehouse
- Go to Inventory > Products > Products
- In the search bar type `foo` string > Search Warehouse for `foo`
- Click on any product Kandan record
- Click on the Forecast smart button of the product
> Traceback

Cause of the issue:

Thanks to the dummy `warehouse_id` field of the `product.template` model
a `warehouse` context key can be set in the context from the search bar:
https://github.com/odoo/odoo/blob/a72763acfc4d83ae2aadad2e807547c5b1819002/addons/stock/models/product.py#L685
https://github.com/odoo/odoo/blob/a72763acfc4d83ae2aadad2e807547c5b1819002/addons/stock/views/product_views.xml#L78
This trick is notably used in order to take the warehouse into account
in the computation of the various quantity fields associated to products
by generating custom location domains:
https://github.com/odoo/odoo/blob/a72763acfc4d83ae2aadad2e807547c5b1819002/addons/stock/models/product.py#L136-L137
https://github.com/odoo/odoo/blob/a72763acfc4d83ae2aadad2e807547c5b1819002/addons/stock/models/product.py#L250-L255
However, since that context key was added via the search bar, it type
might be: a string, an integerId or a list of either/both other types.
(E.g. to create a list of a string and integer type and select a string
and then type a string allowing you to find a real warehouse id that you
can select after clicking on the dropdown arrow). Therefore, in order to
be properly used, this context key needs to be parsed to be used
properly as done in the `_get_domain_locations` for instance. However,
the `warehouse` context key is used at many other places in the code,
each time expecting a single integer id, and since the warehouse context
key is not cleaned from one action to an other you are technically able
to provide a string where the code is expecting a an integer.

Fix:

The proper fix of this use case would be to change the context key name
used by the search view to only match flows expecting such a context.
However, this change is not stable as it requires to modify a view
and hence can't be applied before master (18.1). As such, in prior
versions, we add a context parser for to extract a single integer id
from the `warehouse` context key where it is used but unexpected to
match something else.

Note:

This patch will improve other fixes and hopefully prevent yet unoticed
error raised by this issue:
903d8beeea5d332e556ef81e231a3b8b4c51cd45 and afa7c6bf25c9de5fc0c878faa29e1cc35bc11805

opw-4290818

closes odoo/odoo#187808

X-original-commit: 856409a1fb35c6c49fe4c404931587a95d99d370
Related: odoo/enterprise#74188
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>