[FIX] website_sale: fix layout shift in cart
Product images are lazy-loaded on the cart page, causing them to be displayed
later than the rest of the page. As a result, some parts of the page were
shifted once the image was loaded (since the image had no size before loading).

This change wraps the image in a fixed-size container to avoid the layout shift.

Incidentally, this change also fixes some alignment issues when the cart
contains images with various aspect ratios.

opw-4029017

closes odoo/odoo#187297

Signed-off-by: Louis Tinel (loti) <loti@odoo.com>