[FIX] web_editor, website: fix traceback when clicking on insert media
Steps to reproduce:

- In Website edit mode.
- Drag and drop a "Text" block into the page.
- Click on the first paragaph of the "Text" block.
- click the "Insert media" button of the text toolbar.
- A traceback occurs.

The bug was introduced in commit [1], where we checked that
"this.props.media" was not undefined before calling the "closest"
function on it. However, this was incorrect because, in cases where
"this.props.media" was false, the "closest" function was still called,
causing the traceback.

With this fix, the "closest" function is not called if
"this.props.media" is false.

[1]: https://github.com/odoo/odoo/commit/625c1aa71e17882848dd827f3be6162e49cbc8ca

opw-4335833
opw-4328280
opw-4334279
opw-4328490
opw-4334012
opw-4332500
opw-4332028
opw-4332684
opw-4329687
opw-4331635
opw-4331371
opw-4331287
opw-4332437
opw-4331144
opw-4329646
opw-4328364
opw-4323197
opw-4325959
opw-4322945

closes odoo/odoo#187505

X-original-commit: 6539be6062620bc532a1b0007ec0cc931634df67
Signed-off-by: Colin Louis (loco) <loco@odoo.com>
Signed-off-by: Benjamin Vray (bvr) <bvr@odoo.com>