[FIX] account_edi_ubl_cii: create partner during import
Partner was not created during import because `bank_details` was an
iterator instead of a list/collection. So once it was consummed for the
search, nothing was created.
Moreover, this function is called in the context of a file import which
may have a default journal_id which is not the same as the journal of
the partner.

closes odoo/odoo#187464

X-original-commit: 5511f475c94023d198ff007b5a6149609ba0396e
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>