[FIX] pos_restaurant: test_05_tip_screen tour
Before this commit, we check the order of lines but what is important
is to check the price. When the tour is too fast, the order of commands
can be not well ordered.
In this commit, we adapt trigger to ensure the price is good but not
the order.

Part-of: odoo/odoo#187678
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>