[FIX] mrp: expected duration with different BoM UoM
Problem: When a BoM has a different UoM than the product, it will calculate the number of cycles needed for the work order using the product’s UoM. However, during the cycle time calculation, it uses the BoM’s UoM to calculate the number of cycles used in the previous work orders. It then uses this cycle number calculation to calculate the duration of each cycle. This can cause the expected duration for a work order to be calculated incorrectly.

Purpose: Changing the cycle time’s calculation to use the product’s UoM will make it consistent with the number of cycles calculation on the work order.

Steps to Reproduce on Runbot:

1. Create a new storable product.
2. Create a BoM for this product, but set the UoM to Dozens.
3. Navigate to the Operations tab and add an operation.
4. Set the Duration Computation to Compute based on tracked time.
5. Create a manufacturing order and confirm it.
6. Navigate to the Work Orders tab, set the Real Duration to 20:00, and mark it as done.
7. Create a new manufacturing order.
8. Navigate to the Work Orders tab and observe the Expected Duration is 240:00.

opw-4239248

closes odoo/odoo#189445

X-original-commit: 8b7a5283ef74a788a84ab33c7fe40de43a9786b0
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Richard Prosser (prri) <prri@odoo.com>