[MOV] test_mail: create a base class for mailgateway tests
Just to prepare and ease (future) split of tests between main features.
And ease addition of loop-related tests in a specific class.

Move loop test in its own class, as we are about to add some of them.

Note that diff is ugly, but we just moved a test and some setup, there
is no test change in this commit. Bisous.

Task-3895869

Part-of: odoo/odoo#187696
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>