[FIX] mail: mark messages as read when opening a channel on mobile
Before this commit:

Unread messages in a channel or chat were not being marked as read when the user
opened the channel in mobile view. This occurred because `thread.markAsRead()`
was not being called upon opening the channel.

After this commit:

Opening a chat or channel in mobile view correctly marks unread messages as
read, ensuring consistent behavior.

Task-4320251

closes odoo/odoo#189616

X-original-commit: cc1209fc9c269b06cbe6f8ff1f2439b1bd422bb1
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>