[FIX] stock: compute dirty move lines in get_move_line_quant_match
Issue:

The `get_move_line_quant_match` was introduced in cc64e99c047807b3691a1a80cdd09464b89aff64
to reconcile js and db data's during the edition of stock moves form.
However, the assignment of deleted move lines is not correct:
https://github.com/odoo/odoo/blob/0fc3752cacf25b6fd26b85302b1f23d5a35c8cd7/addons/stock/models/stock_move_line.py#L296
Its value should correspond to the move lines present in db but
absent from the JS datas (removed from the form edition but with unsaved
change). The problem here is that self is the set of records present in
JS but flagged as dirty since either the quant or the quantity did
change:
https://github.com/odoo/odoo/blob/0fc3752cacf25b6fd26b85302b1f23d5a35c8cd7/addons/stock/static/src/fields/stock_move_line_x2_many_field.js#L95
but it should be the complete set of move lines present still present in
JS to work properly.

opw-4294650

closes odoo/odoo#187732

Signed-off-by: William Henrotin (whe) <whe@odoo.com>