[FIX] html_editor: applied color should be selected in solid tab
Before this commit:

When there is a text having a text-color or background color, opening
color selector doesn't highlight the button matching with that
selected color in solid color tab.

After this commit:

Opening color selector highlights the button matching with that
selected color in solid color tab.

task-4341528

Part-of: odoo/odoo#187834
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>