[FIX] account: prevent error when opening invoice preview without invoice date
Currently, an error occurs when the user attempts to preview an invoice, and
invoice date is not available.

Step to produce:

- Install the 'account' module.
- Create a new invoice, add a customer name and invoice line, and add a
'Payment Terms' which have an 'Early Discount' available.
- 'Cancel' this invoice.
- Click on the 'Preview' button (ensure that the invoice has no date).

TypeError: unsupported operand type(s) for +: 'bool' and 'relativedelta'

An error occurs when the system attempts to calculate the discount days with the
invoice date, but the invoice date is not available there.

To resolve this issue, we hide the preview button on canceled invoices.

Sentry-6006569495

closes odoo/odoo#188134

X-original-commit: d4f427106ca4228eca74e37e9dedce055fc4adfd
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Meet Gandhi (mega) <mega@odoo.com>