[IMP] account: put default taxes on demo products
Right now, when a demo company is created, the demo
products do not get taxes in that company.  (even in
the main company)

It is because of https://github.com/odoo/odoo/pull/173803
It is treating however an exception case with tips that if a
product does not have taxes in the initial company, we won't
give it taxes in the new company.

But this means that all demo data won't have taxes unless we put
them explicitly at least in one company.  That is why we explicitly
put the taxes on demo data, but provide some hooks for exception
cases like tips (and combo products).

We only do it for the main company, so the mechanism in the
referred PR will do its job for the other companies.  That
way we also only need to give exceptions for xml_ids that
do not come from a module that depends on accounting.

X-original-commit: 54058a9db0eb6bf1032938d51ec6232c3c88ba62
Part-of: odoo/odoo#189394
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>