[FIX] account_edi_ubl_cii: Fix wrong warning on Credit Note
The previous warning asking to add a Recipient Bank  was shown
incorrectly in previous versions in case of Credit Notes.
With the rework, we escalated this warning to a danger warning that
prevent from sending, as in case of Customer Invoices it is a
blocking error.

The combinations of those two is now preventing Credit Notes to be sent with
an UBL XML file.

See: https://github.com/odoo/odoo/commit/9e769e1b11f22890e5245859053bc8dd31e42634

task-no

closes odoo/odoo#187782

Signed-off-by: Laurent Smet (las) <las@odoo.com>