[FIX] html_editor: used custom colors appearing twice
When a text having a text-color is selected, opening
custom color picker shows that color twice at the top.
It happens because that specific color of selected text
is stored in this.usedCustomColors when opening color
selector. When color selector component is renedered,
it shows usedCustomColors button with currentCustomColor
and button appears twice.

task-4341528

closes odoo/odoo#187834

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>