[FIX] core : _name_search on fields with inconsistent types
Steps to reproduce the issue:
- go to currencies;
- search for a rate with something else than a date;
- it crashes.

This happens because the currency rate model is searched with two
fields, a date field (name) and a float field (rate).  This crashes when
converting the value to match against, because the value cannot be
serialized in SQL both as a date and a float.

What we do in such a case is to explicitly convert the value to the
field's type.  If the conversion fails, we simply ignore that part of
the domain.

opw-4278234

closes odoo/odoo#190495

X-original-commit: 44fbe4d6497e0c440ead93b5f5af7311b5de1459
Signed-off-by: Raphael Collet <rco@odoo.com>