[FIX] product: backport uom_category_id to match purchase/sale UoMs
In the purchase view, the user can see UoMs of all categories for the
vendor bill, but can only pick the one in the same category as the
sales price unit, as other options will get silently reverted by an
onchange. This is logical, but confusing for the user. The addition of
a domain for the UoM category prevents units of other categories from
appearing in the view.

Partial backport of 33d7474b07abf8ae161364bcf29ae72cd04384ba

closes odoo/odoo#189159

Opw: 4212814
X-original-commit: 5d3d4e66d8c63179c59e1b8a6cffb3ca00e1f0ec
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>