[FIX] web: allow event propagation on autocomplete input click
**Problem**:
This commit:
https://github.com/odoo/odoo/commit/655e1ea63cf469937c53762e93f93da40b6b9099#diff-fa2db531fe59b8ff5fce5cd3d124590810738769806d28e79c5c585d822938d8R12288
breaks the behavior of the `list_renderer` when an autocomplete is present in the same view. Specifically, it prevents discarding a row when clicking outside the input.

**Solution**:
Revert the commit, as the issue it aimed to fix (opw-3180055) is now resolved by the addition of the "Search more" functionality, rendering the commit unnecessary.

**Steps to reproduce**:
1. Open the *Purchase* tab on a product form.
2. Add a vendor line (with or without selecting a vendor).
3. Focus on the vendor line, then click to add *Vendor Taxes*.
4. After selecting a tax, observe that the focus incorrectly shifts back to the vendor line instead of remaining on the *Vendor Taxes* input.

opw-4263276

closes odoo/odoo#188820

X-original-commit: 68589c2a64236490bbd27f3476e99d548f5c095d
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Walid Sahli (wasa) <wasa@odoo.com>