[FIX] barcodes_gs1_nomenclature: clean FNC1
Issue
=====

When a GS1 barcode is scanned, if it starts with a FNC1, it can't be
parsed.

How to reproduce
================

- In Inventory Settings, select "Default GS1 Nomenclature" as the used
  nomenclature;
- Create a product with a valid EAN13 as barcode (eg. 1234567891231);
- Open the Barcode app, then the Inventory Adjustments;
- Scan a GS1 barcode for the created product who start with the FNC1
  (eg. "\x1D0101234567891231") -> The product's barcode is not decoded.

Solution
========

To fix this issue, before to be parsed, the given barcode will be
cleaned and if it starts with a FNC1, it will be removed from the
string.

opw-4118637

closes odoo/odoo#188879

X-original-commit: 4d69fd2b39e0c92477fbebee7ff128eedc9572b0
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>