[FIX] account: prevent home menu redirection on outstanding click
Steps to reproduce
==================

- Click on an invoice
- Register a payment where the amount exceeds the invoice amount
- Switch to another invoice with the same customer
- Switch the active notebook to "Journal Items"
- Click on "Outstanding credits"

=> You are redirected to the homepage

Cause of the issue
==================

Before 16.0, it used to be possible to click on a link where the target is
inside another notebook.
In that case, the notebook was switched before scrolling to the anchor.
With owl, it is no longer possible to query where the target is as the
DOM inside other notebooks is not yet rendered before switching to them.

Solution
========

If the anchor is visible, the link will work as intended, otherwise, the
redirection to the homepage will be prevented

opw-4344125

closes odoo/odoo#189488

X-original-commit: b7439e821f0f6fccaf96b64e4bb24b8712c73600
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>