[FIX] mail: mention should not overlap in multi-lines on Safari
Before this commit, rendering of `@`mentions in Discuss on Safari
was bigger than line height, which results in unintended UI overlap.
This is mostly visible and problematic when 2 successive text lines
have mentions in the same horizontal range.

This happens because discuss mentions are `<a>` and anchor links
have `display: inline`. As a reminder, `inline` does not respect
top/bottom padding and margin of element in its container so this
is very likely prone to overlap vertically with sibling nodes.
`inline-block`, on the other hand, take vertical padding and margin
into account so it doesn't overlap. If the element takes more space,
then siblings are pushed accordingly.

This commit fixes the issue by setting `display: inline-block` to
all discuss mentions, so that there's no risk of overlap in
multi-line especially on Safari.

closes odoo/odoo#187893

Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>