[REF] web_tour: replace _legacyIsVisible by hoot isVisible
In this commit, we replace the use of _legacyIsVisible by the visible
option of hoot which refers to isVisible function of hoot-dom.js.
This change implies changes in a few tours.
For information, _legacyIsVisible accepts elements with a width of 0 or
a length of 0 (which is a bit absurd) while isVisible does not, which
implies the changes where we just add :not(:visible) to the trigger.
The other main modifications implies tours where you have to hover the
trigger to then see the element you want to click on.

In this commit, we also improve error logs.

task~3974087

closes odoo/odoo#187894

Related: odoo/enterprise#74171
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>