[FIX] mail: livechat window of visitor correct header avatar spacing
In the backend, the avatar and conversation name is a clickable
button to show more thread actions. When this is clickable, there's
some spacing to align avatar nicely.

In livechat from visitor perspective, however, the avatar was not
spaced properly. This come from avatar and conversation name not
being clickable to show more actions, because all actions are
available in quick on the right side of header.

This commit adds some spacing at beginning of header content
of chat window when there's no actions menu, to compensate with
the padding of the actions menu button.

closes odoo/odoo#187920

Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>