[FIX] website_project: remove tracking on partner_name field
Before this commit, the `partner_name` was tracked in
`project.task` model because `name` of `res.partner` model is
tracked and so the related setup will keep the value set on `name`
field of `res.partner` to also make the related field tracked.

This commit removes the tracking on the related field in task since it
was not expected. Moreover, we will get a duplicated tracking message
with `partner_id` field which is also tracked in `project.task` model.

closes odoo/odoo#187934

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>