[FIX] purchase_stock: decrease received quantity if no return
In case the quantity of a purchase order line is negative the generated
picking will be a delivery but not a return. Meaning validating this
delivery will not update quantity received on the purchase order line.
This commit makes sur the `to_refund` key on stock move impact the value
of `qty_received` only if the stock move is actually a return.

closes odoo/odoo#187946

Opw: 4259022
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>