[IMP] mail: update template order in full mail composer
In Odoo 18.0's full mail composer, we replaced the default template
selector with a custom dropdown menu, enabling users to create, update,
delete, or select mail templates directly within the composer. The
dropdown initially suggests up to seven templates. When the template
isn't among the suggestions, people have to click on the "Search More..."
button to select the template they want. This adds a few additional steps
in the process of selecting a template.

To avoid that, we'll prioritize displaying templates created by the user
first, as they are more likely to be selected. This change is designed
to enhance user efficiency by making the most relevant templates
immediately accessible in the dropdown menu.

see: https://github.com/odoo/odoo/pull/171942

task-4348578

closes odoo/odoo#187953

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>