[FIX] web: tests - 'contains' error handling
This commit adds better error handling to the `contains` helper, to
ensure that an action is called and that it doesn't produce dead code.

It also adds some missing Hoot helpers to the 'contains' API: 'dblclick',
'keyDown' & 'keyUp'.

A few tests have been added to ensure this all works.

Part-of: odoo/odoo#187034
Related: odoo/enterprise#74093
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>