[FIX] product_matrix: avoid crash on generated PO lines
Steps to reproduce:
- Create a product having the following attributes:
-- Color, which has `create_variant`: 'always'
-- Custom, which has `create_variant`: 'no_variant'
- For each attribute, add a few values
- Create a sale order for that product, using a combination of both
attributes
- Go to Inventory/Operation/Procurement/Replenishment
- On the corresponding product, click on 'Order Once'
- On the created Purchase Order, try to edit the product through the
configurator

Issue:
A traceback will appear, as the generated matrix for this product will
be a collection of pairs from [Color,Custom] values. But as the moves
have no `product_no_variant_attribute_value_ids` to store that
information, that part of the information will be lost and the generated
orderpoint from which the PO is made will only be using an existing
product.product.

While this does not solve the issue, it at least allows to open the
product configurator even though the line is faulty.

Note: Same issue appear through a MTO flow, although in this case the
product description will be correct. But trying to open the configurator
would still lead to a traceback anyway.

opw-4197302

closes odoo/odoo#188031

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>