[FIX] stock: set the current sn location on sml
Using the fields `lot_ids` on `stock.move` will create a line
with the given serial. However, it will use the move's location
as reference. But the serial could be in a sublocation.

Currently the user has a warning that display the real location
of the serial but it's annoying to do it when the sytem know.
So set the current location as default and remove the warning
for this case.

opw-4342448

closes odoo/odoo#188061

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>