[IMP] core: avoid user specific company dependent fallback
when an ir.default record with user_id is created, the result will be used as
the fallback value for undefined company dependent column value. It will cause
user dependent fallback which is logically wrong.

This commit forces the fallback value to be fetched with SUPERUSER to make
fallback always not user dependent and increase the cache hit rate for
_get_model_defaults

closes odoo/odoo#188065

Signed-off-by: Raphael Collet <rco@odoo.com>