[FIX] stock: avoid merging moves coming from diffrent kit
Steps to reproduce:

- In the settings enable Multi-Steps Routes
- Inventory > Configuration > Warehouse Management > Wareohuses
- Put you warehouse in delivery in 2 steps
- Create 4 storable products: Kit 1, Kit 2, content, foo
- Create 2 boms of type kit: one for Kit 1 and one for Kit 2, both with:
  - 1 x content
- Create and confirm a sale order with two lines:
  - 1 x Kit 1
  - 1 x Kit 2
> On the associated pick, you can see 2 distinct moves: one per kit
- On the SO add a line for 1 unit of foo and save the SO
> On the associated pick both kit moves were merged.

Cause of the issue:

When you add 1 unit of foo, the `_action_launch_stock_rule` will be
called in order to create and confirm the associated stock moves. During
the `_action_confirm` of this move a `_merge_moves` will be called to
determine if the move can be merged with any other already existing one.
However, this merge operation will actually be performed on each of the
move present in the "candidate_move_set" which includes all the moves
of the already existing picking:
https://github.com/odoo/odoo/blob/8c5ad7621fe8370da4c28d42a51ba06e15ce42ee/addons/stock/models/stock_move.py#L1015-L1027
https://github.com/odoo/odoo/blob/8c5ad7621fe8370da4c28d42a51ba06e15ce42ee/addons/stock/models/stock_move.py#L1011-L1013
This is problematic since the `disting_fields` used as a merging
criterion are based solely on the records present in self and not
on the candidates on which the merging might be performed:
https://github.com/odoo/odoo/blob/8c5ad7621fe8370da4c28d42a51ba06e15ce42ee/addons/stock/models/stock_move.py#L1021
In our case, the reason why the two kit moves where not merged the first
time that they were confirmed was because the `bom_line_id` was added to
the distinct fields by:
https://github.com/odoo/odoo/blob/8c5ad7621fe8370da4c28d42a51ba06e15ce42ee/addons/mrp/models/stock_move.py#L605-L609
But on the action confirm of the foo move and since this move is not
assocaited with a `bom_line_id` of phantom type, it will not and
the two kit moves will be considered as good condidates to merge into
one an other.

Fix:

The `distinct_fields` used as a merging criterion should be based on the
set of all moves that are  considered to be merged rather than to the
moves that we initially wanted to merge.

opw-4337128

closes odoo/odoo#188101

Signed-off-by: William Henrotin (whe) <whe@odoo.com>