[FIX] l10n_{cz,sk}: View company_registry based on country_code on res.partner view
Using fiscal_country_codes in the invisible condition can result in the field being displayed multiple times, as multiple countries may meet the condition. Instead, we are using country_code, as it is more logical for the field to be visible only if the company is located in that specific country.

closes odoo/odoo#188099

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>