[FIX] hr_holidays: fix activity when time off officer is not set
Steps:
- Install the hr_holiday module
- Configure Time Off Type with validation set to `Approved by Time Off Officer`
  or `Both`
- Create a Time Off request without setting a Time Off Officer.

Description of the issue/feature this PR addresses:
When the validation type is set to `hr` or `both` and no Time Off Officer is
configured, an activity is incorrectly generated upon creating a Time Off
request.

Fix:
This PR resolves the issue by ensuring the responsible user is set to empty when
validation is `hr` or `Both` and no Time Off Officer is configured.

task-4351688

closes odoo/odoo#194386

X-original-commit: b5537170a76e53a911e08a42fa0cb3f53fcb50a0
Related: odoo/enterprise#77430
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>