[FIX] stock_account: modify the category, triple the valuation account
Following commit 0abcb6b10602d68883125c9bdc951507b2fd92e1
When modifying the category on the product template, it doesn't go
through the write of `product.category` but the write of
`product.product`.

So we modify the method to be defensive and act as before the commit
if there is no context.

opw-4315793

closes odoo/odoo#188702

X-original-commit: a5e8a413142a912c6dac8b71400acfa1375d485e
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: William Henrotin (whe) <whe@odoo.com>