[FIX] web: RelationalModel: reset offset if no record
Open a list view with multiple pages. Once opened, simulate some
records are deleted (e.g. in another tab, delete some records).
You must delete enough records s.t. there one page less than when
you loaded the list in the first tab. Then go back to the first
tab, and try to go to the last page.

Before this commit, it displayed an empty list, and the pager
was a bit lost (e.g. it could display something like 4-3/3).
With this commit, we detect that we tried to load records with an
offset, but the result was empty, so we reset the offset to 0 and
try again. We thus land on the first page, with a correct pager.

Task~3736927

X-original-commit: 7d8d4fa18f9bf887aa2ee857d82d7154914b5471
Part-of: odoo/odoo#188518
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>