[FIX] l10n_gcc_invoice: display Arabic and English in product description
Steps to reproduce:
- Install "l10n_sa" and the Arabic language
- Create a product and translate its name to Arabic
- Create an invoice and print it
- In the invoice lines under the description column only one translation displays and not both as it should.

Cause:
The code was changed to only display the current language in this commit (02e2fec#diff-750f6aa571f288e1caf97eb01fc4ecaf8efe8d017a6310df8a603d251e0a7930R81)
The goal of this commit was to prevent the duplication of the product name because it is already contained in line.name.

Solution:
The way line.name is computed is by taking the sales description (or purchase one) and concatenating it with the product name. It can also be changed manually in the invoice form view under "label".
So to display the product name in both languages and this label without any duplicate this commit adds conditions:
- the label is always displayed
- the English name is dipslayed if the label does not contains it
- the Arabic name is displayed if the label does ont contains it and if it is different from the English one (ie if it is translated)

opw-4187577

closes odoo/odoo#188437

X-original-commit: f1fa69f6a924a68bad4b10e6854f8caae89b43c3
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>