[FIX] web: many2many_tags_avatar: wrong background color
Before this commit, the invalid `many2many_tags_avatar` input has two layers
of transparent background-color so the two colors stack one uppon the other
one and so we had a inconsitancy color with others invalid field.

This commit adds the `.o_field_many2many_tags_avatar` so the following CSS rules
are also apply like on the `.o_field_many2many_tags`. And so the child input has
`background-color: transparent` when the input is invalid.
```css
.o_input {
  --o-input-background-color: transparent;
}
```

closes odoo/odoo#188348

X-original-commit: 32d9a3959c0b8f380c91240da3cfcdebee00310c
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>