[FIX] web: DomainSelector: shorten descriptions of long lists
When adding a custom filter, we can have a domain of the form

["id", "in", [
    1415T215,
    1545481,
    ...
]]

with a very long list of ids. In this commit we make the domain description
contain only the 20 first ids of the list. If we don't do that, in some universe,
we obtain a (unscrollable) search bar facet that is not fully visible when
confirming the domain.

closes odoo/odoo#188438

Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>