[FIX] point_of_sale: extend customer search fields
Currently, when searching the through the customers loaded in pos you can
search using the zip code but if the customer is not loaded you cannot search
those in db using the zip code.

Steps to reproduce:
-------------------
* Create a customer and set his zip code.
* Open pos shop
* Search for customers
* Enter the zip code, normally you shouldn't see the customer
* Select "Search more"
> Observation: No customer found

Note: If you search with the customer name it will find it.

Why the fix:
------------
Adding a few search values when loeading customers in the session.

opw-4334412

closes odoo/odoo#189164

Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>