[FIX] website: fix media dialog insert media tour
Since this commit [1], where the "media dialog insert media" was added,
the runbot sometimes fails because the tour contacts the
"media-api.odoo.com" API, which is not allowed in a tour.

This commit fixes the issue by inserting an icon instead of an image, as
the icon does not require any API call.

[1]: https://github.com/odoo/odoo/commit/2231148db55cfc6fa3f3f5b2b5c5250b74f2feac

runbot-107920

closes odoo/odoo#188480

Signed-off-by: Colin Louis (loco) <loco@odoo.com>