[FIX] web: editable grouped list: correctly leave edition
This commit fixes two issues with the editable grouped list view.
In a group, click on a row to switch it into edition and edit an
input field (e.g. a char). Do not blur the input. From this state,
 - clicking on the group header to close the group, or
 - clicking on "Add a line" to add a new record
would both lead to the change being lost.

This commit ensures the keep the change in those flows by correctly
calling `leaveEditMode` before closing the group or adding the new
record.

opw~4174315

closes odoo/odoo#188784

X-original-commit: 3d031364cd3065793b2e2fec30157f4653c26a35
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>