[FIX] html_editor: getAdjacentCharacter
This commit fixes the case where the selection received as argument is
not in a leaf node. In this case, the function would fail to return the
correct character before/next to the cursor, which affects the behavior
of navigating with the arrow keys around invisible characters (like ZWS
of NBZWSP).

task-3522395

Part-of: odoo/odoo#189203
Related: odoo/enterprise#75045
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>