[FIX] account_peppol: Fix bad write override in case of multiple records
When the write is called with multiple companies (like the test TestAccountComposerPerformance),
the values are updated with the values of the very first company. This will write a new peppol_endpoint
on all companies. However, in this test, the very first company is a BE one but another is FR.
Then, this write makes an inconsistency between the original FR EAS and the new BE endpoint.

closes odoo/odoo#188985

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>