[IMP] hw_posbox_homepage: better device list
Before this commit, the device list only showed
the identifiers of each device, which aren't
very readable.

After this commit, the devices are grouped by
device type, with each device type having an icon.
Each device then shows it's name and identifier,
as well as the last sent value if applicable.

task-4364479

closes odoo/odoo#189155

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>