[FIX] project_task,hr_recruitment: include ticket author in stage update email
The issue occurs when an email is sent to create a `project.task`.
If the sender of the email is an existing user, this user is later set as the author
of any outgoing emails related to that task. Since the system is configured to skip
sending emails to the author, no email is delivered to the sender in this scenario.

Reproduce
---
- -i project,contacts
- add `example.com` to domains in the settings
-  create new project with an alias: `project@example.com`
	- Have a stage with some "Email Template" set
- send email creating project tasks as a portal user
- BUG: email about stage update is NOT sent (even tho task appeared in stage)

opw-3941928

closes odoo/odoo#189015

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>