[FIX] website_sale_stock : Remove import website_sale from variant_mixin file
Current behavior before PR:
While working on this https://github.com/odoo/enterprise/pull/74813/commits/337ea18baebdf4c896bc3e503448207fd9095991 we were overriding
a method in VariantMixin but it was shadowed and not executed this was
happening because of the order JS is loading the files when importing
website_sale.

Desired behavior after PR is merged:
We are removing the import of website_sale from variant_mixin.js
and moving the overridden method to website_sale.js to avoid having
this problem in the future.

closes odoo/odoo#190697

X-original-commit: 03cc05fd82e057d80dcaf0d781f6f85fab736883
Related: odoo/enterprise#75658
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>