[FIX] website_sale: show desired amount of alternative products
Versions
--------
- saas-17.4+

Steps
-----
1. Have a published product 'A' with 16+ variants;
2. have a published product 'B' with alternative product 'A';
3. add more alternative products whose names are greater than 'A';
4. open product 'B' in eCommerce.

Issue
-----
Only product 'A' gets displayed.

Cause
-----
The dynamic filter for alternative products applies the default limit of
'16' to the variant count.

This was not an issue in earlier versions, as alternative products
displayed variants instead if templates.

Solution
--------
In the `_prepare_values` method, set the `hide_variants` context value
even if `self.filter_id` is falsy, making it accessible for dynamic
filters as well.

In the `_get_products_*` methods, check this context value. If present,
limit each product template to a single variant before calling `search`.

opw-4318624

closes odoo/odoo#192421

X-original-commit: 22499394fedd26cce5ce09cca0759edb9785ff83
Signed-off-by: Levi Siuzdak <sile@odoo.com>