[FIX] stock: reload reception report with values
Steps to reproduce:
- Enable 'Reception Report' in Inventory Configuration
- Make an outgoing shipment for a storable product
- Make an incoming shipment for that same product
- Open the allocation report
- Reload the page

Issue:
The context is lost when reloading the page, meaning that we lose the
the `default_picking_ids`/`default_production_ids` in the context,
making it unable to open the report.

To avoid this, we add the key and values to the router (and thus the
URL) so it can be properly picked up when reloading the report.

opw-4321072

closes odoo/odoo#189560

Signed-off-by: Steve Van Essche <svs@odoo.com>