[FIX] account_peppol: Make text translation friendly
In previous commit [1] we added text at the bottom of email sent when sending
invoices, this doesn't work well for translation to have <a> tags in the middle
of sentences, in particular in Dutch and German. We move the cliquable URL
at the end of the sentence in an icon.

[1]: https://github.com/odoo/odoo/commit/42cc1ea713aa977794231897b2f04758e30d8e19

task-no

closes odoo/odoo#189283

X-original-commit: 173110af4535d44b1b7db84b7d5452e12fc7bfbd
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>