[FIX] web_editor: adapts forced size in image_shape route
In commit [1], we introduced a "data-forced-size" attribute on "image
shapes" SVGs to adjust SVGs height to ensure the image fits properly
within it (e.g. for "devices" shapes).

However, the "image_shape" route was not updated to match this change,
preventing the use of predefined building blocks containing an image
shape with a "forced-size" attribute.

This commit updates the "image_shape" route to support the
"data-forced-size" attribute.

[1]: https://github.com/odoo/odoo/commit/63cf8a693a5262600d465d70f0383229464d71e7

task-4094393

closes odoo/odoo#189156

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>