[FIX] auth_oauth: exclude routes from readonly transactions
Follow-up of #186319 and #186786

These routes are auth=none but aren't technically read-only, as they can
create the user record and need to record the login event anyway.

The signin one even as a broad catch for Exception, which would prevent
the readonly->read-write fallback to work, leading to a failed login.

closes odoo/odoo#189166

Signed-off-by: Olivier Dony (odo) <odo@odoo.com>