[FIX] spreadsheet: Remove data validation in production
Following #184846, the data validation is no longer necessary in a
production context.Moreover, it currently limits some capacities for the
user to user the history of the spreadsheets. Eg:

- Create a new field with studio
- Add a datasouce in a spreadsheet which relies on that field
- make a bunch of revision
- Delete the field you created
- Go back to the spreadsheet
- Open the version history and try to "make a copy" of some recent revision

=> it will throw as you try to create a spreadsheet with invalid fields

However, the data validation still make sense to alert developpers to
adapts dashboards when they change the schema of some tables.
This revision deactivates the data validation on production which
ensures that the last point is still respected while developping
(will be detected by runbot tests).

Task-4363803

closes odoo/odoo#188792

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>