[FIX] purchase: set_vendor_price_po_company
The issue:
When a vendor has multiple prices for two companies (A and B), and the purchase order's company is set to B (while the user's main company is A), the price incorrectly defaults to the main company A's price.

How to reproduce the issue:
1) Create a new product with 2 vendor pricelist records that have different "Unit Prices" under the "Purchase" tab of the product -> each pricelist record should belong to a different company
2) Create an RFQ for this product with both of these companies active
3) Change the "Company" on the PO under the "Other Information" tab to the company that is NOT shown in the top right-hand corner, but is still selected
4) Add the product to the RFQ order line
- The "Unit Price" will be pulled from the main company shown in the top-right hand corner, thereby ignoring the "Company" selected for the PO.

After this commit, the price from the purchase order's company (B) is prioritized over the main company (A).

opw-4347134

closes odoo/odoo#192918

Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>