[FIX] im_livechat: chatbot answer containing ampersamd
Before this commit, if chatbot had ampersamd character in
question answers, this would lead to following crash:

```
Cannot read properties of undefined (reading 'id')
```

Steps to reproduce:

- configure chatbot with following step scripts:
```
 Message | step type | answers                   | only if
---------+-----------+---------------------------+----------------
 step    | question  | "test & step 2", "step 3" |
 step 2  | text      |                           | "test & step 2"
 step 3  | text      |                           | "step 3"
```

- open conversation with chatbot
- select "test & step 2"
=> crash

This happens because when selecting an answer, chatbot detects the
selected answer by comparing user message content with all possible
answer text content. The message body of user is escaped, whereas
chatbot answers are not. Thus `test & step 2` is not found and
therefore the crash results in `answer.id` where `answer` is
`undefined`.

This commit fixes the issue by escaping the value of answer in the
detection of selected answer by user, so that escaped content is
compared with escaped content too.

opw-4369966

closes odoo/odoo#189313

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>