[FIX] calendar: res_users_settings error in internal
Before this commit, an access error was being triggered when opening
the user form of different internal users than self. This happened
because of the `compute_sudo` flag in the calendar_default_privacy
computation function.

The sudo flag was introduced for allowing the read of the ResUsersSettings
configuration during the computation of calendar_default_privacy. Turns out
that this flag is not needed, since the sudo() flags in the inverse function
is already sufficient. Thus, we're removing it from the computation.

closes odoo/odoo#189426

Issue-from: odoo/odoo#183887
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>