[FIX] account: paid payment state for manual payments
When using Invoicing without Accounting, the payment state never became
`paid` if the payment was made manually, even after reconciling it.

Particulary annoying with the use of batch payments because the batch
payments always have a journal entry in Invoicing only, even when the
purpose of it is to send in a batch.

runbot-104545
runbot-105503
runbot-105504

closes odoo/odoo#185111

Related: odoo/enterprise#72708
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>