[FIX] point_of_sale: raise access denied for user role
We encounter an error when trying to open any POS category from the 'Dashboard,'
if the Administrator is assigned the role of 'User' for 'Point of Sale.'

Steps to reproduce:
- Install 'point_of_sale' module
- Open POS
- Now go to Users and Change the right from 'Admin' -> 'User' in 'Point of Sale'
- Now go to 'Dashboard' and try to open any category

Traceback: ParseError
while parsing /home/odoo/src/odoo/saas-17.4/addons/product/data/product_demo.xml:5, somewhere inside
<record id="base.group_user" model="res.groups">
            <field name="implied_ids" eval="[(4, ref('product.group_product_variant'))]"/>
        </record>

This error occurs when the POS application is opened, as the onWillStart method
is triggered. However, if the user's permissions are changed, the method does
not update accordingly, leading to an error.

sentry-5717539295

closes odoo/odoo#191206

X-original-commit: 434066ed72b2b1466149ba9a4f62b4d3a0c658d1
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>