[FIX] web_editor: transform button should be inactive after resetting
Behaviour before commit:

When transforming an image, if we click on transform button again,
it resets the transformation of that image but transform button
remains active.

Behaviour after commit:

Now, if transformation container is open and user clicks on transform
button then image gets reset and transform button gets inactive as
there is no transformation.

task-4251662

closes odoo/odoo#190556

X-original-commit: a53c923583f126bbcaab38acb2ffe63b898c7701
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>