[FIX] mail: livechat action hover style matches header color
Chat window action mouse hover effect was hardcoded for white and
dark theme.

In livechat, the header color can be customised. With the default
colors (purple background and white text), the hover effect had
white background, which made the action not visible.

This commit fixes the issue by using a slight color invert when
mouse hovering the action of a chat window, so that the color is
slightly different from header color while keeping icon color
distinguishable.

closes odoo/odoo#189629

Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>