[FIX] l10n_be_pos_sale: activate default invoice only with intracom taxes:
Previously, settling a sale.order from the POS of a Belgian company
would auto check the invoice option before paying.

Steps to reproduce:
-------------------
* Install `l10n_be_pos_cert`
* Switch to the Beligian company
* Crete a quotation in **Sale** app
* Add any partner and any product, remove all taxes for the product line
* Save
* Open pos shop
* Settle the order
* Go to the payment screen
> Observation: The invoice button is already selected.

Why the fix:
------------
This invoice option was auto selected since this commit: c760fbb

The following commit 3a5e222
updates the previous one as the invoice enforcement is only needed when we have
intracom taxes on the SO. Thus when no intracom taxes apply we can still
decide wether or not to invoice. We fallback on the default behavior which
is to have the invoice button un-selected.

opw-4334095

closes odoo/odoo#190945

X-original-commit: ae305d1d10c1efabea804748725cbeb48ae6e3ad
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>