[FIX] web: list: reset column widths after resize
In a list view, first manually resize a column. Doing so, the
table can either overflow (if the column has been extended), or be
smaller that the available space. Then, resize the page. Before
this commit, the table kept is previous width, i.e. it could not
benefit from the potential available space that came from resizing
the window. With this commit, we force the whole widths computation
to be reset after a window resize.

Followup of https://github.com/odoo/odoo/pull/170511

opw~4318312
task~4221195

closes odoo/odoo#190036

X-original-commit: 222a2e054fa8a3e599f911ddceca510f59dc2d21
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>