[FIX] website_sale: pass compute price args as kwargs
Versions
--------
- saas-17.4+

Steps
-----
1. Start a fresh database with only `website_sale` installed;
2. configure Website to show strikethrough price;
3. configure Website to let user decide on "Add to Cart";
4. create a pricelist with a discount based on another pricelist;
5. have discount apply to all product categories;
6. make pricelist selectable;
7. go to /shop;
8. add a product to cart.

Issue
-----
> AssertionError:
> Invalid value [...] in domain term ('date_end', '>=', res.currency(124,))

Cause
-----
Commit 4dffe0d changed how versions 17.4+ display strikethrough prices
for discount rules.
In doing so, it added a call to `_compute_price_before_discount`, passing
the parameters positionally.

The problem is that not all helper functions used in the process accept
these arguments in the same order, specifically the `_compute_price_rule`
method of `product.pricelist` which takes:
  currency -> uom -> date

vs. `_compute_base_price` of `product.pricelist.item` which takes:
  uom -> date -> currency.

Solution
--------
Name the arguments, so that they get passed `kwargs` instead of `args`.

opw-4366956

closes odoo/odoo#190386

X-original-commit: 59e86e31665e90e4f9fab727d3dc87596ebedbb5
Signed-off-by: Levi Siuzdak <sile@odoo.com>