[FIX] hr_recruitment: candidates in right company when receiving email
When an application is received through email, an `hr.candidate` is
created without `company_id`. The default for this field is
self.env.company. When processing an email this will be the company of
OdooBot. `hr.applicant` on the other hand, will compute its default
company based on the linked `hr.job`.

Because of this, it's possible that these two records are created in
different companies leading to:

File ".../addons/hr_recruitment/models/hr_applicant.py", line 351, in create
  raise ValidationError(_("You cannot create an applicant in a different company than the candidate"))

Fix it by creating the candidate in the same company the job is in.

opw-4384806

closes odoo/odoo#190002

Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>