[IMP] web: use overlay service for dropzones
The current implementation of the dropzone hook relies on the component
registry, which is anchored to an element within the modal registry anchor.
As a result, when a dropzone is placed inside a modal, it is rendered
beneath the modal, making it invisible to users.

This commit will fix that issue by ensuring that the dropzone hook uses
the overlay service to place the element. The dropzone will then appear
above the modal and people will be able to drop files.

task-4273520

Part-of: odoo/odoo#186617
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>