[FIX] sale_timesheet: fix traceback when opening invoices from tasks in portal
Currently a traceback occurs when the user opens an invoices from
the tasks in the portal.

To reproduce this issue:

1) Install sale, timesheet
2) Open a sale order having project or create an SO with product type as service
3) Create multiple invoices for that SO
4) Now open the related tasks of that SO having invoices from portal
5) Click on `invoices` from sid pannel

Error:- "AttributeError: 'list' object has no attribute 'ids'"

This error is occurring from the below line
https://github.com/odoo/odoo/blob/88c0f9d953d077510961243080dcab2cc1920c27/addons/sale_timesheet/controllers/portal.py#L43-L50

Clearly, we  get invoices as a list of recordsets in values from the below line
https://github.com/odoo/odoo/blob/88c0f9d953d077510961243080dcab2cc1920c27/addons/account/controllers/portal.py#L126-L133

Initially, we get the recordsets of invoices in `values['invoices']`.
But because of the recent changes from the below line we getting
the list of recordset values. In which we can access the value of
invoice id.

We can resolve this issue by looping in invoices and access the invoice id.
similarly like this
https://github.com/odoo/odoo/blob/48629b94dac45dad0766fbaeef0f71934a5adcc4/addons/account/controllers/portal.py#L90

sentry-6117277208

closes odoo/odoo#190062

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>