[FIX] hr_timesheet: do not use non timesheetable project as default
This commit's purpose is to prevent the method _get_favorite_project_id
to return a project that is not timesheetable. Currently, we do not
exclude those project from the search, and in some cases we use a non
timesheetable project as default value for timesheets in the timesheet
header.

task - 4268687

closes odoo/odoo#188025

Related: odoo/enterprise#74264
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>