[FIX] account_peppol: Fix deduction of the demo edi mode
Previously, when no EDI proxy user existed, in some case when the handle_demo
ran, it was erronously deducing that the EDI mode was not demo, while it was.
This commit fixes and simplifies the deduction of the EDI mode:
We check the company in the env, if there is a related Peppol proxy user we
take his mode, else we fallback on the dedicated system parameter.

task-no

closes odoo/odoo#190752

X-original-commit: 33fd123dc49e969e45da56caa341224ca7e981b9
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>