[FIX] pos_viva_wallet: add webhook fallback
The Viva Wallet integration relies on Odoo
being called by Viva Wallet via a webhook.
There have been many issues of this webhook
never being called, and therefore payments
getting stuck in Odoo.

In this commit, we add a polling fallback.
Every 5 seconds we query the Viva Wallet
session, and if we get a response (success/failure)
then we update the payment state in the PoS.

closes odoo/odoo#191441

X-original-commit: 70a9867caa308986ade014aba16b3e02ba63ddcb
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>