[FIX] point_of_sale: correctly sync pos payments
Before this commit, if you added a payment line to an order and synced
it to the server, then added some items and removed the payment line to
add a new one, the payment line wasn't removed from the backend. This
caused the order to have different payment lines from what the user
validated. This commit ensures that payment lines are correctly synced
and updated in the backend.

opw-4257663

closes odoo/odoo#190743

X-original-commit: 4b84d4b6f5c838230800e92281fafe5f973d727b
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>