[FIX] hr_expense: correct mapping of expense alias domain in settings
Steps to reproduce:
- Install hr_expense
- Create 2 alias domains
- Go to "Settings / Technical / Email / Aliases"
- Check the values for Expense model
- Go to Expenses settings
- Change the prefix and the domain of the alias
- Save
- Go to "Settings / Technical / Email / Aliases"
- Check the values for Expense model again

Issue:
The prefix has correctly been modified but not the domain.
The domain from Expenses settings and the domain of the Expense alias
are different.

Cause:
The domain field in Expenses settings is in fact the Email Domain
(alias_domain_id) of the company, which makes no sense to combine
the prefix and the domain from 2 different sources.

Solution:
Add a non-stored computed field to map the domain displayed in Expenses
settings with the domain of the Expense alias.
Not great in stable, but there is no other way if we want to keep the
option.

opw-4293936

closes odoo/odoo#192691

X-original-commit: 4a8bdf4a449e6f47833d505a40b41f88526f733b
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>