[FIX] sale: disabled price edition
The new `technical_price_unit` logic was introduced to reduce/prevent
unexpected price recomputations when the price is manually updated
on sale order lines.

Nevertheless, if the price_unit field is made readonly (through studio
or any custom view ...), the price would stop being sent to the server
by the client, resulting in zero-priced amounts because the presence
of the technical price discarded the price recomputation, leading to
the default 0 value.

This commit makes sure that in this situation, the technical price
sent by the client is dropped, ensuring that the price is correctly
computed.

opw-4319211

closes odoo/odoo#190273

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>