[IMP] mail: avoid MemoryError when filtering `template_category`
Before this commit, databases with thousands of mail templates could trigger
a MemoryError when opening the mail template list. This issue occurred because
the default filter on base templates required processing all records batch by
batch in Python.

With this commit, the computation is rewritten to use domain / SQL, eliminating
the Python overhead.

closes odoo/odoo#190326

Signed-off-by: Thibault Francois <tfr@odoo.com>