[IMP] mail: disable email signature in customer email responses
When an external user responds to an email sent via Odoo, Odoo may send
a mail notification with the user's message and automatically add a
signature on their behalf.

Since most external users may have already configured an email signature
in their mail client, the added signature can be redundant. To avoid
this, we will no longer add a signature on behalf of external users.

task-4273520

closes odoo/odoo#185849

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>