[FIX] html_editor: outdent list item with empty nodes
Before this commit, outdenting a list item with empty (or invisible)
nodes would lead to a traceback. This was due to the fact that the
outdentTopLevelLI method was not properly handling the case where the
list item had only empty/invisible nodes, resulting in calling preserve
selection's restore with a node not connected to the document.

task-4398327

closes odoo/odoo#190484

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>