[FIX] web: configuration layout: missing left padding for inputs
Before this commit, inputs were stuck to the right edge of the screen
on mobile. Bootstrap adds negative margins on row that needs to be
compensate by the paddings.
It's why we shouldn't override gutters or paddings.

Paddings are now standardized and the same as those in the generic form view.

Task-4377734

closes odoo/odoo#190030

Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>