[FIX] point_of_sale: add rate to tax_base_line_values
Steps to reproduce the issue:

1. Activate Mexican Localization
2. In Mexican Company, add USD as Currency and give it a Rate
3. Create a Sales Journal and a Pricelist with USD as main Currency
4. _Point of Sale > Configuration > Settings_: Select a Shop and set the USD Journal and Pricelist
5. Open a Session using the USD Shop and complete some Orders, without invoicing them
6. _Point of Sale > Orders > Orders_, select the created Orders and, in the Actions, Create Global Invoice
7. The document should be accepted, but TipoCambio has a rate of 1.000000.

Explanation:

During the creation of the Global Invoice, the rate is retrieved during `pos.order._prepare_tax_base_line_values` and, since no value is given through POS, the default value is used.
https://github.com/odoo/odoo/blob/8fb7e5fd304697aebcce085602a5f3a1ecaf757a/addons/account/models/account_tax.py#L1242
https://github.com/odoo/odoo/blob/8fb7e5fd304697aebcce085602a5f3a1ecaf757a/addons/account/models/account_tax.py#L1212-L1213
https://github.com/odoo/odoo/blob/8fb7e5fd304697aebcce085602a5f3a1ecaf757a/addons/account/models/account_tax.py#L1188-L1193

Fix reasoning:

`pos.order.currency_rate` is the rate of the Currency at the time the order was created. It is the perfect value to use for `TipoCambio`. We must also adapt workflows that were designed with the absence of `rate` in `tax_base_line_values`.
Test in Enterprise

opw-4302180

closes odoo/odoo#189700

Related: odoo/enterprise#75185
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>