[FIX] fleet: allow overriding the behavior on contract easily
Since there are two mechanism that changes the status of a contract (a
cron and the write of a date), we have to override 2 different places
when we want to add a behavior on closing or opening of contracts.

As there are already functions (action_expire, ...) that exist and are
use in the write part, we also use them in the cron, so we'll have only
one place where we need to override it.

closes odoo/odoo#190539

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>