[FIX] point_of_sale: not show discount on price manually set
Currently, when using pricelists, if you change the price of a product inside
the session, it will show a discount value on the receipt.

Steps to reproduce:
-------------------
* Activate advanced pricelist
* Modify the shop's default pricelist
  * Discount policy: Show public price & discounts
  * Add a rule: 10% discount on all products with min qty of 10
* Open shop session
* Add any product qty 1
* Manually change price to a higher amount
* Pay the order
> Observation: On the receipt we can see a negative discount value.

Why the fix:
------------
Changing the price of a product through the price button in the session
should never be considered as a discount. Lines for which the product price
was manually set should not be counted toward overall discount (positive or
negative).

opw-4366651

closes odoo/odoo#191384

X-original-commit: 3e77f003ab5b72cb76c9d1e1846db1464a5fc5da
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>