[FIX] l10n_ee: rename xmlid in upgrade script
In odoo/odoo#180856, a new version of the Estonian tax report is
added with the xmlid tax_report, in order to not overwrite the
tax report for users in stable who didn't update their modules.

When upgrading the module, to avoid dropping and recreating the
report, we add this upgrade script to rename the xmlid of the
tax report.

no-task

closes odoo/odoo#190872

X-original-commit: c96d2b1d1ee917b1c665842010c56c23df91ccd3
Related: odoo/enterprise#75741
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Maira Salazar <srma@odoo.com>