[FIX] website_sale_collect: hide 'Send to Shipper' button for 'in_store' delivery type
This issue occurs when a delivery order is created with the 'Carrier' set to
'Pick up in Store.' After saving and validating the order, click the
'Send to Shipper' button.

Steps to reproduce:
- Install 'website_sale_collect' module
- Create a new 'Delivery Order' and in 'Additional Info' select 'Carrier' as
  'Pick up in Store' > Save and Validate
- Click on the 'Send to Shipper' button

Traceback: TypeError
'NoneType' object is not subscriptable

This commit resolves the issue by adding an XPath to the view to hide the
'Send to Shipper' button for orders with the delivery type set to 'in_store'.

sentry-6040319897

closes odoo/odoo#190772

Signed-off-by: Valeriya Chuprina (vchu) <vchu@odoo.com>