[FIX] web_editor: prevents dropping building blocks into an image field
This commit prevents dropping a building block into an image field.

There is no known case in V16.0 where the issue can be reproduced, but
it must still be fixed in V16.0 because it should not be possible
anyway.

For example, in V16.0, if an image field had a div as a parent, adding
the "row" class to the parent of that div would allow dropping blocks
into the image field. This was because the selector defining where an
image could be dropped is ".row > div:not(.o_grid_item_image) > img" in
V16.0.

However, in V17.0, since this commit [1], the selector was changed, and
the ".row >" part was removed, which caused the bug to appear.

Steps to reproduce from V17.0:

- Go to the "/shop" page.
- Click on one of the products to go to its product page.
- Enter edit mode.
- Bug: It is possible to drop any "inner content" block into the
"product" image field.

[1]: https://github.com/odoo/odoo/commit/507b80a12574ce19bcb925d88868fb377b1c4b06

opw-4273436

closes odoo/odoo#191480

X-original-commit: d58dddc9574fa99640c9b8b58abbe03fc715c21b
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Signed-off-by: Benjamin Vray (bvr) <bvr@odoo.com>