[FIX] web: RelationalModel: pass context to search_count requests
Have a list or kanban views with more than 10k records such that
the count limit is reached, and the pager displays something like
"1-80/10000+". Click on "10000+" to compute the real count. Before
this commit, the context wasn't given to that call, which could
obviously return a wrong result, for instance if the context has
the "active_test" key set to false.

task~4331708

closes odoo/odoo#190807

X-original-commit: a93bbd83a7036ceaba7443ee23e90ed84d43a731
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>