[FIX] product_expiry: update exipiration_date when changing lot_id
Steps to reproduce the bug:
- Enable “Subcontracting” and “Expiration Date” in MRP settings.
- Create a storable product “P1” with the following configurations:
    - tracked by: LOT
    - Expiration date: True
    - BoM:
        - BoM type: subcontracting
        - subcontractor: Azure interior

- Create a lot “L1” with an expiration date set to 01-01-2026.
    - Expiration date: "01-01-2026"

- Configure the operation type::
    - Receipt:
        - use existing SN: True

- Create a receipt:
    - Receive from: Azure interior
    - Product: one unit of P1
    - Mark it as todo
    - Open the detailed operation:
        - Set the “L1”
        - Record production
    - Reopen the detailed operation

Problem:
The expiration date in the stock.move.line is set to today’s date
instead of the lot’s expiration date.

Cause:
When marking the receipt as “To Do,” the SML is created. Since no
"lot_id" is initially set, today’s date is assigned, thanks to the
"_compute_expiration_date" method:

https://github.com/odoo/odoo/blob/9f1d555ac2d9b6e9d3d0d723d4f042eb44cf30b7/addons/product_expiry/models/stock_move_line.py#L37-L40

But When the lot_id is later set, the compute method is triggered again.
However, since the SML already contains a date, the expiration date is
not updated to match the lot’s expiration date.

opw-4394662

closes odoo/odoo#191399

X-original-commit: 787c71f71e865f035e078980079aca5ff0a11ecb
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>