[FIX] web: Field: apply decorations on select
This is a regression since 48ef812a. Before, the field's views were rendered
directly, getting all the decoration classes on them. Since 48ef812a, they are
rendered as a child of a parent `Field` component (a `div`), which gets the
decoration classes instead [^1].

The issue was partially fixed by 2f99f7dc, but only for <input/> and <textarea/>
elements. The same fix is now applied to the <select/> element.

[^1]: https://github.com/odoo/odoo/blob/60c7b65f/addons/web/static/src/views/fields/field.xml#L4-L8

closes odoo/odoo#190788

X-original-commit: 219328de8291612db31d2c27050d3b50797b538f
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>