[FIX] html_editor: move nodes when collaborator on same node
Before this commit:

It was not possible to use the move node handler when the collaborator is on
the same node, because the collaboration avatar is hidden using opacity 0 which
still allowed the avatar to be selectable.

After this commit:

Hide the collaboration avatar using visibility hidden, ensuring it is
not selectable.

task-4348275

closes odoo/odoo#190795

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>