[FIX] mrp: prevent re-count global lead days if not 1-step manufacture
**Current behavior:**
The global lead days system parameter may be counted in the lead
days breakdown in the replenishment report when:
A) product has manufacture route and,
B) WH manufacture steps is not 1-step

**Expected behavior:**
The parameter is counted once.

**Steps to reproduce:**
1. Create a product with a simple BoM and the manufacture route

2. Enable 2 or 3 step manufacturing in WH

3. Create an out move for the product

4. Open the replenishment report in inventory

5. Click the (i) in the orderpoint line for the product -> see
the global vis. days are counted 2x

**Cause of the issue:**
Like the issue solved by 38d8e77
In some instances the `_get_lead_days()` must be called a second
time to capture some additional lead time (in this case because
we have > 1-step manufacturing).

**Fix:**
Use the context added in the referenced similar commit to ignore
the global days in the second call.

opw-4410790

closes odoo/odoo#190808

Signed-off-by: William Henrotin (whe) <whe@odoo.com>