[FIX] stock_landed_costs: exclude uncreated landed cost lines
A stock valuation issue occurs when a landed cost is present in a bill
lines of a purchase order (without the `stock.landed_cost` created)
and the receipt note is validated after confirming the bill.

Steps to reproduce:
- Create a product with automated inventory valuation and AVCO.
- On the purchase tab, set the control policy to “on ordered quantity”.
- Create a purchase order with this product:
  * Quantity: 1
  * Unit cost: $10
- Validate the PO.
- Create a bill and add a landed cost line:
  * Quantity: 1
  * Cost: $25
- Validate the bill.
- Validate the receipt note.
- Check the valuation:
  * The unit value will be incorrect, calculated as $17.5
    ((10 + 25) / 2).

This issue is caused by commit
https://github.com/odoo/odoo/commit/5d067851248aaadf4f447f144a22553f9b188b71.

The solution is to include only landed cost lines from bills
where the `stock.landed.cost` is created and validated.

opw-4384070

closes odoo/odoo#191085

X-original-commit: 859e4fdd124043a30b6da2b70e5b7b7e13125b7c
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>