[FIX] point_of_sale: keep paid order in IndexedDB
Before this commit, if you opened the PoS, made it offline, validated
an order, stayed on the receipt screen, then made the PoS online and
refreshed the page, the paid order would become a draft order. This
caused the loss of tracking for the paid receipt-printed order. This
issue occurred because using readAll prevented the effect from
triggering on data modification.

opw-4371425

closes odoo/odoo#190871

X-original-commit: 05edc2ab717f549b7e6d95c35a2fba40a463e719
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>