[IMP] event: clarify the Timezone field name and tooltips
Issue:

The current descriptions and naming of the `Date` fields (`date_begin`,
`date_end`) and the `Timezone` field (`date_tz`) can be confusing in the
form view of the event model. Specifically:

The timezone used for converting the form dates to the database is
determined by the context and not by the `Timezone` field present on the
form. However, this `Timezone` field determines the timezone used
for displaying the event's date/time on the website. This change tries
to clarify the situation.

opw-4323142

closes odoo/odoo#190867

Related: odoo/documentation#11596
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>