[FIX] web: Update correctly reference char field
Let say we have two record with a reference field with char type.
Record X with reference A
Record Y with reference B

If we open the record x form view directly from the list view the fetch
will works as wanted (reference A)
But if we use the pager to navigate to record Y, the reference field
will not be correct, it will dispay the old one (reference A instead of B)

This is because in the `ReferenceField` setup method we use an
`useRecordObserver` to track record changes and update reference data.

But before this commit, we passed the wrong `props` to `_fetchReferenceCharData`

`useRecordObserver` has a second argument, which is the new `props`,
so we can use it instead of `this.props` still referring to the previous
record

This can be reproduced on `External Identifiers` view (accessible with debug)

opw-4349296

closes odoo/odoo#190970

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>