[FIX] mail: do not crash when having recipients with inactive lang
Currently if a partner to notify has a 'lang' set to an inactive lang
notification process crashes. Indeed it tries to use the inactive lang
to translate layout bits (e.g. buttons, layout, ...) and current ORM
prefers to crash rather than skipping of fallbacking.

We therefore choose to check for the lang to be active before considering
it in notification process.

Part-of: odoo/odoo#190891
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>