[PERF] sale_expense: add missing index on `sale_order_id`
Description:
`sale.order._compute_expense_count` does a `_read_group` with a domain
only based on `sale_order_id`, which is an unindexed FKey, leading
to a guaranteed `Seq.Scan`. This `_compute_expense_count` is used to
compute the count of related expenses for a `sale.order`, shown as a
smart button on the form view, which is a frequent operation.

Reference:
task-3977983

closes odoo/odoo#191004

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>