[FIX] web_editor: remove listeners on destroy
Steps to reproduce:
1. Open the website builder.
2. Add the "Items" block.
3. Double click an image and change it.
4. Via the link popover on that image, unlink.
=> Traceback.

When destroying the related `ReplaceMedia` snippet option after the
image change, some listeners were not cleared. As a result, when
focusing on a new instance of `ReplaceMedia`, the previous listeners
remained active with stale context (old `$target`), which no longer had
a parent since it was removed during the `img` change.

opw-4412300

closes odoo/odoo#191142

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>