[FIX] mail: make suggestions feel snappier
Before this commit, suggestions were fetched sequentially, meaning
that adding characters to the search would only take effect after the
current RPC call finished. The intent was to provide initial results
quickly, as the previous RPC had already started.

However, the first suggestion fetch is often based on very few
characters, which can take a long time to process. This creates the
impression that no item matches the search terms.

With more characters, the search becomes much faster. This commit
adapts the behavior to cancel the previous fetch instead of waiting
for it to finish. It also debounces the fetch to ensure it starts with
as many characters as possible.

closes odoo/odoo#191119

Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>