[FIX] html_editor: remove link in uneditable zone
Empty links are removed by default on clean for save and when the link
popover is closed. This commit adds a check to prevent removing empty
links in uneditable zones.

In particular, this is the case in the download box produced by the
/file command, where the link is contenteditable=true so that it's label
can be edited, while its parent is not editable. This commit prevents
thus the removal of such link, which would make the download box
useless.

task-3522395

Part-of: odoo/odoo#189203
Related: odoo/enterprise#75045
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>