[FIX] website: stop product carousel auto sliding
In 17.2 and prior, the product carousel did not auto slide. However,
after upgrading to Bootstrap 3.5.5 in 17.4, commit [1] changed the
behavior of the carousel and made it auto slide after user interactions,
using `data-bs-ride="true"` attribute.

This issue is known, and fixes were provided for some carousels in [2]
and [3].

Here we apply the same fix for the products carousel.

[1]: https://github.com/odoo/odoo/commit/7da319c81f0a0d8a29c68567f563c501f7dfbe65
[2]: https://github.com/odoo/odoo/commit/9eff9ae1904e0583f5dc60c6d925c52b48b208ec
[3]: https://github.com/odoo/odoo/commit/1db386b088699456f42424e9dd7dd7f74651fcdc

opw-4391294
opw-4404683

closes odoo/odoo#191277

X-original-commit: https://github.com/odoo/odoo/commit/d7d59ab45470d930102cd3e247279d952a2a3c9d
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Signed-off-by: Hadi El Yakhni (hael) <hael@odoo.com>