[FIX] website_mail: `mail` translations available in frontend
Portal chatter now uses discuss, thus it relies on translation of
Discuss in `mail` module.

Frontend translations need to be available, they aren't by default
unless the module is prefixed by `website`. This isn't the case for
`mail`, and the module `website_mail` does not share the
translations.

This commit fixes the issue by enabling `mail` in the translations
of available translations in the frontend

Task-4423556

closes odoo/odoo#191534

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>