[FIX] sale,website_sale: pricelist valid dates not applied
If you create a pricelist rule with a discount that has a valid date range,
that discount is only applied if the SO was created in that range.
Even if it is confirmed within the valid date range.

Fix:
For website_sale orders we consider the date to be the current time
when computing the price.

opw-4375643

closes odoo/odoo#191538

Signed-off-by: Lina Eweis (liew) <liew@odoo.com>