[FIX] website: allows editing dynamic snippet link
Steps to reproduce:

- Install the e-commerce module.
- Go to the homepage.
- Enter edit mode.
- Drag and drop a "Products" block onto the page.
- Click on the "See all" link.
- Edit the link (e.g., change "/shop" to "/contactus").
- Save the page.
- Click on the "See all" link.
- Bug: You are redirected to the "/shop" page instead of the
"/contactus" page as expected.

The bug occurs since commit [1], where the "See all" link was added to
"Dynamic" snippets. Each time a "Dynamic" snippet is rendered during a
page load, the href of the "See all" link is reset to the main module
page. While this behavior is useful when the snippet is first rendered
after being dropped onto the page, it is problematic afterward. Once the
user has edited the link with a different href, it should not be reset
on every page load.

[1]: https://github.com/odoo/odoo/commit/bbcff746075485be61807d08ad8232dbad347713

opw-4418048

closes odoo/odoo#191689

Signed-off-by: Outagant Mehdi (mou) <mou@odoo.com>