[FIX] l10n_es_edi_tbai: only use regime key 52 for simplified invoice
Steps to reproduce:
- Install l10n_es_edi_tbai
- Switch to a Spanish company (e.g. ES Company)
- Create an invoice with "Simplified Invoice Partner (ES)" as customer
- Confirm the invoice
- Process the invoice with "TicketBAI (ES)" service

Issue:
The generated XML has 2 regime keys (02 and 52) for
<ClaveRegimenIvaOpTrascendencia> that are not compatible with each
other.
From the documentation, regime key "52" (for simplified invoices)
can only be associated with regime key "51", which is not supported
for the moment.

Solution:
If the "Simplified Invoice Partner" is used, only the regime key "52"
is used.

opw-4332052
opw-4355424

closes odoo/odoo#192825

X-original-commit: 0c0ff0641ee47745ff83b5d5ce1d15a3d208fc6c
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>