[FIX] im_livechat: correct author name display for visitor messages in demo data
Before this commit:

Messages posted by visitors in demo data incorrectly displayed the author name
as 'Odoobot' instead of 'Visitor'. This happened because the messages created
for visitors in the demo data do not have the author_id field explicitly set to
False, which causes the _message_compute_author() method to compute the author.

After this commit:

The author name for messages posted by visitors in demo data is correctly
displayed as 'Visitor'.

Task-4420677

closes odoo/odoo#192535

X-original-commit: 900846ff73bbe61746c91c6fa46a04a2bc25414d
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>