[FIX] project: wrong account company in quick create of project form
Before this commit, when creating a new analytic account in any plan from the project form, the default company of the account was set to the current company of the user.
The problem that arises is when the project has no company, and the account is set to a company. Which creates inconsistency between the project and the account.

After this commit, we set the company of the created account to the company of the project by default.

task-4438410
version-18.0

closes odoo/odoo#192021

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>