[FIX] l10n_ch: non 256 size barcode has cross centered
Scenario:

- install l10n_ch
- go to /report/barcode/?barcode_type=QR&value=&width=150&height=150&mask=ch_cross

=> the ch_cross mask (swiss cross that should be in the center) is not
in the center

Issue:

Reportlab QrCodeWidget has a fixed size to 32mm (with mm that is 72(ppi)
/ 25.4(1mm to inch) so ~2.83 pixels) and to get the widget to the full
size, a zoom is applied on the Drawing content (so on the QrCodeWidget).

Here are the zoom that are applied for several Drawing size:

- for 100 pixels: 100 / (32 * mm) => 1.012…
- for 256 pixels: 256 / (32 * mm) => 2.822…
- for 400 pixels: 400 / (32 * mm) => 4.409…

But in the code, the zoom is just hardcoded to mm (~2.83) which
corresponds to a barcode size of mm * 32 * mm => 257.127 pixels which is
kind of good enough for 256 pixels, but wrong for most other size.

In base code we only use 256 x 256 pixels so this is not much of an
issue unless someone want to use it for something else.

Fix:

With this commit, we apply the Drawing zoom (on which we have the
QrCodeWidget) to the ch_cross image.

Note:

Without the fix, the added test only fail on the third assertion for
each loop iteration, the farer we are from 257x257 pixels the bigger
the error is, for example for 256x256:

Result:   (38.45140, 38.45140, 52.25725984251969, 52.25725984251969)
Expected: (38.28288, 38.28288, 52.02823111111111, 52.02823111111111)

opw-4307177

closes odoo/odoo#192315

X-original-commit: c1f0d7593275b3551d517bda16a69a52cd4c32aa
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>