[IMP] product: strengthen combo constraints
By design, nested combos aren't supported (i.e. products of type `combo` can't
be used as combo items). However, since we used a python constraint to enforce
this, it was still possible to change the type of a product to `combo` after
using it as a combo item. To prevent this, we changed the python constraint to
an SQL constraint.

closes odoo/odoo#192172

Signed-off-by: Louis Tinel (loti) <loti@odoo.com>