[FIX] mail: can dismiss 'Turn on notifications' in messaging menu
This commit adds a "X" next to "Turn on notifications" to quickly
dismiss the suggestion to enable push notifications.

Note that this keeps the push permission to "Ask", as it cannot
be changed programmatically. So if the user wants to change the
push permissions, it should be manually changed to "Allow" or
"Denied" from browser settings or should clear the local storage
content in order to display "Turn on notifications" in messaging
menu again.

Task-4446924

closes odoo/odoo#192336

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>