[FIX] account: non updated db don't use name_placeholder in views
Description of the issue this commit addresses:

Since the deployment of the new dynamic placeholder feature for move names, the
code has been deployed on stable versions starting from 18.0 but the views are
still only updated if the user manually does it meaning that moves the account
move form view might have a unwanted behavior since the lack of change in the
view while the code was updated will show the name when we don't want to and
the other way around too in some cases.

Desired behavior after this commit is merged:

This commit modifies the _get_view of the account.move model to make sure the
invisible attribute of the name is correctly set.

no task

closes odoo/odoo#192365

Signed-off-by: William André (wan) <wan@odoo.com>