[FIX] project: project grouped by stages now respects the pagination
Steps to reproduce
==================

- Open any project
- Create many project stages (80+)
- Switch to the list view
- Group by stage
- Switch to to next page
=> The pager keeps increasing

Cause of the issue
==================

`stage_id` has a group_expand: `_read_group_stage_ids`

This method is used to include empty stages in the result for the kanban
view, so that we can drag tasks to an empty stage.

Solution
========

Add a context key `project_kanban` and only add empty stages in that
case.

opw-4408061

closes odoo/odoo#193759

X-original-commit: 4a371118c2644ed2a44992887f783d6629af4232
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>