[FIX] project: fix changeDescriptionContentAndSave
When you chain several uses of changeDescriptionContentAndSave() in a
tour, you don't necessarily wait for the editor to be actually saved
before continuing.
So, it always works for the first use but not for the following ones
if the tour engine goes too "fast".
So this fix add a step in which we make sure that the form is actually
saved before continuing.

closes odoo/odoo#192490

Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>