[FIX] pos_loyalty: mitigate race condition
During tours a race condition can occur when loyalty rewards are being
calculated, the race condition happens within the following lines:

https://github.com/odoo/odoo/blob/16.0/addons/pos_loyalty/static/src/js/Loyalty.js#L648-L680

If the method is called again between the computing of
`changesPerProgram` and `await this._couponForProgram` changes to
`couponPointChanges` are committed while the intent was to drop them
(c.f. `dropPrevious`).

Runbot Error 56943

closes odoo/odoo#192877

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>