[FIX] account_peppol: Fix external calls from tests
Context:
In Peppol, we have multiple EDI mode:
- `production` that will call the real Peppol infrastructure,
- `test` that will call the sandbox Peppol,
- `demo` that makes no call at all, everything is mocked.

By default, runbots and demo data are set to `demo`.
In tests, the value is therefore also `demo`, which is what we want:
No call to external services.

Before this commit, we were doing some weird and useless override in the handling
of demo functions, which is useless: if you want to test Odoo in a particular mode, just
set it (this is what we do in Peppol tests), for all other tests, keeping the
default demo mode should be preferred.

This was a problem because tests ran by other modules with Peppol installed were trying
to make external calls.

task-no

closes odoo/odoo#192560

Signed-off-by: William André (wan) <wan@odoo.com>