[FIX] base: fix access error to currencies when in child company
Steps to reproduce (db with Accounting installed):
- create a child company
- give a user (e.g. Marc Demo) access to the child company, set
their "Accounting" rights to "Accountant" and "Administration" to None.
- log as this user and select only the child company in the company
selector.
- try to access "Accounting/Configuration/Accounting/Currencies"
-> Access Error

It is happening as in the `res.currency` model, several methods try to
access the root company, but there's only the child company in
`allowed_company_ids` from the context.

The fix here is to:
1) When accessing the root company in order the get its currency, just
access the currency of the child company, as it is the same.
2) Use sudo when accessing other fields from the root company.

opw-4352595

closes odoo/odoo#192632

Signed-off-by: William André (wan) <wan@odoo.com>