[FIX] sale_stock: return section in portal
When cusomers access portal view, they are able to see deliveries/returns
made for their sales order. Currently in the returns list, incoming
deliveries are shown, without checking if it is a real return. This will
show incoming deliveries from third party vendor that a customer should
not see

- Activate Route MTO
- In Route Buy, open the first rule and set
  - Propagation of Procurement Group: Propagate
- Create a product [TEST] with:
  - Product Type: Goods
  - Invoicing Policy: Ordered Quantities
  - [Purchase tab] Set a Vendor
  - [Inventory Tab] Routes Buy, MTO
- Create a Sales Order with [TEST]
- Confirm
- Open the Purchase order and confirm it
- Back to the SO, 2 deliveries are created
- Open portal

Issue: Under 'Returns' it is listed the delivery from the vendor to the
company.

opw-4420364

closes odoo/odoo#192729

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>