[FIX] account: Fix quick encoding with reverse charge tax
Suppose a tax 15% +100 -100.
Use the quick encoding with an amount of 100.0.
=> The base of the created invoice line is 86.96 and the created tax lines are +13.04 -26.08.
That's wrong. Both have to produce the same amount.

In 17.0, we get a base of 100 and +15 -15 as tax amounts.
In 18.0, we changed the management of reverse charge so the total excluded amount is computed
as 100 / 1.15 ~= 86.96.
It produces tax lines computed on 86.96. 86.96 * 0.15 ~= 13.04.
But at the end, the quick encoding algorithmn "fixes" the tax amount.
It expect a total of 100 and it has a missing amount of 100 - 86.96 - 13.04 + 13.04 = 13.04 to reach the objective.
So it adds 13.04 in the biggest tax line.

opw-4428842

closes odoo/odoo#192949

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>