[FIX] l10n_my_*: Enable invoicing foreign customers.
The current logic is not working when trying to invoice foreign customers.
A few fields that could be useful are hidden, and we don't add the id information
in the xml file which causes it to fail.

In practice, it is valid to invoice a foreign customer and use a generic TIN
if they don't have a malaysian one when reporting to the tax authorities.
This change will allow just that; we add the id information to the xml which
allow to correctly issue a valid invoice for a foreign customer (tested on pre-
production)

opw-4438259

X-original-commit: 6a3a2a17ec7ba8ed3c08a044b8294bba1819c275
Part-of: odoo/odoo#193162
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>