[FIX] point_of_sale: handle custom one2many fields
Before this commit, if there was a custom one2many field with the same
related field as an existing one, it would cause issues in PoS when
capturing an order. For example, if there was a custom field in
pos.order with a related field of order_id, adding products to the
order would not show them in the order summary.

opw-4462991

closes odoo/odoo#193605

X-original-commit: 82721c502e0a55b1d4851c6f35a9795924229a28
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>