[IMP] mail: discrete spinner on real-time connection lost
The "Real-time connection lost" warning is now displayed 15 seconds
after disconnect, as to not be annoying in case of micro-disconnects.

However apps such as Discuss do not work as expected while being
disconnected. 15 seconds can be long to see the issue, and even
micro-disconnects contributes to annoyances. No UI feedback gives
bad impression of overall system stability.

This commit adds a discrete spinner at the usual place of the
"Real-time connection lost". This is barely visible, and shows
immediately when the connection is lost. When apps like Discuss do
not work, this spinner helps to grasp there are some issues and
there's reconnection attempt under-the-hood.

Part of Task-4453176

closes odoo/odoo#192705

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>