[FIX] project_todo: avoid clicking in step of project_task_activities_split tour
Before this commit, the test running the `project_task_activities_split`
tour could failed because a step clicks on a task displayed in the
activity view instead of just checking if the task expected is present
on the view.

This commit removes the click event on that step to just check if we
found the node with the trigger defined in that step to be sure the task
is on the activity view. By doing that, the next step checking if
another task is also there in the activity.

Closes #192995

closes odoo/odoo#193066

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>