[FIX] project_todo: allow user to edit tag and users in todo on mobile
Before this commit, when the user is on mobile and try to edit/create a
new todo, the user cannot edit a tag or users because those fields are
displayed in actions dropdown and that dropdown is closed each time the
user clicks on one of those fields.

This commit prevents to close the dropdown in todo form view since that
dropdown contains fields and not buttons to let the user to edit those
fields.

closes odoo/odoo#193177

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>