[FIX] rpc: re-raise exceptions for http.py
Exceptions in RPC are handled and a Response is returned directly. It is
a different behaviour than jsonrpc where the exception is raised and the
dispatcher wraps it is a Response after logging the exception. The
xmlrpc code should also raise an exception so that it can be handled in
http.py. Adding here a way to detect if a Response was already generated
for an exception in the exception handler of HTTP.

closes odoo/odoo#193659

Related: #193208
X-original-commit: 4375971f8424eddf26d9f20aaff50e643268c1f3
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>