[FIX] hw_drivers: wrong foot pedal name
Devices detected through usb don't always have there manufactured and
product name which can be retrieved.
For example, our foot pedal in the office doesn't which leads to it
having a name "None - None" in Odoo.

After this PR if a device doesn't have a product name / manufacturer
name it will be saved as "Unknown input device"

task-4472585

closes odoo/odoo#193277

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>