[FIX] point_of_sale: assign lot_id on quants when applicable
Currently, when buying a product in pos and using a new lot id, the
quantities do not reflect this lot id.

Steps to reproduce:
-------------------
* Create a product:
  * Storable
  * Available in POS
  * Tracked by lots
* Open shop session
* Select the product just created
* Enter anything as the lot number
* Finalize order
* Close session
* In the back-end, navigate to the product page
* Select the smart button **On hand**
> Observation: The lot is not reflected

Why the fix:
------------
When the session in closed, `stock.lot` are created but not reflected on
the inventory quantity updates.

We can add the `lot_id` when creating move lines as where we mention it,
it will always have a related existing lot as per:
https://github.com/odoo/odoo/blob/ba21b59d1234d13b5d4460facd60c586648048b8/addons/point_of_sale/models/stock_picking.py#L290

opw-4359435

closes odoo/odoo#193649

X-original-commit: 2b57964415c82d6f50e4947b34531c5dd2bb623a
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>