[IMP] l10n_es_edi_tbai: allow setting reversed invoices in outgoing refunds
When doing a refund for an outgoing invoice, sometimes the user will use Odoo's built-in refund features and the reversed_entry_id field will be filled appropriately. That's OK.

However, other times they will just create a new invoice from scratch. For example, when refunding one line for an invoice that had 100 lines, it's easier to create a new invoice with one line than to create a refund and delete 99 lines from it. In these cases, the user can't link it with the invoice it's refunding, resulting in an error for TicketBAI.

See https://github.com/odoo/odoo/pull/191303#discussion_r1910034083 for details.

@moduon MT-4966

closes odoo/odoo#193392

Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>