[FIX] mail: broken live call participant badge
Before this commit the `LIVE` badge on the minimized call participant card
had some styling issues. This was caused by missing padding on the related
div tag and non-transparent background on the parent tag.
This commit fixes the issue by introducing some padding and removing the
css class that adds background opacity.

closes odoo/odoo#193516

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>