[FIX] account: cash basis journal for last payment
If you pay an invoice with multiple payments some roundings are made
for the taxes. On the last payment, the error from those roundings
are put on a move to make sure that the exact amount of the invoice is
paid.

That move was put on the wrong journal. It was put in the 'Exchange
Difference' journal instead of the 'Cash Basis Taxes'.

opw-4236755

closes odoo/odoo#193994

Signed-off-by: William André (wan) <wan@odoo.com>