[FIX] sale_pdf_quote_builder: use correct datetime format
Versions
--------
- saas-17.4+

Steps
-----
1. Create a new quotation template;
2. add a header which uses the `date_order` field;
3. configure the mapping from this form field to the record field;
4. print the quotation template.

Issue
-----
The datetime is formatted as "Jan 16, 2025, 5:59:26 PM", which is
different from how dates are formatted, or how datetimes are usually
formatted elsewhere in Odoo.

Additionally, it may also be too long to fit in the form field.

Cause
-----
The `format_datetime` function is used without a `dt_format` specified,
instead falling back on the default `'medium'`.

Solution
--------
In order to get the format defined for the relevant locale in Odoo,
`dt_format=False` should be specified in the `format_datetime` call.

opw-4414688

closes odoo/odoo#194113

X-original-commit: 6d26db11839aa651927dcdc4c08449fb5f76b7b0
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>