[FIX] mail: draggable participant cards on mobile
Before this commit, when dragging participant cards on mobile they
will get dragged off screen.
1. Open discuss on mobile
2. Get into a call
3. Move the participant cards -> they will disappear

This happens because the `drag` function gets called without checking
if the participant card is in the inset state (the state in which it
can be moved around).
This commit fixes the issue by checking that the component is in that
state before being able to drag it.

task-4491597

closes odoo/odoo#194075

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>