[FIX] base_automation: hide domains for on_change's
Before this commit, the base.automation form view displays the
filter_domain and filter_pre_domain fields when on_change trigger type
is selected but the code does not take into account those fields values
for this type of trigger.

After this commit, these fields are simply hidden in that case.

closes odoo/odoo#194321

Task: opw-4492554
X-original-commit: fcae692a1cb1521629553e7635235a1f1f309ad5
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>