[FIX] mrp: duplicate MO and change date
Steps to reproduce:
-----

- Create a MO
- Duplicate it
- Change the scheduled date
- Confirm

Issue:
---

This (changes)[https://github.com/odoo/odoo/commit/7c808beaf36853b4d9171ef0981d1ec9c4b73a44],
is trying to timedelta between str and datetime leading to an error.

Fix:
---
To fix this the date_start is set as a datetime and removing the conversion
later in the code.

opw-4489618

closes odoo/odoo#194376

Signed-off-by: William Henrotin (whe) <whe@odoo.com>