[IMP] core: disable downloading during tours
It's not *entirely* clear which and when, but some tours apparently
trigger downloads, which by default will make a mess of the user's
Downloads folder when running tests locally.

This has mostly been observed in 18.0 / master with Studio exports but
there are a few others as well (e.g. a few reports). Either way, seems
like a good idea to just configure chrome to block them.

closes odoo/odoo#194417

X-original-commit: 4d57b5408239a8492e58153675d688c387d3bc28
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>