[FIX] base_automation: hide domain for on_change's
Before this commit, the base.automation form view displays the
filter_domain field when on_change trigger type is selected but the code
does not take into account the filter_domain value for this type of
trigger.

After this commit, this field is simply hidden in that case.

closes odoo/odoo#194520

Task: opw-4492554
X-original-commit: 516c7634edaa3cd8101d1006e44a0cc3e12355ae
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>