[FIX] l10n_ch: print Swiss Qrcodes when the transaction is possible
Steps to reproduce:
- Install "l10n_ch" and switch to a Swiss company
- Create an invoice for a Swiss partner
- Send it
- The QR code appears in the generated PDF

Cause:
The bank eligibility is not checked when printing the QR code.

Solution:
If the fiscal_country is not Switzerland then we check the account validity.
A valid account is an IBAN account of this type type CHXX 3000 0XXX XXXX with the number in the middle being between 30000 and 31999.

opw-4380520

closes odoo/odoo#194563

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>