[FIX] l10n_my_edi: XML format and test mode
Fixes an issue with the xml file where the tax
exchange rate node would be added before the
payment terms node if both features are used,
which render the xml wrong.

Also add a sudo when getting the system param
for the test url; as this shouldn't block a non
administrator user from testing the feature.

opw-4425472
opw-4505562

closes odoo/odoo#195044

X-original-commit: 71dc1262fefd3052e616d3c01bee3573431ccb98
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>